% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeEg2.17.R
\name{AndradeEg2.17}
\alias{AndradeEg2.17}
\title{Comprimento e Peso de \emph{Penaeus Paulensis}}
\format{Um \code{data.frame} com 18 observações e 2 variáveis, em que

\describe{

\item{\code{Xc}}{Comprimento do Corpo em mm.}

\item{\code{Xp}}{Peso em gramas.}

}}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Exemplo 2.17, pág.
    128)
}
\description{
Comprimento do corpo \eqn{X_c}, dado em mm, e peso
    \eqn{X_p} de fêmeas, dado em g, de Penaeus paulensis (Crustacea,
    Decapoda, Penaidae), obtidos nas despescas dos viveiros do Centro
    de Ciências Agrárias (CCA) da Universidade Federal de Santa
    Catarina (UFSC).
}
\examples{

data(AndradeEg2.17)
str(AndradeEg2.17)

library(lattice)
xyplot(jitter(Xp) ~ Xc,
       data = AndradeEg2.17,
       type = c("p", "r"),
       xlab = "Comprimento do Corpo (em mm)",
       ylab = "Peso (em gramas)")

}
\keyword{AAS}

