% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeEg4.11.R
\name{AndradeEg4.11}
\alias{AndradeEg4.11}
\title{Dispersão de Formigas da Espécie \emph{Acromyrmex striatus}}
\format{Um \code{data.frame} com 5 observações e 2 variáveis, em que

\describe{

\item{\code{ninhos}}{Fator de 5 níveis qualitativos, que são os
    os números de ninhos por quadrante.}

\item{\code{quadrantes}}{Número de quadrantes com os respetivos
    números de ninhos.}

}}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Exemplo 4.11, pág.
    228)
}
\description{
Estudo da dispersão de formigas da espécie
    \emph{Acromyrmex striatus}, uma espécie de formiga cortadeira,
    em dunas da praia da Joaquina, Florianópolis, SC, onde uma área
    de 100m x 100m, foi subdividida em 100 quadrantes de 10m x 10m.
    Em cada um desses quadrantes foi contado o número de ninhos de
    \emph{Acromyrmex striatus}. Maiores detalhes sobre esse estudo
    estão em Lopes (2001).
}
\examples{

data(AndradeEg4.11)
str(AndradeEg4.11)

barplot(AndradeEg4.11$quadr,
        col = "darkturquoise",
        names.arg = AndradeEg4.11$ninhos,
        xlab = "Ninhos por Quadrante",
        ylab = "Número de Quadrantes")

}
\keyword{DBC}

