% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeEx2.13.R
\name{AndradeEx2.13}
\alias{AndradeEx2.13}
\title{Estudo dos Ovos de \emph{Biomphalaria Taenagophila}}
\format{Um \code{vetor} numérico com 23 observações.}
\source{
Estatística para as ciências agrárias e biológicas: com
    noções de experimentação / Dalton F. Andrade, Paulo J. Ogliari.
    2. ed. rev. e ampl. - Florianópolis: Ed. da UFSC, 2010.
    (Capítulo 2, Exercício 13, pág. 109)
}
\description{
Estudo do comportamento da variável número de ovos
    inviáveis de \emph{Biomphalaria taenagophila} (caramujo) em
    ambiente poluído. Para isso tomou-se uma amostra de 23 caramujos.
}
\examples{

library(lattice)

data(AndradeEx2.13)
str(AndradeEx2.13)

hist(AndradeEx2.13,
     labels = TRUE,
     prob = TRUE,
     main = NULL,
     col = "darkturquoise",
     ylim = c(0, 0.12),
     xlab = "Número de Ovos",
     ylab = "Densidade")

lines(density(AndradeEx2.13), lwd = 2)

}
\keyword{AAS}

