% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.12.R
\name{AndradeTb2.12}
\alias{AndradeTb2.12}
\title{Relação do Número de Pássaros com Local de Alimentação}
\format{Um \code{data.frame} com 6 observações e 3 variáveis, em que

\describe{

\item{\code{estacao}}{Fator de 2 níveis qualitativos que são as
    estações do ano (primavera e outono).}

\item{\code{local}}{Fator de 3 níveis qualitativos que são os locais
    da Floresta (árvores, arbusto e chão).}

\item{\code{passaros}}{Número de pássaros de uma particular espécie.}

}}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Tabela 2.12, pág. 81)
}
\description{
Estudo sobre o número de pássaros de uma particular
    espécie observados em diferentes locais de alimentação na
    floresta. O estudo foi realizado em duas estações do ano.
}
\examples{

data(AndradeTb2.12)
str(AndradeTb2.12)

(tb <- xtabs(passaros ~ estacao + local, data= AndradeTb2.12))

mosaicplot(tb)

library(lattice)
barchart(tb,
         stack = FALSE,
         horizontal = FALSE,
         auto.key = list(
             title = "Local de alimentação",
             cex.title = 1.1,
             columns = 3))

}
\keyword{contingência}

