% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.16.R
\name{AndradeTb2.16}
\alias{AndradeTb2.16}
\title{Altura dos Calos em Explantes do Cerrado \emph{Mandevilla
    velutina}}
\format{Um \code{vetor} numérico com 37 observações.}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Tabela 2.16, pág. 96)
}
\description{
Avaliação da altura de calos em explantes do cerrado da
    espécie \emph{Mandevilla velutina} após sua formação, durante um
    período de 30 dias \emph{in vitro}. Têm-se interesse que os calos
    não se desenvolvam, pois quanto menor, maior será o
    desenvolvimento das plantas.
}
\examples{

data(AndradeTb2.16)
str(AndradeTb2.16)

summary(AndradeTb2.16)
hist(AndradeTb2.16,
     col = "darkturquoise",
     labels = TRUE,
     xlab = "Altura dos Calos (em cm)",
     ylab = "Frequência",
     xlim = c(0, 1))

}
\keyword{AAS}

