% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.25.R
\name{AndradeTb2.25}
\alias{AndradeTb2.25}
\title{Experimento de Germinação de Cultivares de Cebola}
\format{Um \code{data.frame} com 4 observações e 3 variáveis, em que

\describe{

\item{\code{Cultivares}}{Fator de 2 níveis qualitativos que são as
    cultivares de cebola (Bola Precoce-EMPASC 352 e Norte 14).}

\item{\code{Germinação}}{Fator de 2 níveis qualitativos que indica se
    a semente germinou ou não germinou.}

\item{\code{Freq}}{Frequência conjunta de Cultivares e Germinação.}

}}
\source{
Estatística para as ciências agrárias e biológicas: com
    noções de experimentação / Dalton F. Andrade, Paulo J. Ogliari.
    2. ed. rev. e ampl. - Florianópolis: Ed. da UFSC, 2010. (Tabela
    2.25, pág. 107)
}
\description{
Experimento com o objetivo de avaliar o poder
    germinativo de duas cultivares de cebola. Foram utilizadas para o
    teste de germinação quatro repetições de cem sementes,
    totalizando quatrocentas sementes para cada cultivar. A variável
    em estudo é o número de sementes que germinaram.
}
\examples{

library(lattice)

data(AndradeTb2.25)
str(AndradeTb2.25)

xyplot(Freq ~ Germinacao,
       groups = Cultivares,
       data = AndradeTb2.25,
       type = "b",
       auto.key = list(title = "Cultivares",
                       cex.title = 1.1,
                       columns = 2),
       xlab = "Germinação",
       ylab = "Frequência")

}
\keyword{AAS}

