% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.26.R
\name{AndradeTb2.26}
\alias{AndradeTb2.26}
\title{Local de Coleta e Peso de Mexilhões}
\format{Um \code{data.frame} com 6 observações e 3 variáveis, em que

\describe{

\item{\code{Local}}{Fator de 2 níveis qualitativos que são os locais
    de coleta dos mexilhões.}

\item{\code{Peso}}{Fator de 3 níveis qualitativos que indica os pesos
    dos mexilhões, em gramas.}

\item{\code{Freq}}{Frequência conjunta de local de coleta e peso de
    mexilhões.}

}}
\source{
Estatística para as ciências agrárias e biológicas: com
    noções de experimentação / Dalton F. Andrade, Paulo J. Ogliari.
    2. ed. rev. e ampl. - Florianópolis: Ed. da UFSC, 2010. (Tabela
    2.26, pág. 107)
}
\description{
Distribuição conjunta das variáveis local de coleta e
    peso de mexilhões.
}
\examples{

library(lattice)

data(AndradeTb2.26)
str(AndradeTb2.26)

xyplot(Freq ~ Local,
       groups = Peso,
       data = AndradeTb2.26,
       type = "b",
       auto.key = list(title = "Peso",
                       cex.title = 1.1,
                       columns = 3),
       xlab = "Local de Coleta",
       ylab = "Frequência")

}
\keyword{AAS}

