% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.27.R
\name{AndradeTb2.27}
\alias{AndradeTb2.27}
\title{Estudo da Eficiência de uma Vacina Antigripal}
\format{Um \code{data.frame} com 6 observações e 3 variáveis, em que

\describe{

\item{\code{Saude}}{Fator de 2 níveis qualitativos que são os estados
    de saúde dos habitantes.}

\item{\code{Vacina}}{Fator de 3 níveis qualitativos que são as
    dosagens utilizadas da nova vacina antigripal.}

\item{\code{Freq}}{Frequência conjunta do estado de saúde e da
    vacinação dos habitantes.}

}}
\source{
Estatística para as ciências agrárias e biológicas: com
    noções de experimentação / Dalton F. Andrade, Paulo J. Ogliari.
    2. ed. rev. e ampl. - Florianópolis: Ed. da UFSC, 2010. (Tabela
    2.27, pág. 108)
}
\description{
Estudo  para avaliar a eficiência de uma nova vacina
    antigripal, a qual foi administrada aos membros de uma pequena
    comunidade, em duas doses, ao longo de duas semanas. Algumas
    pessoas tomaram as duas doses, outras tomaram apenas a primeira
    dose e outras não tomaram nenhuma dose. Os tratamentos foram
    aplicados em 1.000 habitantes dessa comunidade.
}
\examples{

library(lattice)

data(AndradeTb2.27)
str(AndradeTb2.27)

xyplot(Freq ~ Saude,
       groups = Vacina,
       data = AndradeTb2.27,
       type = "b",
       auto.key = list(title = "Vacinação",
                       cex.title = 1.1,
                       columns = 3),
       xlab = "Estado de Saúde",
       ylab = "Frequência")

}
\keyword{AAS}

