% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.36.R
\name{AndradeTb2.36}
\alias{AndradeTb2.36}
\title{Produção de Matéria Seca e Radiação de uma Cultura}
\format{Um \code{data.frame} com 10 observações e 2 variáveis, em que

\describe{

\item{\code{Prod}}{Produção de matéria seca em \eqn{g.m^{-2}}.}

\item{\code{Rad}}{Radiação fotossintética em \eqn{W.m^{-2}}.}

}}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Tabela 2.36, pág. 140)
}
\description{
Dados de produção de matéria seca de uma cultura e
    quantidade de radiação fotossintética ativa.
}
\examples{

data(AndradeTb2.36)
str(AndradeTb2.36)

library(lattice)
xyplot(Prod ~ Rad,
       data = AndradeTb2.36,
       type = c("p", "r"),
       xlab = expression(Radiação ~ fotossintética ~ (W / m^2)),
       ylab = expression(Produção ~ (g / m^2)))

}
\keyword{AAS}

