% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.46.R
\name{AndradeTb2.46}
\alias{AndradeTb2.46}
\title{Condutividade e Salinidade da Lagoa da Conceição}
\format{Um \code{data.frame} com 6 observações e 3 variáveis, em que

\describe{

\item{\code{estacao}}{Fator de 6 níveis qualitativos, que são as
    estações da região III da Lagoa da Conceição.}

\item{\code{cond}}{Condutividade da região III da Lagoa da Conceição,
    em mho.}

\item{\code{salin}}{Salinidade da região III da Lagoa da Conceição,
    em g/l.}

}}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Tabela 2.46, pág. 149)
}
\description{
Valores de condutividade (mho) e salinidade (g/l) para a
    região III da Lagoa da Conceição.
}
\examples{

data(AndradeTb2.46)
str(AndradeTb2.46)

library(lattice)
xyplot(cond ~ salin,
       data = AndradeTb2.46,
       type = c("p", "r"),
       xlab = "Salinidade (g/l)",
       ylab = "Condutividade (mho)")

}
\keyword{AAS}

