% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.6.R
\name{AndradeTb2.6}
\alias{AndradeTb2.6}
\title{Experimento de Cultura de Feijão de Vagem}
\format{Um \code{data.frame} com 50 observações e 2 variáveis, em que

\describe{

\item{\code{H}}{Fator de 2 níveis qualitativos que são os hábitos de
    crescimento, sendo: 3 = "indeterminado trepador"; 4 =
    "indeterminado prostrado".}

\item{\code{P}}{Fator de 3 níveis qualitativos que são os portes,
    sendo: Tr = "trepador"; EB = "ereto na base"; Pr = "prostrado".}

}}
\source{
Estatística para as ciências agrárias e biológicas: com
    noções de experimentação / Dalton F. Andrade, Paulo J. Ogliari.
    2. ed. rev. e ampl. - Florianópolis: Ed. da UFSC, 2010. (Tabela
    2.6, pág. 74)
}
\description{
Experimento no qual um pesquisador está procurando
    verificar se existe associação entre hábito de crescimento e
    porte na cultura de feijão de vagem.
}
\examples{

library(lattice)
library(plyr)

data(AndradeTb2.6)
str(AndradeTb2.6)

tb <- xtabs(~H+P, data= AndradeTb2.6)

barplot(t(tb), names.arg = c("Prostrado", "Trepador"),
        beside=TRUE,
        space = c(0.2,1),
        col = c("darkturquoise","lawngreen","blue"),
        ylim = c(0, 30))

legend("topleft",
       legend = c("Eb", "Pr", "Tr"),
       col = c("darkturquoise", "lawngreen","blue"),
       pch = 8)

}
\keyword{AAS}

