% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BanzattoQd1.2.3.R
\name{BanzattoQd1.2.3}
\alias{BanzattoQd1.2.3}
\title{Diâmetro à Altura do Peito de \emph{Eucalyptus saligna}}
\format{Um vetor numérico com 20 elementos.}
\source{
Banzatto, D. A., & Kronka, S. D. (2013). Experimentação
    Agrícola (4th ed.). Jaboticabal, SP: Funep. (Quadro 1.2.3 pág. 3)
}
\description{
Valores de diâmentro à altura do peito (DAP, cm) de uma
    amostra aleatória de árvores de \emph{Eucalyptus saligna} de um
    povoamento com 15 anos de idade.
}
\examples{
data(BanzattoQd1.2.3)

hist(BanzattoQd1.2.3, prob = TRUE,
     xlab = expression(Diâmetro~à~altura~do~peito~(cm)),
     ylab = "Densidade", main = NULL)
lines(density(BanzattoQd1.2.3), lwd = 2)
rug(BanzattoQd1.2.3)

boxplot(BanzattoQd1.2.3,
        ylab = expression(Diâmetro~à~altura~do~peito~(cm)))

mean(BanzattoQd1.2.3)
sd(BanzattoQd1.2.3)
fivenum(BanzattoQd1.2.3)

}
\keyword{amostra}

