% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banzatto.R
\name{BanzattoQd3.4.1}
\alias{BanzattoQd3.4.1}
\title{Produtividades de Cultivares de Mandioca}
\format{Um \code{data.frame} com 25 observações e 3 variáveis, em que

\describe{

\item{\code{cult}}{Fator de níveis nominais. Cultivares de mandioca
    em competição no experimento.}

\item{\code{rept}}{Inteiro que identifica a unidade experimental em
    cada cultivar.}

\item{\code{prod}}{Produção observada em cada unidade experimental,
    em ton ha\eqn{^{-1}}.}

}}
\source{
BANZATTO; KRONKA (2013), Quadro 3.4.1, pág. 50.
}
\description{
Produção de mandioca em experimento de competição de
    cultivares de mandioca. O experimento foi instalado em
    delineamento inteiramente casualizado devido a área ser homogênea
    quanto às condições experimentais.
}
\examples{

library(lattice)

data(BanzattoQd3.4.1)

aggregate(prod ~ cult,  data = BanzattoQd3.4.1,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

xyplot(prod ~ cult, data = BanzattoQd3.4.1,
       xlab = "Cultivares de mandioca",
       ylab = expression(Produtividade~de~mandioca~(t~ha^{-1})))

}
\keyword{DIC}
