% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banzatto.R
\name{BanzattoQd4.5.2}
\alias{BanzattoQd4.5.2}
\title{Efeito do Promalin sobre Furtos de Macieira}
\format{Um \code{data.frame} com 20 observações e 3 variáveis.

\describe{

\item{\code{promalin}}{Fator de níveis nominais que representa a
    aplicação de promalin. Os níveis são a dose de promalin (ppm)
    aplicado em plena floração (12.5, 25, e 50), na floração e na
    frutificação (12.5 + 12.5) e a testemunha (sem aplicação).}

\item{\code{bloco}}{São os blocos do experimento, considerados para
    fazer controle local.}

\item{\code{peso}}{Peso (g) médio baseado em um total de 250 frutos
    por parcela.}

}}
\source{
Banzatto, D. A., Kronka, S. D. (2013). Experimentação
    Agrícola (4th ed.). Jaboticabal, SP: Funep. (Quadro 4.5.2
    pág. 84)

Mestriner,  J. C. (1980). Estudos dos efeitos do promalin sobre
    frutos de macieiras (\emph{Malus} ssp) cultivares Brasil e
    Rainha. Trabalho de Conclusão de Curso (Graduação em Agronomia) -
    Faculdade de Ciências Agrárias e Veterinárias,  Universidade
    Estadual Paulista, Jaboticabal - SP.
}
\description{
Resultados de um experimento instalado na Fazenda
    Chapadão, no município de Angatuba - SP. O delineamento
    experimental foi o de blocos casualizados, sendo as parcelas
    constituídas de 4 plantas espaçadas de 6 x 7 metros, com 12 anos
    de idade na época da instalação do experimento.
}
\examples{

library(lattice)

data(BanzattoQd4.5.2)

addmargins(with(BanzattoQd4.5.2,
                tapply(X = peso,
                       INDEX = list(promalin, bloco),
                       FUN = sum)))

xyplot(peso ~ bloco, data = BanzattoQd4.5.2,
       groups = promalin, type = "b",
       xlab = "Bloco",
       ylab = "Peso médio dos frutos (g)",
       auto.key = list(title = "Forma de aplicação do Promalin",
                       cex.title = 1, columns = 2))
}
\keyword{DBC}

