% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banzatto.R
\name{BanzattoQd5.3.7}
\alias{BanzattoQd5.3.7}
\title{Efici\enc{ê}{e}ncia na Utiliza\enc{çã}{ca}o de P por Cultivares de trigo}
\format{Um \code{data.frame} com 48 observações e 5 variáveis.
\describe{

\item{\code{cult}}{Cultivares de trigo usadas no experimento: BR 20
    Guató - tolerante ao alumínio, BR 36 Ianomami - sensível ao
    alumínio, BR 40 Tuiúca - moderadamente sensível ao alumínio.}

\item{\code{calag}}{Quantidade aplicada de calcário feita para
    correção da acidez do solo (calagem): 0 e 4,4 ton ha\eqn{^{-1}}.}

\item{\code{P}}{Quantidade aplicada de fósforo: 0 e 87 mg de P por kg
    de solo.}

\item{\code{rept}}{Identifica a repetição de cada cela experimental.}

\item{\code{efic}}{Variável resposta que a eficiência na utilização
    do fósforo, dado pelo quociente da massa seca da parte aérea
    produzida pela quantidade de fósforo absorvida pelas plantas.}

}}
\source{
Banzatto, D. A., Kronka, S. D. (2013). Experimentação
    Agrícola (4th ed.). Jaboticabal, SP: Funep. (Quadro 5.3.7,
    pág. 120)
}
\description{
Experimento em delineamento inteiramente casualizado que
    avaliou o efeito a eficiência na utilização de fósforo por três
    cultivares de trigo cultivadas em solo com e sem calagem.
}
\examples{

library(lattice)

data(BanzattoQd5.3.7)

addmargins(with(BanzattoQd5.3.7,
                tapply(X = efic,
                       INDEX = list(interaction(cult, calag, P),
                                    rept),
                       FUN = sum)))

xyplot(efic ~ factor(P) | cult, data = BanzattoQd5.3.7,
       groups = calag, type = c("p", "a"),
       layout = c(NA, 1), jitter.x = TRUE,
       auto.key = list(title = expression("Calagem"~(ton~ha^{-1})),
                       cex.title = 1, columns = 2),
       xlab = expression("Fósforo"~(mg~kg^{-1}~de~solo)),
       ylab = expression("Eficiência no uso do Fósforo"))

}
\keyword{DIC}
\keyword{FAT3}

