% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banzatto.R
\name{BanzattoQd5.5.1}
\alias{BanzattoQd5.5.1}
\title{Aduba\enc{çã}{ca}o Mineral no Milho com Uso de Confundimento}
\format{Um \code{data.frame} com 27 observações e 4 variáveis, em que
    \describe{

\item{\code{N}}{Níveis codificados de nitrogênio (N) que representam
    0, 30 e 60 kg ha\eqn{^{-1}}. A fonte do N é o sulfato de amônio
    (20\% de N).}

\item{\code{P}}{Níveis codificados de fósforo (P) que representam 0,
    30 e 60 kg ha\eqn{^{-1}}. A fonte do P é o superfosfato simples
    (18\% de P\eqn{_{2}}O\eqn{_{5}}).}

\item{\code{K}}{Níveis codificados de potássio (K) que representam 0,
    30 e 60 kg ha\eqn{^{-1}}. A fonte do K é o cloreto de potássio
    (60\% de K\eqn{_{2}}O).}

\item{\code{bloc}}{Fator de níveis nominais que identifica os
    blocos.}

\item{\code{prod}}{Variável resposta observada que é a produção de
    grãos de milho (kg ha\eqn{^{-1}}).}

}

Cada parcela foi constituída de 6 linhas de plantio com 10 metros de
comprimento, num total de 60 m\eqn{^{2}} e, como área útil, foram
consideradas as 4 linhas centrais com 8 metros de comprimento (por
desconsiderar 1 m em cada extremidade), perfazendo 32 m\eqn{^{2}}. O
híbrido de milho utilizado foi o HMD-7974.}
\source{
Banzatto, D. A., Kronka, S. D. (2013). Experimentação
    Agrícola (4th ed.). Jaboticabal, SP: Funep. (Quadro 5.5.1,
    pág. 131)

Vilalta, O. A. (1972). Avaliação da produção de milho (\emph{Zea
    mays} L.) em função da adubação NPK em um latossolo
    vermelho-escuro -- fase arenosa. Trabalho de Conclusão de Curso
    (Graduação em Agronomia). Faculdade de Ciências Agrárias e
    Veterinárias, Universidade Estadual Paulista, Jaboticabal, SP.
}
\description{
Experimento realizado em Jaboticabal - SP por Vilalta
    (1972) que, com um arranjo fatorial triplo, estudou a combinação
    de nitrogênio, fósforo e potássio, cada um com 3 níveis (27
    combinações ao todo), sobre a produção de grãos de milho. Como o
    número de tratamentos é 27, não seria viável, em termos de
    controle local, ter blocos desse tamanho. Portanto, usou-se a
    técnica de confundimento para distribuir os 27 tratamentos em 3
    blocos de tamanho 9. A técnica foi usada para confundir dois
    graus de liberdade da interação tripla com o efeito dos blocos. O
    grupo de confundimento adotado foi o W.
}
\examples{

library(lattice)

data(BanzattoQd5.5.1)

a <- with(BanzattoQd5.5.1,
          tapply(X = prod, INDEX = list(N, P, K), FUN = sum))
addmargins(apply(a, MARGIN = c(1, 2), FUN = sum))
addmargins(apply(a, MARGIN = c(1, 3), FUN = sum))
addmargins(apply(a, MARGIN = c(2, 3), FUN = sum))

xyplot(prod ~ factor(N) | factor(P), data = BanzattoQd5.5.1,
       groups = K, type = c("p", "a"), layout = c(NA, 1),
       auto.key = list(title = "Potássio (K)", cex.title = 1,
                       columns = 3),
       strip = strip.custom(strip.names = TRUE,
                            var.name = "Fósforo (P)"),
       xlab = "Nitrogênio (N)",
       ylab = expression("Produção de grãos de milho"~(ka~ha^{-1})))

}
\keyword{DBC}
\keyword{FAT3}
\keyword{confundimento}

