% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BanzattoQd6.3.4.R
\name{BanzattoQd6.3.4}
\alias{BanzattoQd6.3.4}
\title{\enc{É}{E}pocas de Plantio e Colheita na Produ\enc{çã}{ca}o de Beterraba}
\format{Um \code{data.frame} com 72 observações e 5 variáveis.
    \describe{

\item{\code{data}}{Fator de níveis ordinais que representa as datas
    de plantio de beterraba. Os níveis do fator data foram
    casualizados nas parcelas.}

\item{\code{inset}}{Variável binária que representa a aplicação de
    inseticida (1) ou não (0). Este fator foi casualizado as
    subparcelas.}

\item{\code{epoc}}{Fator de níveis ordinais que representa a época de
    colheita das beterrabas. Esse fator foi casualizado nas
    sub-subparcelas.}

\item{\code{bloco}}{Fator de níveis nominais que presenta os blocos
    do experimento.}

\item{\code{prod}}{Produção de beterraba. A unidade de medida não foi
    informada.}

}}
\source{
Banzatto, D. A., Kronka, S. D. (2013). Experimentação
    Agrícola (4th ed.). Jaboticabal, SP: Funep. (Quadro 6.3.4,
    pág. 156)
}
\description{
Experimento estudar o efeito de datas de plantio, épocas
    de colheita e aplicação de inseticidas na produção de
    beterraba. Os tratamentos foi arranjados em parcelas
    sub-subdivididas no delineamento de blocos casualizados.
}
\examples{

library(lattice)

data(BanzattoQd6.3.4)

str(BanzattoQd6.3.4)

ftable(with(BanzattoQd6.3.4,
            tapply(prod,
                   list(data = data,
                        inset = inset,
                        epoc = epoc,
                        bloco = bloco),
                   FUN = mean)))

xyplot(prod ~ epoc | data, groups = inset, data = BanzattoQd6.3.4,
       as.table = TRUE, layout = c(NA, 1),
       jitter.x = TRUE, type = c("p", "a"),
       auto.key = list(title = "Aplicação de inseticida",
                       cex.title = 1, columns = 2),
       strip = strip.custom(var.name = "Época", strip.names = TRUE),
       xlab = "Ordem das datas de plantio",
       ylab = "Produção de beterraba")

}
\keyword{DBC}
\keyword{PSS}

