% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BanzattoQd7.3.1.R
\name{BanzattoQd7.3.1}
\alias{BanzattoQd7.3.1}
\title{Di\enc{â}{a}metro e Altura de Tubetes na Forma\enc{çã}{ca}o de Mudas de Eucalipto}
\format{Um \code{data.frame} com 27 observações e 4 variáveis.

\describe{

\item{\code{dt}}{Diâmetro dos tubetes, em cm.}

\item{\code{at}}{Altura dos tubetes, em cm.}

\item{\code{bloco}}{Variável nominal que identifica os blocos do
    experimento.}

\item{\code{alt}}{Altura das mudas 75 dias após a semeadura.}

}}
\source{
Banzatto, D. A., Kronka, S. D. (2013). Experimentação
    Agrícola (4th ed.). Jaboticabal, SP: Funep. (Quadro 7.3.1,
    pág. 177)
}
\description{
Experimento fatorial 3 \eqn{\times} 3 que avaliou o
    efeito das dimensões do tubete (diâmetro e altura, 3 níveis cada)
    na produção de mudas de eucalipto. O experimento foi instalado em
    delineamento de blocos casualizados e a resposta medida foi a
    altura das mudas 75 dias após a semeadura.
}
\examples{

library(lattice)

data(BanzattoQd7.3.1)

str(BanzattoQd7.3.1)

xtabs(~at + dt, data = BanzattoQd7.3.1)

with(BanzattoQd7.3.1, addmargins(tapply(alt, list(at, dt), FUN = mean)))

xyplot(alt ~ at, data = BanzattoQd7.3.1,
       groups = dt, type = c("p", "a", "g"),
       auto.key = list(title = "Diâmetro dos tubetes (cm)",
                       cex.title = 1, columns = 3),
       xlab = expression("Altura do tubete"~(cm)),
       ylab = expression("Altura das mudas"~(cm)))

}
\keyword{DBC}
\keyword{FAT2}

