% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banzatto.R
\name{BanzattoQd7.3.3}
\alias{BanzattoQd7.3.3}
\title{Experimento Fatorial com Fatores Qualitativos e Quantitativos}
\format{Um \code{data.frame} com 32 observações e 4 variáveis, em que

\describe{

\item{\code{varied}}{Níveis nominais do fator variedade.}

\item{\code{adub}}{Níveis métricos do fator adubação, em kg
    ha\eqn{^{-1}}.}

\item{\code{bloco}}{Níveis nominais do fator bloco.}

\item{\code{prod}}{Produção, em ton ha\eqn{^{-1}}.}

}}
\source{
BANZATTO; KRONKA (2013), Quadro 7.3.3, pág. 182.
}
\description{
Dados fictícios de um experimento fatorial duplo com um
    fator qualitativo e outro quantitativo. O experimento está em
    delineamento de blocos casualizados e a resposta é a produção da
    cultura.
}
\examples{

library(lattice)

data(BanzattoQd7.3.3)

str(BanzattoQd7.3.3)

xtabs(~adub + varied, data = BanzattoQd7.3.3)

with(BanzattoQd7.3.3,
     addmargins(tapply(prod, list(adub, varied), FUN = mean)))

xyplot(prod ~ adub, data = BanzattoQd7.3.3,
       groups = varied, type = c("p", "a", "g"),
       auto.key = list(title = "Variedades",
                       cex.title = 1, columns = 2),
       xlab = expression("Adubação"~(kg~ha^{-1})),
       ylab = expression("Produção"~(kg~ha^{-1})))

}
\keyword{DBC}
\keyword{FAT2}
