% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banzatto.R
\name{BanzattoQd8.4.1}
\alias{BanzattoQd8.4.1}
\title{Grupo de Ensaios de Competi\enc{çã}{ca}o de Gen\enc{ó}{o}tipos de Cebola}
\format{Um \code{data.frame} com 60 observações e 3 variáveis.

\describe{

\item{\code{exper}}{Fator de níveis categóricos que representa os
    experimentos.}

\item{\code{genot}}{Fator de níveis categóricos que representa os
    genótipos de cebola.}

\item{\code{prod}}{Produção média de bulbos dos genótipos em cada
    experimento (média das repetições), em t ha\eqn{^{-1}}.}

}}
\source{
Banzatto, D. A., Kronka, S. D. (2013). Experimentação
    Agrícola (4th ed.). Jaboticabal, SP: Funep. (Quadro 8.4.1,
    pág. 200)

Santos, G. M. (2003). Adaptabilidade e estabilidade de genótipos de
    cebola. Tese (Doutorado em Genética e Melhoramento de
    Plantas). Faculdade de Ciências Agrárias e Veterinárias,
    Universidade Estadual Paulista, Jaboticabal: SP.
}
\description{
Grupo de experimentos de competição de genótipos de
    cebola realizado por Santos (2003). O delineamento considerado em
    cada ensaio não foi informado.
}
\examples{

library(lattice)

data(BanzattoQd8.4.1)

str(BanzattoQd8.4.1)

with(BanzattoQd8.4.1,
     addmargins(tapply(prod, list(genot, exper), FUN = mean)))

xyplot(prod ~ genot, groups = exper, data = BanzattoQd8.4.1,
       type = c("p", "a", "g"),
       auto.key = list(title = "Experimento",
                       cex.title = 1, columns = 3),
       scales = list(x = list(rot = 90)),
       xlab = "Genótipos de cebola",
       ylab = expression(
           "Produção média de bulbos"~(t~ha^{-1})))

}
\keyword{GE}

