% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barbin.R
\name{BarbinEx18}
\alias{BarbinEx18}
\title{Altura de Plantas de Esp\enc{é}{e}cies de Eucalipto em
    V\enc{á}{a}rios Locais}
\format{Um \code{data.frame} com 100 observações e 4 variáveis, em
    que

\describe{

\item{\code{loc}}{Fator categórico que representa os locais
    (municípios) onde foram feitos os ensaios.}

\item{\code{espec}}{Fator categórico que representa as espécies de
    eucalipto.}

\item{\code{rept}}{Números que identificam as repetições de cada
    espécie dentro de cada local.}

\item{\code{alt}}{Altura (m) de plantas de eucalipto com 3 anos de
    idade.}

}}
\source{
BARBIN (2013), Exercício 18, pág. 210.
}
\description{
Dados referentes à altura de plantas de eucalipto, com 3
    anos de idade, em ensaios inteiramente casualizados estudando o
    efeito da espécie de eucalipto em vários municípios (grupo de
    experimentos). Os ensaios foram conduzidos pelo Instituto
    Florestal, em Tupi/SP. Todos os ensaios receberam uma calagem e
    uma adubação completa.
}
\examples{

library(lattice)

data(BarbinEx18)

str(BarbinEx18)

xyplot(alt ~ espec | loc,
       data = BarbinEx18, type = c("p", "a"), jitter.x = TRUE,
       as.table = TRUE,
       xlab = "Espécies de eucalipto",
       ylab = "Altura de plantas com 3 anos de idade (m)",
       scales = list(x = list(rot = 90)))

}
\keyword{DIC}
\keyword{GE}

