% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarbinPg114.R
\name{BarbinPg114}
\alias{BarbinPg114}
\title{Influência do Aparelho e Operador na Medição de Árvores}
\format{Um \code{data.frame} com 200 observações e 4 variáveis.

\describe{

\item{\code{apar}}{Fator de níveis categóricos que identifica os
    aparelhos utilizados pelos operadores para medição da altura
    total das árvores.}

\item{\code{oper}}{Fator de níveis categóricos que identifica os
    operadores que utilizaram os aparelhos.}

\item{\code{bloc}}{Fator de níveis categóricos que representa as
    árvores que foram medidas. Cada árvore funciona como um bloco
    onde todos os operadores x aparelhos foram empregados.}

\item{\code{alt}}{Altura total da árvore (média de duas medidas
    consecutivas, em metros) determinada por cada aparelho ou
    instrumento de medida e operador.}

}}
\source{
Barbin, D. (2013). Planejamento e Análise Estatística de
    Experimentos Agronômicos (2nd ed.). Londrina, PR:
    Mecenas. (pág. 114)

Simões, J. W., Mello, H. A., Barbin, D. (1967). Eficiência dos
    aparelhos e e influência do operador na medição de altura total
    de árvores. O Solo, Piracicaba, v.2, p.57-63.
}
\description{
Dados de um experimento fatorial 5 \eqn{\times} 4, em
    delineamento de blocos casualizados, sobre e influência do
    aparelho e do operador na determinação de altura de árvores, aos
    7 anos de idade, de \emph{Eucaliptus saligna} (Simões,  Mello e
    Barbin, 1967).

Cinco aparelhos ou instrumentos de mensuração (hipsômetro de
    Blume-Leiss, hipsômetro de Haga, hipsômetro Weise, prancheta
    dendrométrica e trena) foram testados por 4 operadores resultando
    em 20 combinações. Diante de uma árvore era sorteado um número no
    universo de 1 a 20 que representava um operador \eqn{\times}
    aparelho. Cada operador com um aparelho efetuava duas medições
    consecutivas da mesma árvore (bloco) e a média era registrada. O
    sorteio continuava até as combinações fossem realizadas na
    árvore, então passava-se para a próxima árvore. Evidentemente, a
    medida feita com o trena ficava sempre para o final (testemunha).
}
\examples{

library(latticeExtra)

data(BarbinPg114)

str(BarbinPg114)

xyplot(alt ~ oper | bloc, data = BarbinPg114,
       groups = apar, type = "b",
       as.table = TRUE,
       auto.key = list(space = "top", columns = 2,
                       title = "Aparelhos", cex.title = 1.1),
       strip = strip.custom(strip.names = TRUE,
                            sep = " ", var.name = "Árvore"),
       xlab = "Operadores",
       ylab = "Altura total (m)")
    layer(panel.abline(h = mean(y), lwd = 2))

xyplot(alt ~ oper | apar, data = BarbinPg114,
       groups = bloc, type = "b",
       as.table = TRUE,
       auto.key = list(space = "right",
                       title = "Árvore", cex.title = 1.1),
       xlab = "Operadores",
       ylab = "Altura total (m)")

}
\keyword{DBC}
\keyword{FAT2}

