% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BarbinPg137.R
\name{BarbinPg137}
\alias{BarbinPg137}
\title{Adubação NPK na Produção de Algodão Herbáceo}
\format{Um \code{data.frame} com 54 observações e 6 variáveis.
    \describe{

\item{\code{N}}{Dose de nitrogênio (kg ha\eqn{^{-1}}) usada na
    adubação.}

\item{\code{P}}{Dose de fósforo (kg ha\eqn{^{-1}}) usada na
    adubação.}

\item{\code{K}}{Dose de potássio (kg ha\eqn{^{-1}}) usada na
    adubação.}

\item{\code{bloc}}{Fator de níveis categóricos considerado para
    controle local.}

\item{\code{rept}}{Número inteiro que identifica as repetições das
    celas experimentais dentro dos blocos.}

\item{\code{prod}}{Produção de algodão herbáceo, em kg
    ha\eqn{^{-1}}.}

}}
\source{
Barbin, D. (2013). Planejamento e Análise Estatística de
    Experimentos Agronômicos (2nd ed.). Londrina, PR:
    Mecenas. (pág. 137)

Cavalcanti, F. B. (1977). A adubação mineral na cultura do algodão
    (\emph{Gossypium hirsutum} L.) no Estado de
    Pernambuco. Dissetação de Mestrado, Escola Superior de
    Agricultura "Luiz de Queiroz"/USP, Piracicaba.
}
\description{
Produção de algodão herbáceo de um ensaio de adubação
    NPK \eqn{3^3} com confundimento (grupo W) de dois graus de
    liberdade da interação tripla. Cada combinação NPK teve duas
    repetições por bloco e o ensaio foi conduzido por Cavalcanti
    (1977) em Parnambuco.
}
\examples{
library(lattice)

data(BarbinPg137)

str(BarbinPg137)

x <- c("N", "P", "K")
forms <- unlist(lapply(mapply(FUN = combn,
                              m = 1:2,
                              MoreArgs = list(x = x)),
                       FUN = function(x) {
                           paste0("~",
                                  apply(x,
                                        MARGIN = 2,
                                        FUN = paste,
                                        collapse = "+"))
                       }))

# A frequência das combinações simples e duplas.
sapply(forms, xtabs, data = BarbinPg137)

xyplot(prod ~ as.factor(N) | as.factor(P), data = BarbinPg137,
       groups = K, type = c("p", "a"),
       auto.key = list(space = "top", columns = 3,
                       title = expression(
                           "Potássio"~(K * ", " * kg~ha^{-1})),
                       cex.title = 1.1),
       strip = strip.custom(strip.names = c(TRUE), sep = " = ",
                            var.name = expression(
                                "Fósforo"~(P * ", " * kg~ha^{-1}))),
       xlab = expression("Nitrogênio"~(N * ", " * kg~ha^{-1})),
       ylab = expression(
           "Produção de algodão herbáceo"~(kg~ha^{-1})))
}
\keyword{DBC}
\keyword{FAT3}
\keyword{confundimento}

