% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarbinPg156.R
\name{BarbinPg156}
\alias{BarbinPg156}
\title{Aplicação de Adubação Fosfatada na Produção de Milho}
\format{Um \code{data.frame} com 48 observações e 4 variáveis, em que

\describe{

\item{\code{aplic}}{Fator categórico cujos níveis são a forma de
    aplicação da adubação fosfatada.}

\item{\code{adub}}{Fator métrico cujos níveis representam a
    quantidade de adubação fosfatada, em kg ha\eqn{^{-1}}, feita com
    \eqn{\textrm{P}_2 \textrm{O}_5}.}

\item{\code{bloc}}{Fator de níveis categóricos cujos níveis
    representam os blocos do experimento.}

\item{\code{prod}}{Produção de milho (kg ha\eqn{^{-1}}).}

}}
\source{
Barbin, D. (2013). Planejamento e Análise Estatística de
    Experimentos Agronômicos (2nd ed.). Londrina, PR:
    Mecenas. (pág. 156)
}
\description{
Dados de um ensaio de adubação fosfatada em milho,
    realizado no esquema de parcelas subdivididas em delineamento de
    blocos ao acaso, que avaliou o efeito da forma de aplicação da
    adubação (cova, sulco ou lanço) e da dose de fósforo na produção
    de milho.
}
\examples{

library(lattice)

data(BarbinPg156)

str(BarbinPg156)

xyplot(prod ~ adub | aplic, data = BarbinPg156,
       groups = bloc, type = "o",
       as.table = TRUE,
       xlab = expression(
           "Dose de adubação fosfatada"~(kg~P[2]*O[5]~ha^{-1})),
       ylab = expression("Produção de milho"~(kg~ha^{-1})))

}
\keyword{DBC}
\keyword{PS}

