% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barbin.R
\name{BarbinPg25}
\alias{BarbinPg25}
\title{Enraizamento de Estacas de P\enc{ê}{e}ssego}
\format{Um \code{data.frame} com 20 observações e 2 variáveis, em que

\describe{

\item{\code{cult}}{Fator de níveis categóricos que representa as
    cultivares de pêssego estudadas.}

\item{\code{estac}}{Número de estacas enraizadas. Total de estacas
    plantadas foi 20 em cada parcela.}

}

O atributo \code{missings} é um vetor com números que indicam a
    posição das observações consideradas perdidas para se ilustrar a
    análise do experimento com número não igual de repetições.}
\source{
BARBIN (2013), pág. 25.

Sambão, J. C., Sampaio, V. R., Barbin, D. (1982). Enraizamento de
    estacas herbáceas de pessegueiro (\emph{Prunus persica}
    L.). Anais da E.S.A. "Luiz de Queiróz", Piracicaba, v.39,
    1039-1045.
}
\description{
Dados adaptados de Zambão, Sampaio e Barbin (1982), onde
    foram comparadas quatro cultivares de pêssego quando ao
    enraizamento de estacas. Experimento foi instalado em
    delineamento inteiramente casualizado.
}
\examples{

library(lattice)

data(BarbinPg25)

str(BarbinPg25)

xyplot(estac ~ cult, data = BarbinPg25, jitter.x = TRUE,
       xlab = "Cultivares de pêssego",
       ylab = "Número de estacas enraizadas do total de 20")

}
\keyword{DIC}
