% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrosDiasEg3.2.R
\name{BarrosDiasEg3.2}
\alias{BarrosDiasEg3.2}
\title{Dados da produção de frutos de 43 cacaueiros.}
\format{data.frame com 43 observações e 1 variável, em que

\describe{

\item{prod}{Número de frutos produzidos por cacaueiro.}

}}
\source{
Barros, W. S. &  Dias, L. A. S. (2009).
    Biometria Experimental, UFV Viçosa, MG.
}
\description{
Número de frutos produzidos por cacaueiro.
}
\examples{

boxplot(BarrosDiasEg3.2)

}
\keyword{cacauiero}
\keyword{produção}

