% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrosDiasEg3.4.R
\name{BarrosDiasEg3.4}
\alias{BarrosDiasEg3.4}
\title{Erros da produção de frutos de 43 cacaueiros}
\format{data.frame com 43 observações e 1 variável, em que

\describe{

\item{prod}{Erros de produção,por arvore,de frutos de cacau}

}}
\source{
L.A.S. Dias W.S. Barros,Biometria Experimental,
Viçosa MG, 2009.
}
\description{
Amostra aleatória de 43 arvores onde são 
     avaliados os erros de produção de frutos.
}
\examples{

library(lattice)

boxplot(BarrosDiasEg3.4)

}
\keyword{erros}
\keyword{produção}

