% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrosDiasEg3.6.R
\name{BarrosDiasEg3.6}
\alias{BarrosDiasEg3.6}
\title{alturas de 100 plantas de milho.}
\format{data.frame com 100 observações e 1 variável, em que

\describe{

\item{altura}{altura das plantas de milho}

}}
\source{
L.A.S. Dias W.S. Barros,Biometria Experimental,
     Viçosa MG, 2009.
}
\description{
amostras das alturas de 100 plantas de milho
    tomadas aleatoriamente.
}
\examples{

library(lattice)

boxplot(BarrosDiasEg3.6)

}
\keyword{altura}
\keyword{milho}

