% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorgesTb1.3.R
\name{BorgesTb1.3}
\alias{BorgesTb1.3}
\title{Tensiômetro}
\format{Um \code{data.frame} de 9 linhas e 2 colunas.

\describe{

    \item{\code{altura}}{ Altura da câmara do tensiômetro medida 
    em milímetros (mm). }
    
    \item{\code{tensao}}{ Tensão da água no solo medida em mb }

}}
\source{
Borges, C. G., Demétrio, & Zocchi, S. S. (2011). 
Modelo de Regressão. Piracicaba, SP: USP.
}
\description{
Foi observado a construção de um tensiômetro de leitura
direta, obtendo-se os resultados de altura da câmara e tensão da 
água no solo.
}
\examples{

## Carregando pacotes necessários -----------------------------------

library(lattice)

## Carregando o banco de dados --------------------------------------

data(BorgesTb1.3)

## Plotando o gráfico -----------------------------------------------

xyplot(tensao ~ altura, data = BorgesTb1.3,
    main = "Altura VS Tensão",
    xlab = "Altura do Tensiômetro",
    ylab = "Tensão da Água")
    
## Medidas de Posição -----------------------------------------------

summary(BorgesTb1.3)

}
\keyword{tensiômetro}

