% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEg.12.2.R
\name{CharnetEg.12.2}
\alias{CharnetEg.12.2}
\title{Relação do tecido adiposo e área abdominal}
\format{Um \code{data.frame} com 29 linhas e 5 colunas.

\describe{

\item{\code{tomo}}{Tomografia.}

\item{\code{cint}}{Medida da circunferência da cintura.}

\item{\code{us1}}{Medida 1 do tecido adiposo por ultrassonografia.}

\item{\code{us2}}{Medida 2 do tecido adiposo por ultrassonografia.}

\item{\code{pc}}{Medida de pregas cutâneas.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações 
    (2nd ed., p. 356). Editora Unicamp (Capítulo 12, exemplo 12.2, 
    página 286)
}
\description{
Relação entre medidas de tecido adiposo obtida por 
    tomagrafia computadorizada de área abdominal.
}
\examples{

data(CharnetEg.12.2)

library('ggplot2')

cintura <- qplot(us1, us2, data=CharnetEg.12.2, color=cint,
                 xlab='Tecido adiposo 1', ylab='Tecido adiposo 2',
                 main='Relação do tecido adiposos e a circunferência
                 da cintura')

cintura  + geom_abline(intercept = 0, slope = 1)

}
\keyword{regressão}
\keyword{tomografia}

