% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEg4.2.R
\name{CharnetEg4.2}
\alias{CharnetEg4.2}
\title{Avaliação de Vendedores pelos Clientes}
\format{Um \code{data.frame} com 2 colunas e 10 linhas.

\describe{

\item{\code{notas}}{Nota atribuída ao vendedor, em escala de 0 a 
    100.}

\item{\code{vendas}}{Volume anual de vendas, em milhões de reais.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Capítulo 4, exemplo 4.2,
    página 79)
}
\description{
Análise das vendas através das notas atribuídas pelos 
    clientes aos vendedores de uma empresa de seguro.
}
\examples{

data(CharnetEg4.2)

plot(CharnetEg4.2)

}
\keyword{RS}

