% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEg5.2}
\alias{CharnetEg5.2}
\title{Efeito de um Desinfetante}
\format{Um \code{data.frame} com 2 colunas e 16 linhas.

\describe{

\item{\code{conc}}{Concentração do desinfetante, em percentagem.}

\item{\code{bact}}{Contagem de bactérias, em unidade.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 5, exemplo 2,
    pág. 95)
}
\description{
Medição da quantidade de bactérias em um estudo sobre o
    efeito de um desinfetante diluído em quantidade fixas de água,
    em concentrações de 1 a 8%.
}
\examples{

data(CharnetEg5.2)

plot(CharnetEg5.2)

}
\keyword{RS}

