% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEg6.4.R
\name{CharnetEg6.4}
\alias{CharnetEg6.4}
\title{Efeito de um Produto Químico no Peso de Girassóis}
\format{Um \code{data.frame} com 2 colunas e 20 linhas.

\describe{

\item{\code{peso}}{Peso da planta, em gramas.}

\item{\code{dose}}{Dose da aplicação, em \eqn{\log(g cm^{-2})}.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 6, exemplo 4,
    pág. 136)
}
\description{
Medição do efeito da adição de uma solução de um certo
   produto químico na regação de girassóis.
}
\examples{

data(CharnetEg6.4)

plot(peso ~ dose, data = CharnetEg6.4)

}
\keyword{RS}

