% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEg7.3.R
\name{CharnetEg7.3}
\alias{CharnetEg7.3}
\title{Relação entre Salário, Tempo de Experiência e Sexo}
\format{Um \code{data.frame} com 2 colunas e 15 linhas.

\describe{

\item{\code{salario}}{Salário, em mil reais.}

\item{\code{exp}}{Tempo de experiência, em anos completos.}

\item{\code{sexo}}{Sexo do gerente de agência bancária.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 7, exem. 3, p. 152)
}
\description{
Dados de salário, tempo de experiência e sexo dos 
    gerentes de agências bancárias de um grande banco.
}
\examples{

data(CharnetEg7.3)

with(CharnetEg7.3, plot(salario, exp, col = sexo))

}
\keyword{RM}

