% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEg8.2}
\alias{CharnetEg8.2}
\title{Conjunto de Dados Gen\enc{é}{e}rico}
\format{Um \code{data.frame} com 2 colunas e 15 linhas, em que

\describe{

\item{\code{x}}{Variável independente, sem interpretação.}

\item{\code{y}}{Variável dependente, sem interpretação.}

}}
\source{
CHARNET et al. (2008), Capítulo 8, exercício 2, pág. 191.
}
\description{
Um conjunto de dados qualquer para exercício de ajuste
    de modelos de regressão linear com diferentes preditores.
}
\examples{

data(CharnetEg8.2)
str(CharnetEg8.2)

plot(CharnetEg8.2)
plot(CharnetEg8.2 - 1.1 * min(CharnetEg8.2), log = "xy")

}
\keyword{RS}
