% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEg9.2}
\alias{CharnetEg9.2}
\title{Rela\enc{çã}{ca}o entre Poluentes Atmosf\enc{é}{e}ricos}
\format{Um \code{data.frame} com 3 colunas e 7 linhas, em que

\describe{

\item{\code{CO}}{Monóxido de carbono, sem especificação.}

\item{\code{HC}}{Hidrocarbonetos, sem especificação.}

\item{\code{NO}}{Óxido de nitrogênio, sem especificação.}

}}
\source{
CHARNET et al. (2008), Capítulo 9, exemplo 2, pág. 216.
}
\description{
Relação de poluentes atmosféricos emitidos por carros
    durante uma coleta realizada em sete períodos diferentes em um
    grande centro.
}
\examples{

data(CharnetEg9.2)
str(CharnetEg9.2)

plot(CharnetEg9.2)
cor(CharnetEg9.2)

}
\keyword{RM}

