% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx1.20.R
\name{CharnetEx1.20}
\alias{CharnetEx1.20}
\title{Testes de Avaliação de Personalidade}
\format{Um \code{data.frame} com 2 colunas e 20 linhas.

\describe{

\item{\code{t1}}{Resultados do teste I, em pontos.}

\item{\code{t2}}{Resultados do teste II, em pontos.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Capítulo 1, exercício 20,
    página 26)
}
\description{
Dois testes de avaliação de personalidade aplicados a 20
    indivíduos.
}
\examples{

data(CharnetEx1.20)

plot(CharnetEx1.20)

}
\keyword{RS}

