% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx11.3.R
\name{CharnetEx11.3}
\alias{CharnetEx11.3}
\title{Relação do Preço de Venda de Imóvel e suas Características}
\format{Um \code{data.frame} com 5 colunas e 20 linhas.

\describe{

\item{\code{preco}}{Preço de venda do imóvel, em reais.}

\item{\code{area}}{Área total de construção, em m\eqn{^2}.}

\item{\code{tempo}}{Tempo de construção, em anos.}

\item{\code{nquar}}{Número de quartos.}

\item{\code{pisc}}{Fator que indica se há piscina (\code{S}) ou não
    (\code{N}) no imóvel.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 11, exercício
    3, pág. 274)
}
\description{
Estudo observacional onde o interesse é explicar o valor
    de venda de imóveis de um mesmo bairro a partir de suas
    características presença de piscina, tempo de construção, área
    total e número de quartos.
}
\examples{

data(CharnetEx11.3)

plot(CharnetEx11.3)

}
\keyword{RM}

