% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx2.11}
\alias{CharnetEx2.11}
\title{Estudo do Efeito de Droga sobre o Ritmo Card\enc{í}{i}aco}
\format{Um \code{data.frame} com 3 colunas e 6 linhas, em que

\describe{

\item{\code{dia}}{Dia do experimento, em contagem.}

\item{\code{dose}}{Número de doses aplicada no dia.}

\item{\code{bpm}}{Número de batimentos cardíacos por minuto.}

}}
\source{
CHARNET et al. (2008), Capítulo 2, exercício 11, pág. 48.
}
\description{
Estudo do efeito de droga sobre o ritmo cardíaco em um
    paciente voluntário por um período de 6 dias.
}
\examples{

data(CharnetEx2.11)
str(CharnetEx2.11)

plot(CharnetEx2.11)

}
\keyword{RM}
