% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx2.13.R
\name{CharnetEx2.13}
\alias{CharnetEx2.13}
\title{Compara\enc{çã}{ca}o entre Alturas de Pais e Filhos}
\format{Um \code{data.frame} com 2 colunas e 7 linhas.

\describe{

\item{\code{pai}}{Altura do pai, em centímetros.}

\item{\code{filhos}}{Altura do filho, em centímetros.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 2, exercício
    13, pág. 48)
}
\description{
Comparação entre alturas de 12 pais e respectivos 12
   filhos.
}
\examples{

data(CharnetEx2.13)

plot(CharnetEx2.13)

}
\keyword{RS}

