% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx2.14}
\alias{CharnetEx2.14}
\title{Produto Interno Bruto Trimestral do Brasil}
\format{Uma série temporal \code{ts}, com 32 observações, sendo 4
    observações por ano (trimestralmente) de 1990 a 1997.}
\source{
CHARNET et al. (2008), (Capítulo 2, exercício
    14, pág. 49)
}
\description{
Comparação do Produto Interno Bruto (PIB) trimestral do
    Brasil em 8 anos, de 1990 a 1997.
}
\examples{

data(CharnetEx2.14)
str(CharnetEx2.14)

CharnetEx2.14

plot(CharnetEx2.14, type = "o")

}
\keyword{TS}

