% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx2.8.R
\name{CharnetEx2.8}
\alias{CharnetEx2.8}
\title{Dados Genéricos para Regressão Linear Simples}
\format{Um \code{data.frame} com 3 colunas e 34 linhas.

\describe{

    \item{\code{conj}}{Fator que indica a qual conjunto a observação
    pertence.}

    \item{\code{x}}{Variável independente, sem interpretação.}

    \item{\code{y}}{Variável dependente, sem interpretação.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 2, exercício
    8, pág. 47)
}
\description{
Três conjuntos de dados apresentados para
    exercício. Objetivo do exercício é visualizar casos em que só
    o termo linear não é satisfatório.
}
\details{
Este conjunto de dados agrupa dados onde pressupõe-se que
    sejam de populações distintas. Portanto ao utilizá-lo separe-o
    com base na variável \code{conj}.
}
\examples{

data(CharnetEx2.8)

library(ggplot2)

ggplot(CharnetEx2.8, aes(x = x, y = y))
    geom_point()
    facet_grid(~conj)
    stat_smooth(method = "lm")

}
\keyword{RS}

