% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx3.4.R
\name{CharnetEx3.4}
\alias{CharnetEx3.4}
\title{Compara\enc{çã}{ca}o da Velocidade M\enc{á}{a}xima e Peso de Carros de Corrida}
\format{Um \code{data.frame} com 2 colunas e 24 linhas.

\describe{

\item{\code{peso}}{Peso do veículo, em quilogramas.}

\item{\code{velo}}{Velocidade máxima, em km/h .}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 3, exercício
    4, pág. 66)
}
\description{
Dados provenientes de um estudo onde se observou a
    velocidade máxima que carros de corrida de uma certa categoria
    conseguem atingir durante o percurso em comparação com seu peso.
}
\examples{

data(CharnetEx3.4)

plot(CharnetEx3.4)

}
\keyword{RS}

