% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx3.9}
\alias{CharnetEx3.9}
\title{Consumo de Combust\enc{í}{i}vel por Velocidade}
\format{Um \code{data.frame} com 2 colunas e 28 linhas, em que

\describe{

\item{\code{velo}}{Velocidade, em milhas por hora}

\item{\code{cons}}{Consumo de combustível, em milhas por galão}

}}
\source{
CHARNET et al. (2008), Capítulo 3, exercício 9, pág. 68.
}
\description{
Teste feito com um certo tipo de carro, comparando o
    consumo de combustível (em milhas por galão) e sua velocidade
    (em milhas por hora).
}
\examples{

data(CharnetEx3.9)
str(CharnetEx3.9)
xtabs(~velo, data = CharnetEx3.9)

plot(CharnetEx3.9)

}
\keyword{RS}
