% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx4.10.R
\name{CharnetEx4.10}
\alias{CharnetEx4.10}
\title{Custo de Produção por Tamanho do Lote}
\format{Um \code{data.frame} com 2 colunas e 10 linhas.

\describe{

\item{\code{npecas}}{Quantidade de peças produzidas no lote, em
unidades.}

\item{\code{custo}}{Custo do total de peças do lote.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 4, exer. 10, p. 85,
    cap. 6, exer. 10, p. 146.)
}
\description{
Custo de produção pelo número de peças produzidas em uma
    amostra de 10 lotes.
}
\examples{

data(CharnetEx4.10)

plot(CharnetEx4.10)

}
\keyword{RS}

