% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx4.8}
\alias{CharnetEx4.8}
\title{Taxa de Acidentes de Trabalho}
\format{Um \code{data.frame} com 2 colunas e 9 linhas, em que

\describe{

\item{\code{ano}}{Ano em que foi registrado a taxa.}

\item{\code{taxa}}{Taxa de acidentes de trabalho, em milhão de
    horas/homem.}

}}
\source{
CHARNET et al. (2008), Capítulo 4, exercício
    8, pág. 84; Capítulo 6, exercício 9, pág. 146.
}
\description{
Taxa de acidentes de trabalho por milhão de horas/homem
    de exposição ao risco, durante 9 anos.
}
\examples{

data(CharnetEx4.8)
str(CharnetEx4.8)

plot(CharnetEx4.8, type = "o")

}
\keyword{RS}

