% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx5.10}
\alias{CharnetEx5.10}
\title{Rela\enc{çã}{ca}o entre a Taxa de Desemprego e \enc{Í}{I}ndice de Suic\enc{í}{i}dios}
\format{Um \code{data.frame} com 2 colunas e 14 linhas.

\describe{

\item{\code{ano}}{Ano do registro.}

\item{\code{des}}{Taxa de desemprego.}

\item{\code{su}}{Índice de suicídio, para 1000 habitantes.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 5, exercício
    10, pág. 112)
}
\description{
Dados do jornal Los Angeles Times, de 13 de dezembro de
    1980, informando as taxas de desemprego e o índice de suicídios
    nos EUA em 14 diferentes anos, no período de 1950 a 1977.
}
\examples{

data(CharnetEx5.10)

plot(CharnetEx5.10)

}
\keyword{RS}

