% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx5.11.R
\name{CharnetEx5.11}
\alias{CharnetEx5.11}
\title{Relação do Lucro de uma Loja e Gastos com Publicidade}
\format{Um \code{data.frame} com 2 colunas e 19 linhas.

\describe{

\item{\code{lucro}}{Lucro, em milhares de reais.}

\item{\code{gastos}}{Gasto em publicidade, em milhares de reais.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 5, exer. 11, p. 112,
    cap. 6, exer. 5, p. 143.)
}
\description{
Relação do lucro de uma loja de eletrônicos e seu gasto
    com publicidade num período de 19 meses.
}
\examples{

data(CharnetEx5.11)

plot(CharnetEx5.11)

}
\keyword{RS}

