% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx5.13.R
\name{CharnetEx5.13}
\alias{CharnetEx5.13}
\title{Danos aos Pulm\enc{õ}{o}es em Pacientes com Enfisema}
\format{Um \code{data.frame} com 2 colunas e 16 linhas.

\describe{

\item{\code{anos}}{Tempo, em anos, que o paciente fumou.}

\item{\code{danos}}{Danos gerados nos pulmões, em escala de 0 a 100.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 5, exercício
    1, pág. 113)
}
\description{
Danos aos pulmões em 16 pacientes com enfisema pulmonar
    conforme o número de anos que ele fumou.
}
\examples{

data(CharnetEx5.13)

plot(CharnetEx5.13)

}
\keyword{RS}

