% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx5.3}
\alias{CharnetEx5.3}
\title{Precis\enc{ã}{a}o de um Veloc\enc{í}{i}metro}
\format{Um \code{data.frame} com 2 colunas e 8 linhas, em que

\describe{

\item{\code{vr}}{Velocidade real, em km/h.}

\item{\code{va}}{Velocidade anotada no velocímetro, em km/h.}

}}
\source{
CHARNET et al. (2008), Capítulo 5, exercício 3, pág. 109.
}
\description{
Medição da precisão de um velocímetro de locomotiva.
}
\examples{

data(CharnetEx5.3)
str(CharnetEx5.3)

plot(CharnetEx5.3, asp = 1)
abline(a = 0, b = 1, lty = 2)

}
\keyword{RS}

