% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx5.5}
\alias{CharnetEx5.5}
\title{Respostas a um An\enc{ú}{u}ncio de Emprego}
\format{Um \code{data.frame} com 2 colunas e 14 linhas, em que

\describe{

\item{\code{linhas}}{Número de linhas do anúncio de emprego.}

\item{\code{resp}}{Respostas obtidas com o anúncio.}

}}
\source{
CHARNET et al. (2008), Capítulo 5, exercício
    5, pág. 110; Capítulo 6, exercício 4, pág. 143.
}
\description{
Relação da resposta a um anúncio de emprego com o número
    de linhas do anúncio.
}
\examples{

data(CharnetEx5.5)
str(CharnetEx5.5)

plot(CharnetEx5.5)

}
\keyword{RS}
