% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx5.6}
\alias{CharnetEx5.6}
\title{Consumo de Combust\enc{í}{i}vel e Peso do Autom\enc{ó}{o}vel}
\format{Um \code{data.frame} com 2 colunas e 14 linhas, em que

\describe{

\item{\code{peso}}{Peso do veículo, em kg.}

\item{\code{cons}}{Consumo do veículo num determinado trecho,
   em litros.}

}}
\source{
CHARNET et al. (2008), Capítulo 5, exercício 6, pág. 110.
}
\description{
Consumo de combustível para percorrer determinado trecho
    conforme o peso dos veículos de passeio, de mesmo ano,
    selecionados aleatoriamente de uma grande empresa.
}
\examples{

data(CharnetEx5.6)
str(CharnetEx5.6)

plot(CharnetEx5.6)

}
\keyword{RS}
