% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx6.3}
\alias{CharnetEx6.3}
\title{Sal\enc{á}{a}rio Mensal de Formandos em Economia}
\format{Um \code{data.frame} com 2 colunas e 35 linhas, em que

\describe{

\item{\code{ano}}{Ano do registro do salário.}

\item{\code{salario}}{Salário mensal.}

}}
\source{
CHARNET et al. (2008), Capítulo 6, exercício 3, pág. 143.
}
\description{
Pesquisa de uma faculdade de economia sobre a evolução
    do salário mensal de 5 de seus formandos no período de 87 a 93.
}
\examples{

data(CharnetEx6.3)
str(CharnetEx6.3)

plot(CharnetEx6.3)

}
\keyword{RS}
