% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx6.6}
\alias{CharnetEx6.6}
\title{Conjunto de Dados Gen\enc{é}{e}rico}
\format{Um \code{data.frame} com 2 colunas e 15 linhas.

\describe{

\item{\code{x}}{Variável independente, sem interpretação.}

\item{\code{y}}{Variável dependente, sem interpretação.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 6, exercício
    6, pág. 144)
}
\description{
Conjunto de dados qualquer para exercício de ajuste de
    um modelo de regressão linear simples, com transformação na
    variável resposta e retirando os pontos considerados
    discrepantes.
}
\examples{

data(CharnetEx6.6)

plot(CharnetEx6.6)

}
\keyword{RS}

