% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx6.7}
\alias{CharnetEx6.7}
\title{Peso de Correspond\enc{ê}{e}ncias}
\format{Um \code{data.frame} com 2 colunas e 11 linhas.

\describe{

\item{\code{peso}}{Peso total das correspondências no dia, em kg.}

\item{\code{cor}}{Total de correspondências no dia, em milhares.}

}}
\source{
CHARNET et al. (2008), Capítulo 6, exercício 7, pág. 145.
}
\description{
Relação do peso das correspondências levantados por uma
    agência de correios durante 11 dias.
}
\examples{

data(CharnetEx6.7)
str(CharnetEx6.7)

plot(CharnetEx6.7)

}
\keyword{RS}
