% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx7.1}
\alias{CharnetEx7.1}
\title{Efeito da Radia\enc{çã}{ca}o ao Oz\enc{ô}{o}nio em Sementes de
    Soja}
\format{Um \code{data.frame} com 3 colunas e 12 linhas, em que

\describe{

\item{\code{n}}{Nível de ozônio, fator com níveis \code{a} ou
    \code{b}.}

\item{\code{rad}}{Quantidade de radiação solar aplicada, unidade de
    medida não informada.}

\item{\code{peso}}{Peso das sementes de soja, unidade de medida não
    informada.}

}}
\source{
CHARNET et al. (2008), Capítulo 7, exercício 1, pág. 163;
    Capítulo 10, exercício 1, pág. 253.
}
\description{
Efeito da radiação solar em dois níveis de ozônio e
    impacto nos pesos de sementes de soja.
}
\examples{

data(CharnetEx7.1)
str(CharnetEx7.1)

with(CharnetEx7.1, plot(peso, rad, col = n))

}
\keyword{RS}
\keyword{dummy}

