% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx7.2.R
\name{CharnetEx7.2}
\alias{CharnetEx7.2}
\title{Impacto de Impurezas em um Reator Químico}
\format{Um \code{data.frame} com 4 colunas e 20 linhas.

\describe{

\item{\code{rea}}{Reator analisado, um fator com dois níveis \code{I}
    ou \code{II}.}

\item{\code{t}}{Tempo, em minutos.}

\item{\code{imp}}{Porcentagem de impurezas, em escala \eqn{\log}.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 7, exer. 2, p. 163,
    Cap 10, exer. 2, p. 253)
}
\description{
Relação entre a porcentagem de impurezas dentro de um 
    reator químico e o tempo de permanência da substância dentro 
    deste reator.
}
\examples{

data(CharnetEx7.2)

with(CharnetEx7.2, plot(tempo ~ imp, col = rea))

}
\keyword{RM}

