% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx8.1.R
\name{CharnetEx8.1}
\alias{CharnetEx8.1}
\title{Influência da Publicidade e Capital Investido no Lucro Anual}
\format{Um \code{data.frame} com 3 colunas e 12 linhas.

\describe{

\item{\code{lucro}}{Lucro anual, em 100 mil reais.}

\item{\code{capi}}{Capital investido no ano, em 100 mil reais.}

\item{\code{publi}}{Gasto com publicidade no ano, em 100 mil reais.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 8, exer. 1, p. 195,
    cap. 9, exer. 6, p. 230.)
}
\description{
Influência das variáveis capital investido (\code{capi})
    e gasto em publicidade (\code{publi}) no lucro anual
    (\code{lucro}) de 12 empresas observadas.
}
\examples{

data(CharnetEx8.1)

plot(CharnetEx8.1)
lm(lucro ~ ., data = CharnetEx8.1)

}
\keyword{RM}

