% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx8.2.R
\name{CharnetEx8.2}
\alias{CharnetEx8.2}
\title{Relação da Temperatura e Produção de um Produto Químico}
\format{Um \code{data.frame} com 2 colunas e 11 linhas.

\describe{

\item{\code{temp}}{Temperatura codificada.}

\item{\code{peso}}{Peso do produto, em kg.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 8, exer. 2, p. 195,
    cap. 9, exer. 11, p. 233.)
}
\description{
Uma indústria química está interessada em maximizar a
    sua produção de um certo processo químico. Para isso, obteve os
    seguintes resultados da temperatura codificada, \code{temp}, e do
    peso do produto, \code{peso}, em kg.
}
\examples{

data(CharnetEx8.2)

plot(CharnetEx8.2)

}
\keyword{RP}
\keyword{RegSeg}

