% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx8.4.R
\name{CharnetEx8.4}
\alias{CharnetEx8.4}
\title{Peças Defeituosas por Produção Média e Tempo de Reparo}
\format{Um \code{data.frame} com 3 colunas e 15 linhas.

\describe{

\item{\code{nitens}}{Itens produzidos com defeito num dia, em peças.}

\item{\code{prod}}{Produção média por hora, em peças.}

\item{\code{tempo}}{Tempo, em semanas, decorrido do último reparo da
    máquina.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 8, exercício
    4, pág. 196, Capítulo 9, exercício 7, pág. 230)
}
\description{
Relação da quantidade de peças defeituosas por produção
    média e tempo decorrido desde o último reparo da máquina.
}
\examples{

data(CharnetEx8.4)

plot(CharnetEx8.4)

}
\keyword{RM}

