% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Costa.R
\name{CostaEx5.7.3}
\alias{CostaEx5.7.3}
\title{Efeito de Aduba\enc{çã}{ca}o Nitrogenada na Cultura do Milho}
\format{Um \code{data.frame} com 15 observações e 3 variáveis.

\describe{

\item{\code{dose}}{Quantidade aplicada de adubação nitrogenada (kg
    ha\eqn{^{-1}})}

\item{\code{rept}}{Identifica as repetições de cada dose.}

\item{\code{prod}}{Produção de milho (ton ha\eqn{^{-1}}).}

}}
\source{
Costa, J. R. (2003). Técnicas experimentais aplicadas às
    ciências agrárias (Documentos 163). Seropédica, RJ: Embrapa
    Agrobiologia. (Exemplo 5.7.3.1, pág. 95)
}
\description{
Experimento instalado em delineamento inteiramente
    casualizado que mediu o efeito da adubação nitrogenada na
    produção de milho.
}
\examples{

library(lattice)

data(CostaEx5.7.3)

str(CostaEx5.7.3)

xyplot(prod ~ dose, data = CostaEx5.7.3,
       type = c("p", "smooth", "g"),
       xlab = expression("Dose de nitrogênio"~(kg~ha^{-1})),
       ylab = expression("Produção de milho"~(ton~ha^{-1})))

}
\keyword{DIC}
\keyword{RegSeg}

