% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CostaTb4.R
\name{CostaTb4}
\alias{CostaTb4}
\title{Massa seca de parte aérea em cana-de-açúcar}
\format{Um \code{data.frame} com 24 observações e 3 variáveis.

\describe{

\item{\code{varied}}{Fator nominal. Variedades de cana-de-açúcar.}

\item{\code{rept}}{Inteiro que representa a unidade experimental em
    cada variedade.}

\item{\code{mspa}}{Variável resposta observada (contínua) que é a
    massa seca de parte aérea de cada unidade experimental (g
    parcela\eqn{^{-1}})}

}}
\source{
Costa, J. R. (2003). Técnicas experimentais aplicadas às
    ciências agrárias (Documentos 163). Seropédica: Embrapa
    Agrobiologia. (pg 58)
}
\description{
Resultados de um experimento conduzido em casa de
    vegetação em delineamento inteiramente casualizado para avaliar a
    massa seca de parte aérea de variedades de cana-de-açúcar. A
    unidade experimental (parcela) era um vaso com 3 plantas.
}
\examples{
library(lattice)

data(CostaTb4)

aggregate(mspa ~ varied,  data = CostaTb4,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

xyplot(mspa ~ varied, data = CostaTb4,
       xlab = "Variedade de cana-de-açúcar",
       ylab = expression(Massa~seca~de~parte~aérea~(g~parcela^{-1})))
}
\keyword{DIC}

