% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Costa.R, R/CostaTb8.R
\name{CostaTb8}
\alias{CostaTb8}
\title{Irriga\enc{çã}{ca}o no Tamanho de Frutos de Variedades de Banana}
\format{Um \code{data.frame} com 32 observações e 4 variáveis.

\describe{

\item{\code{varied}}{Fator de níveis nominais que representa as
    variedades de banana. As variedade de banana é o fator com níveis
    casualizados nas parcelas dentro dos blocos.}

\item{\code{irrig}}{Fator de níveis número que indica o número de
    linhas de irrigação recebidas pelas plantas, 1 ou 2 linhas. A
    irrigação é o fator casualizado depois das variedades e dentro
    das parcelas que são de alguma das variedades. Portato, é o fator
    com níveis nas subparcelas.}

\item{\code{bloco}}{Fator de níveis nominais que representam os
    blocos do experimento.}

\item{\code{comp}}{Comprimento (cm) do fruto central da terceira
    penca do cacho de bananas.}

}}
\source{
Costa, J. R. (2003). Técnicas experimentais aplicadas às
    ciências agrárias (Documentos 163). Seropédica, RJ: Embrapa
    Agrobiologia. (Tabela 8, pág. 76)

Costa, J. R. (2003). Técnicas experimentais aplicadas às
    ciências agrárias (Documentos 163). Seropédica, RJ: Embrapa
    Agrobiologia. (Tabela 8, pág. 76)
}
\description{
Experimento em parcelas subdivididas que estudou o
    comprimento de frutos de variedades de banana em função do nível
    de irrigação fornecido.

Experimento em parcelas subdivididas que estudou o
    comprimento de frutos de variedades de banana em função do nível
    de irrigação fornecido.
}
\examples{

library(lattice)

data(CostaTb8)

str(CostaTb8)

ftable(with(CostaTb8,
            tapply(comp,
                   list(varied = varied,
                        irrig = irrig,
                        bloco = bloco),
                   FUN = mean)))

xyplot(comp ~ varied, groups = irrig, data = CostaTb8,
       as.table = TRUE, layout = c(NA, 1),
       jitter.x = TRUE, type = c("p", "a"),
       auto.key = list(title = "Linhas de irrigação",
                       cex.title = 1, columns = 2),
       xlab = "Variedade de banana",
       ylab = expression("Comprimento do fruto"~(cm)))


library(lattice)

data(CostaTb8)

str(CostaTb8)

ftable(with(CostaTb8,
            tapply(comp,
                   list(varied = varied,
                        irrig = irrig,
                        bloco = bloco),
                   FUN = mean)))

xyplot(comp ~ varied, groups = irrig, data = CostaTb8,
       as.table = TRUE, layout = c(NA, 1),
       jitter.x = TRUE, type = c("p", "a"),
       auto.key = list(title = "Linhas de irrigação",
                       cex.title = 1, columns = 2),
       xlab = "Variedade de banana",
       ylab = expression("Comprimento do fruto"~(cm)))

}
\keyword{DBC}
\keyword{PS}

