% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioEx1.4.1.2.R
\name{DemetrioEx1.4.1.2}
\alias{DemetrioEx1.4.1.2}
\title{Peso Médio de Galinhas}
\format{Um \code{data.frame} de 10 linhas e 2 colunas.

\describe{

    \item{\code{peso}}{Peso médio.}
    
    \item{\code{consumo}}{Consumo de alimentos.}
    
}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Exercício 1.4.1.2 pág. 14)
}
\description{
Foi mensurado o peso médio e consumo de alimentos de 50
    galinhas para 10 linhagens \emph{White Leghorn}.
}
\examples{

data(DemetrioEx1.4.1.2)

library(lattice)

xyplot(consumo ~ peso, data = DemetrioEx1.4.1.2,
     main = "Peso VS Consumo",
     xlab = "Peso",
     ylab = "Consumo",
     type = c("p", "r"), col.line = 3)
     
}
\keyword{RS}

