% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioEx1.4.1.4}
\alias{DemetrioEx1.4.1.4}
\title{Volume das Cerejeiras}
\format{Um \code{data.frame} de 31 linhas e 3 colunas, em que

\describe{

\item{\code{dia}}{Diâmetro da cerejeira a 4.5 pés do solo, medido em
    polegadas.}

\item{\code{alt}}{Altura das cerejeiras, medida em pés.}

\item{\code{vol}}{Volume das cerejeiras, medido em pés cúbicos.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011) (Exercício 1.4.1.4 pág. 14)
}
\description{
Foram mensurados o diâmetro, a altura e o volume de 31
    cerejeiras com o objetivo de verificar a relação entre estas
    variáveis, tendo em vista a predição do volume de madeira em uma
    área de floresta.
}
\examples{

data(DemetrioEx1.4.1.4)
str(DemetrioEx1.4.1.4)

library(lattice)

pairs(~ dia + alt + vol, data = DemetrioEx1.4.1.4,
      main = "Gráfico de Pares")

xyplot(vol ~ dia, data = DemetrioEx1.4.1.4,
       xlab = "Diâmetro",
       ylab = "Volume",
       type = c("p", "r"), col.line = 3)

}
\keyword{RM}
