% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioEx1.4.1.5}
\alias{DemetrioEx1.4.1.5}
\title{N\enc{ú}{u}mero de Ovos por Fol\enc{í}{i}culos Ovulados}
\format{Um \code{data.frame} de 14 linhas e 2 colunas, em que

\describe{

\item{\code{ovo}}{Número de ovos.}

\item{\code{foli}}{Número de folículos.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Exercício 1.4.1.5 pág. 15.
}
\description{
Foi contado o número de ovos postos e o número de
    folículos ovulados.
}
\examples{

data(DemetrioEx1.4.1.5)
str(DemetrioEx1.4.1.5)

library(lattice)

xyplot(foli ~ ovo, data = DemetrioEx1.4.1.5,
       xlab = "Número de ovos",
       ylab = "Número de folículos",
       type = c("p", "r"), col.line = 3)

}
\keyword{RS}
