% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioEx2.12.15.R
\name{DemetrioEx2.12.15}
\alias{DemetrioEx2.12.15}
\title{Valores Simulados}
\format{Um \code{data.frame} de 6 linhas e 2 colunas.

\describe{

    \item{\code{x}}{Sem interpretação.}
    
    \item{\code{y}}{Sem interpretação.}
    
}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Exercício 2.12.15 pág. 63)
}
\description{
Os dados estão presentes no exercício com o objetivo de
    desenhar o gráfico de dispersão e realizar o ajuste do modelo.
}
\examples{

data(DemetrioEx2.12.15)

library(lattice)

xyplot(y ~ x, data = DemetrioEx2.12.15,
     main = "x vs y",
     xlab = "x",
     ylab = "y",
     type = c("p", "r"), col.line = 3)
     
}
\keyword{TODO}

