% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioEx5.4.2.R, R/DemetrioEx5.4.5.R
\name{DemetrioEx5.4.5}
\alias{DemetrioEx5.4.5}
\title{Méis Silvestres}
\format{Um \code{data.frame} de 94 linhas e 3 colunas.

\describe{

    \item{\code{condut}}{Condutividade elétrica, em \eqn{\mu}S.}
    
    \item{\code{n}}{Proporção de Nitrogênio proteico.}
    
    \item{\code{cinzas}}{Proporção de cinzas.}
    
}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Exercício 5.4.2 pág. 169)

Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Exercício 5.4.5 pág. 165)
}
\description{
Análise físico química de méis silvestres, produzidos por 
    Apis mellifera em 1999, provenientes de 94 localidades de
    São Paulo.

Dados referentes a avaliação visual (realizada por um
    pesquisador) e real do grau de infestação de folhas de amendoim.
    
    A avaliação exige um treinamento específico para a cultura e
    a doença em questão. Por esse motivo, foram desenvolvidos
    programas computacionais que geram imagens de folhas com 
    diferentes porcentagens de infestação para o pesquisador estimar
    visualmente e em seguida mostram as porcentagens reais.
}
\examples{

data(DemetrioEx5.4.2)

pairs(~ condut + n + cinzas, data = DemetrioEx5.4.2,
    main = "Gráfico de Pares")
     

data(DemetrioEx5.4.5)

library(lattice)

xyplot(visual ~ real, data = DemetrioEx5.4.5,
     main = "Real vs Visual",
     xlab = "Real",
     ylab = "Visual",
     type = c("p", "r"), col.line = 3)
     
}
\keyword{RS}
\keyword{TODO}

