% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DemetrioTb1.2.R
\name{DemetrioTb1.2}
\alias{DemetrioTb1.2}
\title{Irrigação em Batata}
\format{Um \code{data.frame} com 12 linhas e 2 colunas.

\describe{

    \item{\code{lamina}}{Espessura da lâmina de água medida em
    milímetros (mm).}

    \item{\code{prod}}{Produtividade medida em toneladas de batatas
     por hectare (t/ha).}

}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 1.2 pág. 9)
}
\description{
Este experimento refere-se a irrigação em batata
    plantada em terra roxa estruturada. Foram medidas as lâminas de
    água a diferentes distâncias do aspersor. O objetivo foi mensurar
    a produtividade, já que, no tipo de solo utilizado no experimento
    (solo argiloso), espera-se que o excesso de água diminua a
    produtividade.
}
\examples{
library(lattice)

data(DemetrioTb1.2)

xyplot(prod ~ lamina, data = DemetrioTb1.2,
       main = "Produção VS Lâmina de Água",
       xlab = "Lâmina de água (mm)",
       ylab = "Produção (t/ha)",
       type = c("p", "r"), col.line = 3)
}
\keyword{aspersor}
\keyword{batata}
\keyword{produtividade}

