% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioTb1.4}
\alias{DemetrioTb1.4}
\title{Concentra\enc{çã}{ca}o de F\enc{ó}{o}sforo}
\format{Um \code{data.frame} com 18 linhas e 3 colunas, em que

\describe{

\item{\code{cfi}}{ Conteúdo de fósforo inorgânico no solo.}

\item{\code{cfo}}{ Conteúdo de fósforo orgânico no solo.}

\item{\code{conteudo}}{ Conteúdo de fósforo nas plantas crescidas no
    solo.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Tabela 1.4 pág. 11.
}
\description{
Dados referentes a medidas de concentrações de fósforo
    inorgânico e orgânico no solo, onde posteriormente mediu-se o
    conteúdo de fósforo nas plantas crescidas no local. O objetivo do
    experimento é estudar a relação existente entre o conteúdo de
    fósforo na planta e as duas fontes do elemento no solo.
}
\examples{

data(DemetrioTb1.4)
str(DemetrioTb1.4)

pairs(DemetrioTb1.4, main = "Dispersão em Pares")

library(lattice)

xyplot(conteudo ~ cfi, data = DemetrioTb1.4,
       xlab = "Fósforo Inorgânico",
       ylab = "Conteúdo na Planta",
       type = c("p", "r"), col.line = 3)

xyplot(conteudo ~ cfo, data = DemetrioTb1.4,
       xlab = "Fósforo Orgânico",
       ylab = "Conteúdo na Planta",
       type = c("p", "r"), col.line = 3)

}
\keyword{RM}

