% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioTb1.6.R
\name{DemetrioTb1.6}
\alias{DemetrioTb1.6}
\title{Resposta de Milho ao Fosfato}
\format{Um \code{data.frame} de 14 linhas e 4 colunas.

\describe{

    \item{\code{milho}}{Dados de resposta da cultura do milho ao
    fosfato em porcentagem.}
    
    \item{\code{prod}}{Produtividade da cultura na parcela
    testemunha, em lb/acre.}
    
    \item{\code{satu}}{Saturação de bases em porcentagem.}
    
    \item{\code{ph}}{pH do solo.}

}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 1.6 pág. 13)
}
\description{
Neste estudo foram obtidos dados sobre a resposta da
    cultura do milho ao fosfato, produtividade na testemunha,
    porcentagem de saturação de bases e pH do solo. Como todas as
    variáveis computadas são aleatórias, o interesse principal é em
    estudar as correlações entre as variáveis observadas.
}
\examples{

data(DemetrioTb1.6)

pairs(~ milho + prod + satu + ph, data = DemetrioTb1.6,
     main = "Dispersão duas a duas")

cor(DemetrioTb1.6)

}
\keyword{TODO}

