% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioTb10.2}
\alias{DemetrioTb10.2}
\title{Estudo em Plantas Nicotianas}
\format{Um \code{data.frame} de 18 linhas e 3 colunas, em que

\describe{

\item{\code{caule}}{Comprimento do caule.}

\item{\code{ramo}}{Comprimento do ramo.}

\item{\code{basal}}{Comprimento do caule basal.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Tabela 10.2 pág. 161; Exercício
    5.4.7 pág. 164.
}
\description{
Dados referentes ao comprimento do caule, do ramo e do
    caule basal de plantas do gênero Nicotiana.
}
\examples{

data(DemetrioTb10.2)
str(DemetrioTb10.2)

pairs(~ caule + basal + ramo , data = DemetrioTb10.2)

}
\keyword{AnaFat}
\keyword{RM}
