% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioTb2.12.R
\name{DemetrioTb2.12}
\alias{DemetrioTb2.12}
\title{Produção de Ruibarbo}
\format{Um \code{data.frame} de 28 linhas e 3 colunas.

\describe{

    \item{\code{data}}{Data de colheita dos ruibarbos. Obs.: Para
    utilizar o formato \code{\link{Date}} foi considerado o ano 1983
    (Date de publicação do livro que referencia os dados).}
    
    \item{\code{bloco}}{Fator que indica o bloco, para controle de
    variação, ao qual a observação pertence.}
    
    \item{\code{prod}}{Valor da produção de ruibarbo.}
    
}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 2.12 pág. 66-67)
}
\description{
Dados de um experimento conduzido em delineamento de
    blocos ao acaso onde foi mensurada a produção de ruibarbos para
    enlatamento, considerando diferentes datas de colheita.
}
\examples{

data(DemetrioTb2.12)

library(lattice)

# Estatísticas descritivas
with(DemetrioTb2.12, tapply(prod, data, summary))

with(DemetrioTb2.12, {
    mu <<- aggregate(prod, list(data), mean)
    des <<- aggregate(prod, list(data), sd)
})

xyplot(prod ~ data, data = DemetrioTb2.12,
       type = c("p", "r"), grid = TRUE,
       panel = function(x, y, ...) {
           panel.points(x = mu$G - 0.5, y = mu$x, pch = 15, col = 1)
           panel.arrows(x0 = mu$G - 0.5, y0 = mu$x - des$x,
                        x1 = mu$G - 0.5, y1 = mu$x + des$x,
                        code = 3, length = 0.05, angle = 90)
           panel.xyplot(x, y, ...)
           })

}
\keyword{TODO}

