% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioTb3.5}
\alias{DemetrioTb3.5}
\title{Dados Gen\enc{é}{e}ricos para Regress\enc{ã}{a}o
    M\enc{ú}{u}ltipla}
\format{Um \code{data.frame} de 6 linhas e 3 colunas, em que

\describe{

\item{\code{x1}}{Variável explicativa, sem interpretação.}

\item{\code{x2}}{Variável explicativa, sem interpretação.}

\item{\code{y}}{Variável dependente, sem interpretação.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Tabela 3.5 pág. 99.
}
\description{
Dados para exercício de análise via modelo de regressão
    linear múltipla.
}
\examples{

data(DemetrioTb3.5)
str(DemetrioTb3.5)

pairs(~ x1 + x2 + y , data = DemetrioTb3.5)

}
\keyword{RM}
