% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioTb3.6}
\alias{DemetrioTb3.6}
\title{Dados Gen\enc{é}{e}ricos para Regress\enc{ã}{a}o
    M\enc{ú}{u}ltipla}
\format{Um \code{data.frame} de 14 linhas e 4 colunas, em que

\describe{

\item{\code{x1}}{Variável explicativa, sem interpretação.}

\item{\code{x2}}{Variável explicativa, sem interpretação.}

\item{\code{x3}}{Variável explicativa, sem interpretação.}

\item{\code{y}}{Variável dependente, sem interpretação.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Tabela 3.6 pág. 99.
}
\description{
Dados para exercício de análise via modelo de regressão
    linear múltipla com três variáveis explicativas.
}
\examples{

data(DemetrioTb3.6)
str(DemetrioTb3.6)

pairs(~ x1 + x2 + x3 + y , data = DemetrioTb3.6)

}
\keyword{RM}

