% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioTb4.5.R
\name{DemetrioTb4.5}
\alias{DemetrioTb4.5}
\title{Dados simulados}
\format{Um \code{data.frame} de 9 linhas e 5 colunas.

\describe{

    \item{\code{x}}{ Formação indisponível. }

    \item{\code{y1}}{ Formação indisponível. }
    
    \item{\code{y2}}{ Formação indisponível. }
    
    \item{\code{y3}}{ Formação indisponível. }
    
    \item{\code{y4}}{ Formação indisponível. } 

}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 4.5 pág. 137)
}
\description{
Dados simulados
}
\examples{

data(DemetrioTb4.5)

pairs(~ x + y1 + y2 + y3 + y4 , data = DemetrioTb4.5,
     main = "Dispersão duas a duas")

}
\keyword{RM}

