% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioTb10.2.R, R/DemetrioTb5.1.R
\name{DemetrioTb5.1}
\alias{DemetrioTb5.1}
\title{Nicotiana}
\format{Um \code{data.frame} de 18 linhas e 3 colunas.

\describe{

    \item{\code{caule}}{Comprimento do caule.}

    \item{\code{ramo}}{Comprimento do ramo.}
    
    \item{\code{basal}}{Comprimento do caule basal.}
    
}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 10.2 pág. 161; Exercício 
    5.4.7 pág. 164)

Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 5.1 pág. 157; Exercício
    5.4.7 pág. 161; Exercício 5.4.7 pág. 167)
}
\description{
Dados referentes ao comprimento do caule,
    do ramo e do caule basal da Nicotiana.

Dados referentes a um estudo sobre a resposta da 
    cultura do milho em função da quantidade de fosfato, porcentagem
    de saturação de bases e sílica em solos ácidos.
    
    Neste estudo a variável resposta, que está em porcentagem,
    foi medida como a diferença entre as produções que receberam
    fosfato e as que não receberam, dividida pelas produções
    das parcelas que receberam fosfato, e multiplicado por 100. 
    Considerando-se esses dados, foi obtida a variável produtividade
    das parcelas que receberam fosfato, dada por 
    \eqn{Y_1 = X_1(1 + \frac{Y}{100})}.
}
\examples{

data(DemetrioTb10.2)

pairs(~ caule + basal + ramo , data = DemetrioTb10.2,
     main = "Dispersão duas a duas")


data(DemetrioTb5.1)

pairs(~ resp + prod + sat + silica , data = DemetrioTb5.1,
     main = "Dispersão duas a duas")

}
\keyword{RM}

