% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEg10.1}
\alias{DiasEg10.1}
\title{Subst\enc{â}{a}ncias Alelo\enc{á}{a}ticas no Percentual de
    Germina\enc{çã}{ca}o de Sementes}
\format{Um \code{data.frame} com 40 observações e 4 variáveis, em que

\describe{

\item{\code{cut}}{Fator de 2 níveis qualitativos, que representa a
    cultivar.}

\item{\code{trat}}{Fator de 5 níveis qualitativos, que representa
    tratamento com substância alelopática.}

\item{\code{rept}}{Inteiro que identifica as repetições.}

\item{\code{pg}}{Percentual de germinação das sementes.}

}}
\source{
BARROS; DIAS (2009), Exemplo 10.1, pág. 269.
}
\description{
Experimento em delineamento inteiramente casualizado
    onde foram avaliados os percentuais de germinação de sementes de
    2 cultivares, submetidas a 5 tratamentos com substâncias
    alelopáticas, sob esquema fatorial 2 \eqn{\times} 5 com 4
    repetições.
}
\examples{

library(lattice)

data(DiasEg10.1)
str(DiasEg10.1)

xtabs(~cult + trat, data = DiasEg10.1)

xyplot(pg ~ trat, groups = cult, data = DiasEg10.1,
       type = c("p", "a"),
       auto.key = list(title = "Cultivares", cex.title = 1.1,
                       columns = 2),
       xlab = "Tratamentos com substâncias alelopáticas",
       ylab = "Percentual de germinação")

}
\keyword{DIC}
\keyword{FAT2}
