% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEg3.2}
\alias{DiasEg3.2}
\title{Produ\enc{çã}{ca}o de frutos em cacaueiro}
\format{Vetor com 43 observações que são o número de frutos
    produzidos por plantas de cacaueiro.}
\source{
BARROS; DIAS (2009), Exemplo 3.2.
}
\description{
Amostra aleatória do número de frutos produzidos por
    plantas cacaueiro.
}
\examples{

stem(DiasEg3.2)

hist(DiasEg3.2)
rug(DiasEg3.2)

plot(density(DiasEg3.2))
rug(DiasEg3.2)

boxplot(DiasEg3.2)
rug(DiasEg3.2, side = 2)

}
\keyword{AAS}
