% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEg5.3}
\alias{DiasEg5.3}
\title{Produtividade de Cultivares de Milho}
\format{Um \code{data.frame} com 15 observações e 3 variáveis, em que

\describe{

\item{\code{cult}}{Fator categórico que representa as cultivares de
    milho.}

\item{\code{rept}}{Inteiro que indica as repetições das cultivares.}

\item{\code{prod}}{Produtividade de grãos (toneladas por hectare).}

}}
\source{
BARROS; DIAS (2009), Exercício 9, Cap. 11, pág. 321.
}
\description{
Resultados de um experimento em delineamento
    inteiramente casualizado que mediu a produtividade de cultivares
    de milho.
}
\examples{

library(lattice)

data(DiasEg5.3)
str(DiasEg5.3)

unstack(DiasEg5.3, prod ~ cult)

xyplot(prod ~ reorder(cult, prod), data = DiasEg5.3,
       type = c("p", "a"), jitter.x = TRUE,
       xlab = "Cultivares de milho",
       yalb = "Produtividade de grãos (ton/ha)")

}
\keyword{DIC}

