% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiasEg6.1.R
\name{DiasEg6.1}
\alias{DiasEg6.1}
\title{Ganhos de Produção de Arroz nos Ciclos de Seleção}
\format{Um \code{data.frame} com 6 observações e 2 variáveis, em que

\describe{

\item{\code{ciclos}}{Fator de 6 níveis quantitativos que são os
    ciclos de produção.}

\item{\code{prod}}{percentuais de ganhos de produção.}

}}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV. (Exemplo 6.1, pág. 157)
}
\description{
Amostra com 6 pares de observações representanto os
    ciclos de seleção e ganhos percentuais de produção
    correspondentes em arroz.
}
\examples{

library(lattice)

data(DiasEg6.1)
str(DiasEg6.1)

xyplot(prod ~ ciclos, data = DiasEg6.1, type = c("p", "r"),
       xlab = "Ciclos de produção",
       ylab = "Percentuais de ganhos de produção")

}
\keyword{DIC}

