% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiasEg6.3.R
\name{DiasEg6.3}
\alias{DiasEg6.3}
\title{Número Total de Frutos Colhidos e Sadios em Clones de
    Cacaueiro}
\format{Um \code{data.frame} com 7 observações e 3 variáveis, em que

\describe{

\item{\code{clones}}{Fator de 7 níveis qualitativos ordinais.}

\item{\code{ntfc}}{Número total de frutos colhidos.}

\item{\code{ntfs}}{Número total de frutos sadios.}

}}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV. (Exemplo 6.3, pág. 173)
}
\description{
Foram avaliados 7 clones de cacaueiros a fim determinar
    a correlação entre o número de frutos sadios e número de frutos
    colhidos.
}
\examples{

library(lattice)

data(DiasEg6.3)
str(DiasEg6.3)

xyplot(ntfs ~ ntfc, data = DiasEg6.3, type = c("p", "r"),
       ylab = "Número de frutos sadios ",
       xlab = "Número de frutos colhidos")

}
\keyword{AAS}

