% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEg7.1}
\alias{DiasEg7.1}
\title{Luz e \enc{Á}{A}gua na Produ\enc{çã}{ca}o de Tomateiros}
\format{Um \code{data.frame} com 18 observações e 3 variáveis, em que

\describe{

\item{\code{luz}}{Fator métrico que identifica, em escala codificada,
    a quantidade de luz aplicada.}

\item{\code{agua}}{Fator métrico que identifica, em escala
    codificada, a quantidade de água aplicada.}

\item{\code{prod}}{Produção de tomateiros.}

}}
\source{
BARROS; DIAS (2009), Exemplo 7.1, pág. 187.
}
\description{
Resultados de um experimento que avaliou o efeito da
    quantidade de luz e de água na produção de tomateiros.
}
\examples{

library(lattice)

data(DiasEg7.1)
str(DiasEg7.1)

xtabs(~luz + agua, data = DiasEg7.1)

xyplot(prod ~ luz, groups = agua, data = DiasEg7.1,
       type = c("p", "a"), jitter.x = TRUE,
       xlab = "Níveis de luz (codificados)",
       ylab = "Produção de tomateiros",
       auto.key = list(title = "Água (codificado)",
                       cex.title = 1.1, columns = 2))

}
\keyword{DIC}
\keyword{FAT2}
