% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEg9.1}
\alias{DiasEg9.1}
\title{Teor Proteico de Cultivares de Feijoeiro e Soja}
\format{Um \code{data.frame} com 33 observações e 3 variáveis, em que

\describe{

\item{\code{cult}}{Fator de 9 níveis qualitativos que representa cada
    um dos cultivares.}

\item{\code{rept}}{Inteiro que identifica as repetições.}

 \item{\code{teor}}{Teor proteico (\%).}

}}
\source{
BARROS; DIAS (2009), Exemplo 9.1, pág. 222.
}
\description{
Em um ensaio em delineamento inteiramente casualizado
    foi avaliado o teor proteico de 10 cultivares de feijoeiro e 1 de
    soja com 3 repetições.
}
\examples{

library(lattice)

data(DiasEg9.1)
str(DiasEg9.1)

xyplot(teor ~ reorder(cult, teor), data = DiasEg9.1,
       xlab = "Cultivares", ylab = "Teor proteico (\%)",
       scales = list(x = list(rot = 90)))

}
\keyword{DIC}
