% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEg9.2}
\alias{DiasEg9.2}
\title{Produ\enc{çã}{ca}o de Cacau}
\format{Um \code{data.frame} com 54 observações e 4 variáveis, em que

\describe{

\item{\code{bacia}}{Fator categórico que representa as bacias onde
    foram feitas as amostras.}

\item{\code{prog}}{Fator categórico que identifica as progênies
    obtidas em cada uma das bacias.}

\item{\code{plant}}{Inteiro que identifica as plantas de cada
    progênie.}

\item{\code{prod}}{Produção de cacau úmido (kg planta\eqn{^{-1}}).}

}}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV. (Exemplo 9.2, pág. 231)
}
\description{
Dados de produção de cacau úmido (kg/planta) obtidos de
    seis progênies formadas por 3 plantas cada e amostradas em 3
    bacias hidrográficas da Amazonia. Este é um exemplo de
    classificação hirerárquica de fatores decorrente do processo de
    amostragem em multinível.
}
\examples{

library(lattice)

data(DiasEg9.2)
str(DiasEg9.2)

xtabs(~bacia + prog, data = DiasEg9.2)

xyplot(prod ~ prog | bacia,
       data = DiasEg9.2, as.table = TRUE,
       xlab = "Progênies",
       ylab = "Produção de cacau úmido (kg/planta)",
       strip = strip.custom(strip.names = TRUE, var.name = "Bacia"))

}
\keyword{ClaHier}

