% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiasEx10.4.10.R
\name{DiasEx10.4.10}
\alias{DiasEx10.4.10}
\title{Massa fresca de Capim Elefante}
\format{Um \code{data.frame} com 72 observações e 4 variáveis, em que

\describe{

\item{\code{clon}}{Fator categórico que representa os clones de capim
    elefante.}

\item{\code{cort}}{Fator categórico que representa as épocas de corte
    do capim, sendo que A - sem corte, B - corte em 10 se setembro, C
    - corte em 30 de novembro e D - corte em 1 de novembro.}

\item{\code{blocos}}{Fator categórico que identifica os blocos do
    experimento.}

 \item{\code{tms}}{Produção de massa fresca de capim elefante
    (t/ha).}

}}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV. (Exercício 10, Cap. 10, pág. 296)
}
\description{
Dados refentes à produção de 3 clones
    de capim elefante em função da época de corte do capim. O
    experimento foi instalado em delineamento inteiramente
    casualizado.
}
\examples{

library(lattice)

data(DiasEx10.4.10)
str(DiasEx10.4.10)

xtabs(~cort + clon, data = DiasEx10.4.10)

xyplot(mfce ~ cort, groups = clon, data = DiasEx10.4.10,
       type = c("p", "a"),
       xlab = "Corte",
       ylab = "Massa fresca de capim elefante (ton/ha)",
       auto.key = list(title = "Clones", cex.title = 1.1,
                       columns = 3))

}
\keyword{DBC}
\keyword{FAT2}

