% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEx10.4.6}
\alias{DiasEx10.4.6}
\title{Ensaio de Competi\enc{çã}{ca}o de Batata Doce}
\format{Um \code{data.frame} com 12 observações e 3 variáveis, em que

\describe{

\item{\code{solo}}{Fator categórico que representa os tipos de solo.}

\item{\code{adub}}{Fator categórico que representa os níveis de
    adubação.}

\item{\code{prod}}{Produtividade.}
}}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV. (Exercício 10.4, Cap. 10,
    pág. 294)
}
\description{
Ensaio de competição de batata doce, instalado em
    delineamento de blocos casualizados com 3 repetições e sob
    esquema fatorial 3 \eqn{\times} 4, onde foram avaliados 3 tipos
    de solos e 4 níveis de adubação.
}
\details{
Não foram exibidos todos os dados, pois dos \eqn{4 \times 3
    \times 3 = 36} dados previstos, apenas 12 estão
    disponíveis. Julga-se que seja a média.
}
\examples{

library(lattice)

data(DiasEx10.4.6)
str(DiasEx10.4.6)

a <- reshape::cast(data = DiasEx10.4.6, solo ~ adub, value = "prod")
addmargins(as.matrix(a[, -1]))

xtabs(~solo + adub, data = DiasEx10.4.6)

xyplot(prod ~ adub, groups = solo,
       data = DiasEx10.4.6,
       type = "o",
       xlab = "Adubação", ylab = "Produção")

}
\keyword{DBC}

