% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dias.R
\name{DiasEx3.6.7}
\alias{DiasEx3.6.7}
\title{Percentual de Germina\enc{çã}{ca}o de Lotes de Sementes de
    Tomate}
\format{Um \code{data.frame} com 2 observações e 1 variável, em que

\describe{

\item{\code{lote}}{Fator categórico de dois níveis que indica os
    lotes de semente.}

\item{\code{pgerm}}{Percentual de germinação das sementes.}

}}
\source{
BARROS; DIAS (2009), (Exercício 7, Cap. 7.6, pág. 102.
}
\description{
Percentuais de germinação de 2 lotes de sementes de
     tomate com 5 repetições com 100 sementes em cada repetição.
}
\examples{

library(lattice)

data(DiasEx3.6.7)
str(DiasEx3.6.7)

xyplot(pgerm ~ lote,
       data = DiasEx3.6.7 ,
       xlab = "Lote",
       ylab = "Percentual germinação")

}
\keyword{DIC}
