% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiasEx6.5.1.R
\name{DiasEx6.5.1}
\alias{DiasEx6.5.1}
\title{Peso em Fun\enc{çã}{ca}o das Idades em Codornas}
\format{Um \code{data.frame} com 10 observações e 2 variáveis, em que

\describe{

\item{\code{idade}}{Idade do animal, em semanas.}

\item{\code{peso}}{Peso do animal (g).}

}}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV. (Exercício 1, Cap. 6, pág. 179)
}
\description{
Dados referentes ao peso em função da idade de codornas.
}
\examples{

library(lattice)

data(DiasEx6.5.1)
str(DiasEx6.5.1)

xyplot(peso ~ idade, data = DiasEx6.5.1, type = c("p", "r"),
       ylab = "Peso (g)", xlab = "Idade (semanas)")

}
\keyword{RL}

