% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiasEx6.5.10.R
\name{DiasEx6.5.10}
\alias{DiasEx6.5.10}
\title{Correla\enc{çã}{ca}o entre Produ\enc{çã}{ca}o e Di\enc{â}{a}metro}
\format{Um \code{data.frame} com 8 observações e 3 variáveis, em que

\describe{

\item{\code{prod}}{Produção em gramas.}

\item{\code{diam}}{Diãmetro em centimetros.}

}}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV. (Exercício 10, Cap. 7, pág. 181)
}
\description{
Dados referentes a 8 pares de valores de produção
    (gramas) e diâmetro (cm).
}
\examples{

library(lattice)

data(DiasEx6.5.10)
str(DiasEx6.5.10)

xyplot(prod ~ diam, data = DiasEx6.5.10, type = c("p", "r"),
       xlab = "Produção (gramas)", ylab = "Diâmetro (cm)")

}
\keyword{ASS}
\keyword{RL}

