% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpprechtTb5.2.R
\name{EpprechtTb5.2}
\alias{EpprechtTb5.2}
\title{Medidas de peças de uma linha de produção}
\format{Um \code{data.frame} com 60 observações e 3 variáveis, 
    em que

\describe{

\item{\code{pc}}{Identificação da peça.}

\item{\code{op}}{Fator que indica o operador que realizou a medição
    da peça (de 1 a 3).}
    
\item{\code{tam}}{Tamanho da peça (µm).}

}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estatístico de Qualidade (2nd ed.). São Paulo - SP:
     Editora Atlas. (pg 152).
}
\description{
Dados referentes às medidas de peças de uma linha de
    produção. O objetivo é analisar a repetitividade e a 
    reprodutividade de um micrômetro com leitura milesimal, usado na
    medição de um componente de um processo de usinagem. Três 
    operadores treinados mediram duas vezes cada uma de 10 peças. 
    A sequência em que cada um dos operadores mede cada uma das 
    peças foi aleatorizada.
}
\examples{

data(EpprechtTb5.2)

boxplot(tam~op, data=EpprechtTb5.2,
        xlab="Operário", 
        ylab="Tamanho", 
        main="Boxplot do tamanho das peças medidas pelos três 
            operários",
        col=c("#F0FFFF","#FFDAB9", "#C1FFC1"))

                  
}
\keyword{ASM}
\keyword{CEQ}

