% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpprechtTb5.9.R
\name{EpprechtTb5.9}
\alias{EpprechtTb5.9}
\title{Medidas de peças de uma linha de produção}
\format{Um \code{data.frame} com 60 observações e 3 variáveis, 
    em que

\describe{

\item{\code{pc}}{Identificação da peça.}

\item{\code{op}}{Fator que indica o operador que realizou a medição
    da peça (1 = primeiro operador, 2 = segundo operador).}
 
\item{\code{tam}}{Tamanho da peça.}

}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estatístico de Qualidade (2nd ed.). São Paulo - SP:
     Editora Atlas. (pg 157).
}
\description{
Após a aquisição de um equipamento de medição mais 
    sofisticado que o anterior, foi realizado um estudo de 
    repetitividade e reprodutividade do processo de medida com o 
    equipamento. Cada peça foi medida três vezes por cada um de
    dois operadores.
}
\examples{

data(EpprechtTb5.9)

EpprechtTb5.9$op <- as.factor(EpprechtTb5.9$op)

grid(nx=NA, ny=NULL)
par(new=TRUE)
boxplot(tam~op, data=EpprechtTb5.9,
        xlab="Operário", 
        ylab="Tamanho", 
        main="Boxplot do tamanho das peças medidas pelos dois operários",
        col=c("#F0FFFF","#FFDAB9"))

                  
}
\keyword{ASM}
\keyword{CEQ}

