% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpprechtTb6.10.R
\name{EpprechtTb6.10}
\alias{EpprechtTb6.10}
\title{Concentração de um processo químico}
\format{Um \code{data.frame} com 100 observações e 2 variáveis, 
    em que

\describe{

\item{\code{amostra}}{Identificação da amostra.}

\item{\code{conc}}{Concentração registrada na amostra.}
 
}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estatístico de Qualidade (2nd ed.). São Paulo - SP:
     Editora Atlas. (pg 182).
}
\description{
Dados referentes à concentração de um processo químico
    registrados a cada 3 minutos.
}
\examples{

data(EpprechtTb6.10)

library(qcc)

obj <- qcc.groups(EpprechtTb6.10$conc, EpprechtTb6.10$amostra)
qcc(obj, type="xbar.one", nsigmas=3)

                  
}
\keyword{CEQ}

