% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpprechtTb6.4.R
\name{EpprechtTb6.4}
\alias{EpprechtTb6.4}
\title{Temperatura do banho químico}
\format{Um \code{data.frame} com 20 observações e 4 variáveis, 
    em que

\describe{

\item{\code{amostra}}{Identificação da amostra.}

\item{\code{ord}}{Fator que indica a ordem dos registros de uma amostra
    (1 = primeira temperatura registrada, 2 = segunda temperatura, 
    3 = terceira temperatura.}
 
\item{\code{temp}}{Temperatura registrada.}
 
}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estat?stico de Qualidade (2nd ed.). S?o Paulo - SP:
     Editora Atlas. (pg 167).
}
\description{
A cada 30 minutos registram-se três temperaturas do 
    banho, espaçadas de 3 minutos. Exemplo: na primeira amostra 
    efetua-se uma medida às 8h00, outra às 8h03 e outra às 8h06. Na 
    segunda amostra serão registradas as temperaturas nos horários 
    8h30, 8h33 e 8h36, e assim por diante.
}
\examples{

data(EpprechtTb6.4)

hist(EpprechtTb6.4$temp)
                  
}
\keyword{CEQ}

