% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epprecht.R
\name{EpprechtTb8.12}
\alias{EpprechtTb8.12}
\title{Controle de Qualidade para a Fra\enc{çã}{ca}o de N\enc{ã}{a}o
    Conformes}
\format{Um \code{data.frame} com 20 observações e 1 variável, em que

\describe{

\item{\code{nconf}}{Número de peças não conformes nas amostras.}

}}
\source{
COSTA et al. (2010), pág. 227.
}
\description{
Dados de 20 amostras, representando o número de peças
    não-conformes em amostras de tamanho 100.
}
\examples{

data(EpprechtTb8.12)
str(EpprechtTb8.12)

library(qcc)

qcc(EpprechtTb8.12, type = "p", size = 100,
    xlab = "Amostra",
    ylab = "Proporção de itens não conformes",
    title = "")

}
\keyword{CEQ}

