% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epprecht.R
\name{EpprechtTb8.16}
\alias{EpprechtTb8.16}
\title{Controle de Qualidade na Produ\enc{çã}{ca}o de Cabos
    El\enc{é}{e}tricos}
\format{Um \code{data.frame} com 10 observações e 2 variáveis, em que

\describe{

\item{\code{met}}{Quantidade de cabo avaliado em cada amostra (em
    metros).}

\item{\code{def}}{Número de defeitos encontrados.}

}}
\source{
COSTA et al. (2010), pág. 233.
}
\description{
Dados referentes ao monitoramento do processo de
    produção de cabos elétricos. Foi registrada a quantidade de
    defeitos encontrados em amostras de determinada quantidade de
    metros de cabo.
}
\examples{

data(EpprechtTb8.16)
str(EpprechtTb8.16)

library(qcc)

qcc(EpprechtTb8.16$def, type = "u", sizes = EpprechtTb8.16$met,
    xlab = "Amostra", ylab = "Quantidade de defeitos", title = " ")

}
\keyword{CEQ}

