% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpprechtTb8.8.R
\name{EpprechtTb8.8}
\alias{EpprechtTb8.8}
\title{Controle de qualidade para o número de não-conformidades em geladeiras}
\format{Um \code{data.frame} com 40 observações e 1 variável, 
    em que

\describe{

\item{\code{nconf}}{Número de não-conformidades.} 

}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estatístico de Qualidade (2nd ed.). São Paulo - SP:
     Editora Atlas. (pg 222).
}
\description{
Dados referentes aos números de não-conformidades  
    em 40 amostras de cinco geladeiras. Foi suposto que o processo
    operava sob controle nesse período.
}
\examples{

data(EpprechtTb8.8) 

library(qcc)

qcc(EpprechtTb8.8, type="c", size=5,
    xlab="Amostra", ylab="Número de não-conformidades", title=" ")
                  
}
\keyword{CEQ}

