% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Faria.R
\name{FariaEg2.9.5}
\alias{FariaEg2.9.5}
\title{Varia\enc{çã}{ca}o de M\enc{é}{e}todos para a Determina\enc{çã}{ca}o da CTC do Solo}
\format{Um \code{data.frame} com 20 observações e 3 variáveis, em que

\describe{

\item{\code{metod}}{Fator que identifica os métodos para a
    determinação aplicados nas amostras de solo.}

\item{\code{ctc}}{CTC da amostra de solo, em cmol kg\eqn{^{-1}}.}

}}
\source{
Faria, J. C. (2009). Notas de aulas expandidas (10th ed.)
    Ilhéus - BA: UESC. (Exemplo 2.9.5, pág 39)
}
\description{
Dois métodos de CTC do solo são usados em uma amostra de
    controle e fornecem os resultados em cmol kg\eqn{^{-1}}.
}
\examples{

library(lattice)

data(FariaEg2.9.5)
str(FariaEg2.9.5)

aggregate(ctc ~ metod, data = FariaEg2.9.5,
          FUN = function(x) {
              c(media = mean(x), variancia = var(x))
          })

xyplot(ctc ~ metod, data = FariaEg2.9.5)

}
\keyword{ASS}

