% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FerreiraEg3.4.R
\name{FerreiraEg3.4}
\alias{FerreiraEg3.4}
\title{Exame de Sangue em Pintores de Carro}
\format{Um \code{data.frame} com 20 observações e 2 variáveis, em que

\describe{

\item{\code{conc}}{Concentração de hemoglobina.}

\item{\code{linf}}{Contagem de linfócitos.}

}}
\source{
Ferreira, D. F. (2011). Estatística Multivariada (2nd
    ed.). Lavras, MG: Editora UFLA. (Exemplo 3.4 pág. 136)

Royston, J. P. (1983). Some techniques for assessing multivariate
    normality based on Shapiro-Wilk. London, Applied Statistics -
    Journal of the Royal Statistical Society - Series C, v. 32, n. 2,
    p. 121-133. (Table 2)
}
\description{
Subconjunto de valores de um hemograma aplicado à
    pintores de carros, apresentado por Royston (1983).
}
\examples{

# Normalidade univariada
par(mfrow = c(1, 2))
with(FerreiraEg3.4, {
    qqnorm(conc, main = "Normal Q-Q Plot\\nConc. Hemoglobina")
    qqline(conc)
    qqnorm(linf, main = "Normal Q-Q Plot\\nCont. Linfócitos")
    qqline(linf)
})
#'
}
\keyword{TODO}

