% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FerreiraEg5.1.R
\name{FerreiraEg5.1}
\alias{FerreiraEg5.1}
\title{Teores de Areia e Argila em Solo na Amazônia}
\format{Um \code{data.frame} com 30 observações e 2 variáveis, em que

\describe{

\item{\code{areia}}{Representa o teor de areia em cada parcela de
    solo.}

\item{\code{argila}}{Representa o teor de argila em cada parcela de
    solo.}

}}
\source{
Ferreira, D. F. (2011). Estatística Multivariada (2nd
    ed.). Lavras, MG: Editora UFLA. (Exemplo 5.1 pág. 193-194)
}
\description{
Conjunto de dados referente aos teores de areia e argila
    de uma amostra com 30 parcelas de solo de capoeira nova na
    Amazônia.
}
\examples{
data(FerreiraEg5.1)

pairs(FerreiraEg5.1)

layout(matrix(c(2, 0, 1, 3), ncol = 2, byrow = TRUE),
       widths = c(4/5, 1/5), heights = c(1/5, 4/5))
areiaPlot <- hist(FerreiraEg5.1$areia, plot = FALSE)
argilaPlot <- hist(FerreiraEg5.1$argila, plot = FALSE)
par(mar = c(5, 5, 0.5, 0.5))
plot(argila ~ areia, data = FerreiraEg5.1, cex = 1.5)
par(mar = c(0, 5, 1, 1))
barplot(areiaPlot$counts, axes = FALSE, space = 0)
par(mar = c(5, 0, 1, 1))
barplot(argilaPlot$counts, axes = FALSE, space = 0, horiz = TRUE)

}
\keyword{TODO}

