% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FerreiraEg8.1.R
\name{FerreiraEg8.1}
\alias{FerreiraEg8.1}
\title{Número de Grãos e Produtividade em Cultivares de Feijão}
\format{Um \code{data.frame} com 20 observações e 4 variáveis.

\describe{

\item{\code{cult}}{Fator nominal com 4 níveis que representa as
    cultivares de feijão.}

\item{\code{rept}}{Inteiro que identifica as unidades experimentais
    de cada cultivar.}

\item{\code{ngrao}}{Número médio de grãos por vagem.}

\item{\code{prod}}{Produtividade de feijão, em kg ha\eqn{^{-1}}.}

}}
\source{
Ferreira, D. F. (2011). Estatística Multivariada (2nd
    ed.). Lavras, MG: Editora UFLA. (Exemplo 8.1, pág. 339)
}
\description{
Dados obtidos de um experimento inteiramente casualizado
    com cinco repetições, realizado no período de seca, onde
    avaliou-se o efeito de diferentes cultivares na produtividade do
    feijão por meio das variáveis produtividade e número médio de
    grãos por vagem.
}
\examples{

data(FerreiraEg8.1)
str(FerreiraEg8.1)

library(lattice)
library(latticeExtra)

p1 <- xyplot(prod ~ cult,
             data = FerreiraEg8.1,
             type = c("p", "smooth"))
p2 <- xyplot(ngrao ~ cult,
             data = FerreiraEg8.1,
             type = c("p", "smooth"))
doubleYScale(p1, p2, add.ylab2 = TRUE)

xyplot(prod ~ ngrao,
       groups = cult,
       grid = TRUE,
       auto.key = list(space = "right",
                       title = "Cultivar", cex = 0.8),
       data = FerreiraEg8.1,
       xlab = "Número médio de grãos por vagem",
       ylab = "Produção de feijão")

}
\keyword{manova}

