% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FerreiraEx7.4.1.R
\name{FerreiraEx7.4.1}
\alias{FerreiraEx7.4.1}
\title{Avalia\enc{çã}{ca}o de Dieta em Animais}
\format{Um \code{data.frame} com 12 observações e 3 variáveis.

\describe{

\item{\code{fase}}{Fator que indica a qual fase do estudo a
    observação pertence. Antes ou depois da dieta.}

\item{\code{peso}}{Peso do animal.}

\item{\code{teor}}{Teor de proteína.}

}}
\source{
Ferreira, D. F. (2011). Estatística Multivariada (2nd
    ed.). Lavras, MG: Editora UFLA. (Exercício 7.4.1 pág. 328)
}
\description{
Dados referentes a um estudo com animais avaliados antes
    e depois de terem sido submetidos a uma dieta balanceada. O
    interesse do estudo é avaliar se houve efeito significativo da
    dieta nas variáveis peso e teor de proteína.
}
\examples{
data(FerreiraEx7.4.1)

aggregate(peso ~ fase, data = FerreiraEx7.4.1, summary)
aggregate(teor ~ fase, data = FerreiraEx7.4.1, summary)

by(FerreiraEx7.4.1[2:3], FerreiraEx7.4.1[1], cov)

with(FerreiraEx7.4.1, {
     par(mar = c(4, 5, 4, 5))
     plot.default(y = peso, xlab = "", ylab = "",
                  x = jitter(as.numeric(fase), factor = 0.5) - 0.15,
                  xlim = c(0.5, 2.5),
                  col = 4,
                  axes = FALSE)
     box()
     axis(side = 2, col.axis = 4)
     mtext(side = 2, text = "Peso", line = 3, col = 4) 
     par(new = TRUE, mar = c(4, 5, 4, 5))
     plot.default(y = teor, xlab = "", ylab = "",
                  x = jitter(as.numeric(fase), factor = 0.5) + 0.15,
                  xlim = c(0.5, 2.5),
                  col = 2,
                  axes = FALSE)
     axis(side = 4, col.axis = 2)
     axis(side = 1, at = 1:2, labels = c("antes", "depois"))
     mtext(side = 4, text = "Teor da proteína",
           line = 3, col = 2)
})

}
\keyword{TODO}

