% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ferreira.R
\name{FerreiraEx9.7.2}
\alias{FerreiraEx9.7.2}
\title{Avalia\enc{çã}{ca}o Nutricional de Tipos de Carne}
\format{Um \code{data.frame} com 5 observações e 5 variáveis, em que

\describe{

\item{\code{ener}}{Valor energético de uma porção, em calorias.}

\item{\code{prot}}{Valor proteico de uma porção, em gramas.}

\item{\code{gord}}{Quantidade de gordura em uma porção, em gramas.}

\item{\code{calc}}{Quantidade de cálcio em uma porção, em
    miligramas.}

\item{\code{ferro}}{Quantidade de ferro em uma porção, em
    miligramas.}

}

    O tipo de carne é indicado conforme nomenclatura das linhas do
    \code{data.frame}. Os tipos de carne marisco, siri e camarão são
    todos enlatados.}
\source{
FERREIRA (2011), Exercício 9.7.2, pág. 409.
}
\description{
Dados provenientes de um estudo onde avaliou-se as
    variáveis energia, proteína, gordura, cálcio e ferro em 5
    diferentes tipos de carnes. O estudo é apresentado em Bussab et
    al., 1990 sob o objetivo de agrupar os tipos de carne com base em
    sua informação nutricional.
}
\examples{

data(FerreiraEx9.7.2)
FerreiraEx9.7.2

(cl2 <- kmeans(FerreiraEx9.7.2, 2))
(cl3 <- kmeans(FerreiraEx9.7.2, 3))
cbind("k = 2" = cl2$cluster, "k = 3" = cl3$cluster)

(D <- dist(FerreiraEx9.7.2))
hc <- hclust(D)
plot(as.dendrogram(hc), main = "Dendograma")
rect.hclust(hc, k = 2, border = 2)
rect.hclust(hc, k = 3, border = 4)
legend("topright", lty = 1, col = c(2, 4), bty = "n",
       legend = c("2 grupos (k = 2)", "3 grupos (k = 3)"))

}
\references{
Bussab, W. O., Miazaki, É. S., Andrade,
    D. F. (1990). Introdução à análise de agrupamentos. São Paulo,
    SP: ABE.
}
\keyword{AnaAgrup}
