% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labestData.R
\docType{package}
\name{labestData}
\alias{labestData}
\alias{labestData-package}
\title{Conjuntos de Dados para Ensino de Estatística}
\description{
O labestData é um projeto coletivo do PET Estatística
    UFPR para desenvolver um pacote R com conjuntos de dados para
    ensino de Estatística.  Os objetivos desse projeto são 1)
    capacitar os membros do PET Estatística a desenvolver pacote R
    com versionamento Git e 2) contribuir com a comunidade científica
    ao reunir, organizar, manter e disponibilizar conjuntos de dados
    disponíveis em livros, páginas de internet e arquivos pessoais.
}

