% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labestDataView.R
\name{labestDataView}
\alias{labestDataView}
\title{Shiny App para Visualização e Download dos Datasets}
\usage{
labestDataView()
}
\value{
Abre uma aplicação web com shiny.
}
\description{
Essa função abre uma interface \pkg{shiny} para
    visualizar as tabelas de dados e consultar a documentação. Além
    disso, tem um botão para salvar o dado em txt.
}
\section{Warning}{
 Para funcionar, é necessário ter o pacote shiny
    instalado. Para abrir a documentação dos datasets é necessário
    informar o número da porta que o navegador usa
    (\code{http://127.0.0.1:XXXXX}) para exibir a documentação em
    html. Para ter esse número, peça a documentação de alguma
    função antes de chamar a aplicação. Veja na seção exemplos como
    fazer.
}
\examples{
\dontrun{

options(help.ports = 1111)
help(BanzattoQd1.2.3, help_type = "html")
help(package = "labestData", help_type = "html")

labestDataView()

}
}
\author{
Walmes Zeviani e Eduardo Ribeiro Jr.
}

