% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyTb1.1.R
\docType{data.frame}
\name{ManlyTb1.1}
\alias{ManlyTb1.1}
\title{Pardais sobreviventes da tempestade}
\format{Um \code{data.frame} com 49 registros e 6 variáveis.

\itemize{

\item \code{ct} Comprimento total do pardal em (mm).

\item \code{ea} Extensão alar em (mm).

\item \code{cbc} Comprimento do bico e cabeça em (mm).

\item \code{cdu} Comprimento do úmero em (mm).

\item \code{cqe} Comprimento da quilhado esterno em (mm).

\item \code{sobre} Pardal Sobrevivel. S = Sim e N = Não 
}}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados- uma introdução.
     Porto Alegre, RS: ARTMED (pg 14 e 15)
}
\usage{
data(ManlyTb1.1)
}
\description{
Estudo em 1898, para a teoria da evolução de Darwin com 
  pardais moribundos, com total de 49 dados.
}
\examples{

data(ManlyTb1.1)

pairs(~ ct + ea + cbc + cdu + cqe + sobre,
      data = ManlyTb1.1, 
      main="Gráfico de dispersão dos pardais")
}
\keyword{Dados}
\keyword{Multivariados}

