% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyTb1.3.R
\name{ManlyTb1.3}
\alias{ManlyTb1.3}
\title{Distribuiçao de uma borboleta}
\format{Um \code{data.frame} com 16 colônias de boroboletas com 11 variáveis.

\describe{

\item{\code{colo}}{Identificação da colônia.}

\item{\code{alt}}{Altitude em pés.}

\item{\code{precip}}{Preciptação anual, em polegadas.}

\item{\code{tempmax}}{Temperatura máxima.}

\item{\code{tempmin}}{Temperatura mínima.}

\item{\code{dg0.4}}{Frequência 0.4 em demobilidade gênica PGi em porcentagem.}

\item{\code{dg0.6}}{Frequência 0.6 em demobilidade gênica PGi.}
 
\item{\code{dg0.8}}{Frequência 0.8 em demobilidade gênica PGi.}

\item{\code{dg1}}{Frequência 1 em demobilidade gênica PGi.}
 
\item{\code{dg1.16}}{Frequência 1.16 em demobilidade gênica PGi.}

\item{\code{dg1.3}}{Frequência 1.3 em demobilidade gênica PGi.}
 
}}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados - uma introdução.
     Porto Alegre, RS: Bookman (pg 19)
}
\description{
Um estudo de colônias de borboletas do tipo Euphydryas editha na Califórnia e Orego no EUA com variáveis ambientais e frequências gênicas.
}
\examples{

data(ManlyTb1.3)

pairs(~ alt + precip + tempmax + tempmin + dg0.4 + dg0.6 + dg0.8 + dg1 + dg1.16 + dg1.3,
      data = ManlyTb1.3, 
      main="Matriz das variáveis das colônias de borboletas")
}
\keyword{CP}

