% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb1.5}
\alias{ManlyTb1.5}
\title{Emprego em paises europeus}
\format{Um \code{data.frame} com 30 registros em 11 variáveis.

\describe{

\item{\code{pais}}{Identificação do país.}

\item{\code{grup}}{Grupo econômico ao qual pertencente o país: União
     Europeia (UE); Área europeia de livre comércio (AELC); Leste; Outro}

\item{\code{afp}}{Porcentagem da população ativa empregada na agricultura,
     florestal e pesca.}

\item{\code{mep}}{Porcentagem empregada na mineração e exploração de pedreiras.}

\item{\code{fab}}{Porcentagem empregada nas fábricas.}

\item{\code{fea}}{Porcentagem empregada nas áreas de fornecimento de
     energia e água.}

\item{\code{con}}{Porcentagem empregada na construção.}

\item{\code{ser}}{Porcentagem empregada na área de serviços.}

\item{\code{fin}}{Porcentagem empregada na área de finanças.}

\item{\code{ssp}}{Porcentagem empregada na área de serviços sociais e
    pessoais.}

\item{\code{tc}}{Porcentagem empregada nas áreas de transportes e
     comunicações.}

}}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados-
     uma introdução. Porto Alegre, RS: Bookman (pg 22)
}
\description{
Porcentagens da força de trabalho de empregados para nove
diferentes campos de trabalho em 30 países europeus
}
\examples{

data(ManlyTb1.5)

pairs(~afp + mep + fab + fea + con + ser + fin + ssp + tc,
      data = ManlyTb1.5,
      main="Matriz das variáveis de força de trabalho"
      )
}
\keyword{AnaAgrup}

