% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyTb10.2.R
\name{ManlyTb10.2}
\alias{ManlyTb10.2}
\title{Variáveis solo e vegetação em Belize}
\format{Um \code{data.frame} com 151 linhas e 8 variáveis.

\describe{

\item{\code{ser}{Porcentagem de solo com enriquecimento constante de calcário.}}

\item{\code{spc}}{Porcentagem de solo de prado com cálcio na água subterrânea.}

\item{\code{smc}}{Porcentagem de solo com matriz de coral son condições de enriquecimento constante de calcário.}

\item{\code{sao}}{Porcentagem de solos aluvial e orgânico adjacentes a rios e solo orgânico salino na costa.}

\item{\code{dfe}}{Porcentagem de floresta decídua estacional com ervas de folhas largas.}

\item{\code{flab}}{Porcentagem de floresta de locais altos e baixos coberta com água, plantas herbáceas em lugares úmidos e pântano.}

\item{\code{fpc}}{Porcentagem de floresta palma de cohune.}

\item{\code{fm}}{Porcentagem de floresta mista.}

}}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados - uma introdução.
     Porto Alegre, RS: Bookman (pg 167 à 170)
}
\description{
Estudo no distrito de Corozal em Belize em 4 variávei de solo e 4 variáveis de vegetação registradas para quadrados de 2,5 x 2,5 km.
}
\examples{

data(ManlyTb10.2)
pairs(~ ser + spc + smc + sao + fde + flab + fpc + fm,
      data = ManlyTb10.2, 
      main="Matriz das variáveis dos solos")

            
}
\keyword{ACC}

