% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb10.4}
\alias{ManlyTb10.4}
\title{Combina\enc{çã}{ca}o das Tabelas 1.5 e 6.7.}
\format{Um \code{data.frame} com 22 países e 18 variáveis.

\describe{

\item{\code{pais}}{Identificação do país.}

\item{\code{cv}}{Consumo de carne vermelha medida em gramas por
    pessoa por dia.}

\item{\code{cb}}{Consumo de carne branca.}

\item{\code{ovo}}{Consumo de ovos.}

\item{\code{leite}}{Consumo de leite.}

\item{\code{peixe}}{Consumo de peixe.}

\item{\code{cere}}{Consumo por cereais.}

\item{\code{carb}}{Consumo de carboidratos.}

\item{\code{gnl}}{Consumo de grãos, nozes e sementes oleaginosas.}

\item{\code{fv}}{Consumo de frutas e vegetais.}

\item{\code{agr}}{Porcentagem da população ativa empregada na
    agricultura, florestal e pesca.}

\item{\code{min}}{Mineração e exploração de pedreiras.}

\item{\code{fab}}{Fabricação.}

\item{\code{fea}}{Fornecimento de energia e água.}

\item{\code{con}}{Construção.}

\item{\code{ser}}{Serviços.}

\item{\code{fin}}{Finanças.}

\item{\code{ssp}}{Serviços sociais e pessoais.}

\item{\code{tc}}{Transportes e comunicações.}

}}
\source{
MANLY (2005), pág. 175.
}
\description{
Dados referentes a consumo de proteína e força de
    trabalho em países europeus.
}
\examples{

data(ManlyTb10.4)
str(ManlyTb10.4)

library(lattice)

splom(ManlyTb10.4[-1])

}
\keyword{AnaCorCan}
