% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyTb11.3.R
\name{ManlyTb11.3}
\alias{ManlyTb11.3}
\title{Dist\enc{â}{a}ncias rodovi\enc{á}{a}rias}
\format{Uma matriz (13x13) com as distâncias rodoviárias entre as 
cidades da ilha Sul da Nova Zelândia.}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados - 
     uma introdução. Porto Alegre, RS: Bookman (pg 182)
}
\description{
Distâncias rodoviárias (em milhas) entre cidades na ilha
Sul da Nova Zelândia
}
\examples{

data(ManlyTb11.3)

require(lattice)

levelplot(ManlyTb11.3, xlab = "", ylab = "", main = "Gráfico das 
distâncias rodoviárias", scales = list(x = list(rot=90)))
}
\keyword{EM}

