% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb11.5}
\alias{ManlyTb11.5}
\title{Vota\enc{çõ}{co}es de Parlamentares}
\format{Uma matriz com os números de votos discordantes entre 15
    parlamentares de Nova Jersey.}
\source{
MANLY (2005), pág. 185.
}
\description{
Número de votos discordantes entre os parlamentares de
    Nova Jersey em leis referentes a problemas ambientais
}
\examples{

data(ManlyTb11.5)
str(ManlyTb11.5)

library(lattice)

levelplot(ManlyTb11.5,
          xlab = "",
          ylab = "",
          scales = list(x = list(rot = 90)))

}
\keyword{AnaAgrup}

