% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb9.7}
\alias{ManlyTb9.7}
\title{Esp\enc{é}{e}cies de Plantas em Lotes}
\format{Um \code{data.frame} com 25 espécie de plantas com 18
    variáveis.

\describe{

\item{\code{esp}}{Espécie}

\item{\code{l1}}{Medidas de abundância das espécies no lote 1.}

\item{\code{l2}}{Medidas de abundância das espécies no lote 2.}

\item{\code{l3}}{Medidas de abundância das espécies no lote 3.}

\item{\code{l4}}{Medidas de abundância das espécies no lote 4.}

\item{\code{l5}}{Medidas de abundância das espécies no lote 5.}

\item{\code{l6}}{Medidas de abundância das espécies no lote 6.}

\item{\code{l7}}{Medidas de abundância das espécies no lote 7.}

\item{\code{l8}}{Medidas de abundância das espécies no lote 8.}

\item{\code{l9}}{Medidas de abundância das espécies no lote 9.}

\item{\code{l10}}{Medidas de abundância das espécies no lote 10.}

\item{\code{l11}}{Medidas de abundância das espécies no lote 11.}

\item{\code{l12}}{Medidas de abundância das espécies no lote 12.}

\item{\code{l13}}{Medidas de abundância das espécies no lote 13.}

\item{\code{l14}}{Medidas de abundância das espécies no lote 14.}

\item{\code{l15}}{Medidas de abundância das espécies no lote 15.}

\item{\code{l16}}{Medidas de abundância das espécies no lote 16.}

\item{\code{l17}}{Medidas de abundância das espécies no lote 17.}

}}
\source{
MANLY (2005), pág. 152.
}
\description{
Estudo de 25 espécies de plantas em 17 lotes de um prado
    de pastagem na Reserva Natural em Steneryd na Suécia. Cada valor
    na tabela é a soma dos valores cobertos em um intervalo de 0 a 5
    por nove quadrantes de amostra, de modo que um valor 45
    corresponde à completa cobertura pela espécie.
}
\examples{

data(ManlyTb9.7)
str(ManlyTb9.7)

euclid <- as.matrix(dist(ManlyTb9.7[, -1]))

heatmap(euclid,
        margins = c(6, 6),
        labRow = ManlyTb9.7$esp,
        labCol = ManlyTb9.7$esp)

}
\keyword{AnaAgrup}
