% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb9.8}
\alias{ManlyTb9.8}
\title{Bens de T\enc{ú}{u}mulos}
\format{Um \code{data.frame} com 45 linhas e 38 colunas.

\describe{

\item{\code{sep}}{Identifica o sepultamento}

\item{\code{tipo}}{Identifica os restos Mortais (1 = Adulto
    masculino; 2 = Adulto Feminino; 3 = Criança.}

\item{\code{obj1}}{Presença (1) ou ausência (0) do objeto 1.}

\item{\code{obj2}}{Presença (1) ou ausência (0) do objeto 2.}

\item{\code{obj3}}{Presença (1) ou ausência (0) do objeto 3.}

\item{\code{obj4}}{Presença (1) ou ausência (0) do objeto 4.}

\item{\code{obj5}}{Presença (1) ou ausência (0) do objeto 5.}

\item{\code{obj6}}{Presença (1) ou ausência (0) do objeto 6.}

\item{\code{obj7}}{Presença (1) ou ausência (0) do objeto 7.}

\item{\code{obj8}}{Presença (1) ou ausência (0) do objeto 8.}

\item{\code{obj9}}{Presença (1) ou ausência (0) do objeto 9.}

\item{\code{obj10}}{Presença (1) ou ausência (0) do objeto 10.}

\item{\code{obj11}}{Presença (1) ou ausência (0) do objeto 11.}

\item{\code{obj12}}{Presença (1) ou ausência (0) do objeto 12.}

\item{\code{obj13}}{Presença (1) ou ausência (0) do objeto 13.}

\item{\code{obj14}}{Presença (1) ou ausência (0) do objeto 14.}

\item{\code{obj15}}{Presença (1) ou ausência (0) do objeto 15.}

\item{\code{obj16}}{Presença (1) ou ausência (0) do objeto 16.}

\item{\code{obj17}}{Presença (1) ou ausência (0) do objeto 17.}

\item{\code{obj18}}{Presença (1) ou ausência (0) do objeto 18.}

\item{\code{obj19}}{Presença (1) ou ausência (0) do objeto 19.}

\item{\code{obj20}}{Presença (1) ou ausência (0) do objeto 20.}

\item{\code{obj21}}{Presença (1) ou ausência (0) do objeto 21.}

\item{\code{obj22}}{Presença (1) ou ausência (0) do objeto 22.}

\item{\code{obj23}}{Presença (1) ou ausência (0) do objeto 23.}

\item{\code{obj24}}{Presença (1) ou ausência (0) do objeto 24.}

\item{\code{obj25}}{Presença (1) ou ausência (0) do objeto 25.}

\item{\code{obj26}}{Presença (1) ou ausência (0) do objeto 26.}

\item{\code{obj27}}{Presença (1) ou ausência (0) do objeto 27.}

\item{\code{obj28}}{Presença (1) ou ausência (0) do objeto 28.}

\item{\code{obj29}}{Presença (1) ou ausência (0) do objeto 29.}

\item{\code{obj30}}{Presença (1) ou ausência (0) do objeto 30.}

\item{\code{obj31}}{Presença (1) ou ausência (0) do objeto 31.}

\item{\code{obj32}}{Presença (1) ou ausência (0) do objeto 32.}

\item{\code{obj33}}{Presença (1) ou ausência (0) do objeto 33.}

\item{\code{obj34}}{Presença (1) ou ausência (0) do objeto 34.}

\item{\code{obj35}}{Presença (1) ou ausência (0) do objeto 35.}

\item{\code{obj36}}{Presença (1) ou ausência (0) do objeto 36.}

\item{\code{obj37}}{Presença (1) ou ausência (0) do objeto 37.}

\item{\code{obj38}}{Presença (1) ou ausência (0) do objeto 38.}

}}
\source{
MANLY (2005), (pág. 153 e 154)
}
\description{
Conjunto de dados referentes a bens de túmulos de um
    cemitério em Bannadi, nordeste da Tailândia. São registros de
    presença ou ausência de 38 diferentes artigos em cada um dos 46
    túmulos, com informação adicional sobre se os restos mortais eram
    de um adulto masculino, feminino ou criança. Os sepultamentos
    estão na ordem de riqueza de diferentes bens (totais variando de
    0 a 11), e os bens estão na ordem de frequência de ocorrência
    (totais variando de 1 a 18).
}
\examples{

data(ManlyTb9.8)
str(ManlyTb9.8)

matdist <- as.matrix(dist(ManlyTb9.8[, -c(1, 2)], method = "binary"))

# Mapa de calor para a matriz de dissimilaridades.
heatmap(matdist,
        margins = c(6, 6),
        labRow = ManlyTb9.8$sep,
        labCol = ManlyTb9.8$sep)

}
\keyword{AnaAgrup}
