% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiAnA3}
\alias{MingotiAnA3}
\title{Dados dos Domic\enc{í}{i}lios}
\format{Um \code{data.frame} com 120 observações e 6 variáveis, em
    que

\describe{

\item{\code{dom}}{Identificação do domicílio.}

\item{\code{loc}}{Fator que indica localidade da residência (de 1 a
    3).}

\item{\code{inst}}{Fator que indica grau de instrução do chefe da
    família (de 1 a 3).}

\item{\code{nres}}{Número de pessoas residentes no domicílio.}

\item{\code{rendm}}{Renda familiar mensal (em quantidade de salários
    mínimos).}

\item{\code{rendpc}}{Renda familiar mensal per capita (em número de
    salários mínimos).}

}}
\source{
MINGOTI (2005), pág. 287.
}
\description{
Dados de uma pesquisa feita em 120 residências de uma
    determinada região. Os dados contém informações dos domicílios e
    dos residentes.
}
\examples{

data(MingotiAnA3)
str(MingotiAnA3)

library(lattice)

bwplot(inst ~ rendpc | loc,
       data = MingotiAnA3,
       pch = "|",
       xlab = "Renda per capita ",
       ylab = "Grau de instrução")

}
\keyword{AnaComPrin}
