% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiAnA4}
\alias{MingotiAnA4}
\title{Dados Relativos \enc{à}{a}s Empresas}
\format{Um \code{data.frame} com 46 observações e 6 variáveis, em que

\describe{

\item{\code{emp}}{Identificação da empresa.}

\item{\code{grp}}{Fator que indica o grupo em que a empresa está
    situada (1 = se a empresa faliu, 2 = se a empresa não faliu).}

\item{\code{x1}}{Fluxo de caixa/ total de débitos.}

\item{\code{x2}}{Rendimento da empresa/ total de patrimônio.}

\item{\code{x3}}{Patrimônio atual/ total de débito.}

\item{\code{x4}}{Patrimônio atual/ rendimento das vendas.}

}}
\source{
MINGOTI (2005), pág. 290.
}
\description{
Dados de 21 empresas, coletados aproximadamente 2 anos
    antes da falência das mesmas, e de outras 25 empresas que não
    faliram no período.
}
\examples{

data(MingotiAnA4)
str(MingotiAnA4)

library(car)

scatterplotMatrix(~x1 + x2 + x3 + x4 | grp,
                  data = MingotiAnA4,
                  reg.line = FALSE,
                  spread = FALSE,
                  smoother = FALSE)

}
\keyword{AnaComPrin}
