% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiAnA6}
\alias{MingotiAnA6}
\title{Dados de funcion\enc{á}{a}rios da empresa}
\format{Um \code{data.frame} com 109 observações e 10 variáveis, 
    em que

\describe{

\item{\code{func}}{Funcionários.}

\item{\code{grupo}}{Grupo de desempenho nos teste (0 = médio, 1 = melhor).}

\item{\code{idade}}{Idade do funcionário (em anos completos).}

\item{\code{sexo}}{Fator que indica sexo (0 = feminino, 1 = masculino).}

\item{\code{ecivil}}{Fator que indica estado Civil 
    (1 = solteiro, 0 = não solteiro).}

\item{\code{nfilho}}{Número de filhos.}

\item{\code{anexp}}{Anos de experiência na função.}

\item{\code{testp}}{Nota do teste psicotécnico (de 0 a 50).}

\item{\code{testc}}{Nota do teste de conhecimento prático da função 
    que exerce (de 1 a 10).}

\item{\code{satisf}}{Satisfação com a vida pessoal ("1" para 
    satisfeito, "0" para não satisfeito).}

}}
\source{
Mingoti, S.A. (2005). Análise de dados através de métodos de 
     estatística multivariada - uma abordagem aplicada. 
     Belo Horizonte, MG: Editora UFMG. (pg 292).
}
\description{
Dados descritivos dos funcionários de uma empresa para
    avaliar qualidade.
}
\examples{

data(MingotiAnA6) 

library(car)

scatterplotMatrix(~ testp + testc + anexp | grupo, data=MingotiAnA6, 
     reg.line=FALSE,
     spread=FALSE, 
     smoother=FALSE,
     main="Dispersão das variáveis de medidas para os dois grupos")
                  
}
\keyword{AnaComp}

