% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiTb2.2}
\alias{MingotiTb2.2}
\title{Notas de Estudantes em Tr\enc{ê}{e}s Provas}
\format{Um \code{data.frame} com 19 observações e 4 variáveis, em que

\describe{

\item{\code{est}}{Identificação do estudante.}

\item{\code{p1}}{Nota obtida na primeira prova variando de 0 a 25.}

\item{\code{p2}}{Nota obtida na segunda prova variando de 0 a 25.}

\item{\code{p3}}{Nota obtida na terceira prova variando de 0 a 25.}

}}
\source{
MINGOTI (2005), pág. 52.
}
\description{
Notas obtidas em uma turma de estudantes em três provas
    de uma determinada disciplina.
}
\examples{

data(MingotiTb2.2)
str(MingotiTb2.2)

pairs(~ p1 + p2 + p3,
      data = MingotiTb2.2)

}
\keyword{TODO}
