% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiTb3.1}
\alias{MingotiTb3.1}
\title{Receitas e Patrim\enc{ô}{o}nio de Empresas}
\format{Um \code{data.frame} com 12 observações e 4 variáveis, em que

\describe{

\item{\code{emp}}{Identificação da empresa.}

\item{\code{gbt}}{Ganho bruto da empresa (em unidades monetárias).}

\item{\code{glq}}{Ganho líquido da empresa (em unidades monetárias).}

\item{\code{patr}}{Patrimônio acumulado (unidades monetárias).}

}}
\source{
MINGOTI (2005), pág. 42.
}
\description{
Dados de 12 empresas sobre ganhos (bruto e líquido) e
    patrimônio acumulado num determinado período.
}
\examples{

data(MingotiTb3.1)
str(MingotiTb3.1)

library(car)

scatterplotMatrix(~gbt + glq + patr,
                  data = MingotiTb3.1)


}
\keyword{TODO}
