% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MingotiTb3.5.R
\name{MingotiTb3.5}
\alias{MingotiTb3.5}
\title{Impressões sensoriais de marcas de coxinha de galinha}
\format{Um \code{data.frame} com 8 observações e 5 variáveis, em que

\describe{

\item{\code{marca}}{Identificação da marca.}

\item{\code{sabor}}{Média das notas dos juízes para o sabor da
    coxinha de galinha.}

\item{\code{aroma}}{Média das notas dos juízes para o aroma da
    coxinha de galinha.}

\item{\code{massa}}{Média das notas dos juízes para a qualidade da
    massa.}

\item{\code{recheio}}{Média das notas dos juízes para a qualidade do
    recheio.}

}}
\source{
Mingoti, S. A. (2005). Análise de dados através de métodos de
    estatística multivariada: uma abordagem aplicada.  Belo
    Horizonte, MG: Editora UFMG. (pg 42).
}
\description{
Dados de 8 marcas de coxinha de galinha, avaliados em 4
    atributos (sabor, aroma, qualidade da massa e qualidade do
    recheio) avaliados por 5 julgadores numa escala de 1 a 5, quanto
    maior a nota melhor a classificação do atributo.  Os dados de
    entrada da tabela referem-se à média das notas dos 5 julgadores,
    para cada marca e cada atributo.
}
\examples{
data(MingotiTb3.5)

library(car)

scatterplotMatrix(~sabor + aroma + massa + recheio,
                  data = MingotiTb3.5,
                  main = "Matriz de dispersão")
}
\keyword{sensorial}

