% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiTb3.7}
\alias{MingotiTb3.7}
\title{Dados de Amostras de Solo}
\format{Um \code{data.frame} com 25 observações e 5 variáveis, em que

\describe{

\item{\code{amst}}{Identificação da amostra.}

\item{\code{areia}}{Porcentagem de areia.}

\item{\code{sed}}{Porcentagem de sedimentos.}

\item{\code{arg}}{Porcentagem de argila.}

\item{\code{morg}}{Quantidade de material orgânico.}

}}
\source{
MINGOTI (2005), pág. 73.
}
\description{
Dados de 25 amostras de determinado tipo de solo.  Para
    cada amostra foram registradas as porcentagens de areia,
    sedimentos, argila e a quantidade de material orgânico.
}
\examples{

data(MingotiTb3.7)
str(MingotiTb3.7)

library(car)

scatterplotMatrix(~ amst + areia + sed + arg + morg,
                  data = MingotiTb3.7)

}
\keyword{AnaComPrin}
