% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MingotiTb6.8.R
\name{MingotiTb6.8}
\alias{MingotiTb6.8}
\title{Índices de desenvolvimento de países}
\format{data.frame com 21 observações e 5 variáveis, em que

\describe{

\item{pais}{Nomes dos 21 países.}

\item{expv}{Índice da esperança de vida que mede a realização 
     relativa de um país na esperança de vida à nascença.}

\item{ed}{Índice da educação que mede a realização relativa de um 
     país tanto na alfabetização de adultos como na escolarização 
     bruta combinada do primário, secundário e superior.}

\item{pib}{Índice do PIB, que é calculado utilizando o PIB
     per capita ajustado (dólares) no qual usa-se o logarítimo.}

\item{estpol}{Índice de percepções da probabilidade de 
         desestabilização (tensões étnicas, conflito armado, etc).}

}}
\source{
MINGOTI, S.A. (2005). Análise de dados através de métodos de 
     estatística multivariada - uma abordagem aplicada. 
     Belo Horizonte, MG: Editora UFMG. (pg 184).
}
\description{
Dados relativos a 21 países com as variáveis que são 
     índices de expectativa de vida, educação, renda (PIB) e 
     estabilidade política e de segurança. Os índices foram 
     construídos por uma metodologia da ONU, e quanto maior 
     esses valores, melhor será a qualidade do país.
}
\examples{

data(MingotiTb6.8)

pairs(MingotiTb6.8)

}
\keyword{Desenvolvimento}
\keyword{Qualidade}
\keyword{de}
\keyword{vida}

