% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiTb8.1}
\alias{MingotiTb8.1}
\title{Renda e o N\enc{ú}{u}mero de Filhos}
\format{Um \code{data.frame} com 3 observações e 5 variáveis, em que

\describe{

\item{\code{renda}}{Renda, medida em 3 categorias (menos de 2000,
     2000 a 5000 e 5000 ou mais), em unidades monetárias.}

\item{\code{i0f}}{Indivíduos que não tem filhos (igual a zero
    filhos).}

\item{\code{i1f}}{Indivíduos que tem 1 filho (igual a um filho).}

\item{\code{i2f}}{Indivíduos que tem 2 filhos (igual a dois filhos).}

\item{\code{m2f}}{Indivíduos que tem mais de 2 filhos (mais que dois
    filhos).}

}}
\source{
MINGOTI (2005), pág. 258.
}
\description{
Dados relativos a 257 indivíduos classificados de acordo
    com a renda e número de filhos.
}
\examples{

data(MingotiTb8.1)
str(MingotiTb8.1)

library(lattice)

xyplot(i0f + i1f + i2f + m2f ~ renda,
       data = MingotiTb8.1,
       ylab = "Número de indivíduos",
       auto.key = TRUE)

}
\keyword{TODO}

