% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEg1.12.6}
\alias{PaulaEg1.12.6}
\title{Sal\enc{á}{a}rio de Executivos}
\format{Um \code{data.frame} com 220 observações e 4 variáveis.

\describe{

\item{\code{sal}}{Salário anual (em mil USD).}

\item{\code{sexo}}{Sexo (masculino, feminino).}

\item{\code{pos}}{Posição na empresa (escore de 1 a 9).}

\item{\code{aexp}}{Experiência (em anos).}

}}
\source{
PAULA (2004), Eg 1.12.6, pág. 97.
}
\description{
Dados referentes ao salário anual de uma amostra
    aleatória de 220 executivos (145 homens e 75 mulheres). O salário
    será relacionado com as variáveis: sexo, anos de experiência no
    cargo e posição na empresa.
}
\examples{

data(PaulaEg1.12.6)
str(PaulaEg1.12.6)

library(lattice)

xyplot(sal ~ aexp | sexo,
       data = PaulaEg1.12.6,
       type = c("p", "smooth"),
       xlab = "Anos de experiência",
       ylab = "Salário")

}
\keyword{RL}
\keyword{dummy}

