% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaEg2.4.2.R
\name{PaulaEg2.4.2}
\alias{PaulaEg2.4.2}
\title{Captura de Peixes via Espinhel de Fundo no Litoral Paulista}
\format{Um \code{data.frame} com 156 observações e 8 variáveis.
    \describe{

\item{\code{frota}}{Fator com dois níveis que indica de qual frota é
    a embarcação, \code{Santos} ou \code{Ubatuba}.}

\item{\code{ano}}{Fator com cinco níveis que representa o ano, de
    \code{1995} a \code{1996}.}

\item{\code{trim}}{Fator com quatro níveis que representa o trimestre
    em estudo, de \code{1} a \code{4}.}

\item{\code{lat}}{Latitude, definida como distância ao Equador medida
    ao longo do meridiano de Greenwich.}

\item{\code{long}}{Longitude, definida como distância ao meridiano de
    Greenwich medida ao longo do Equador.}

\item{\code{dias}}{Dias de pesca.}

\item{\code{captura}}{Quantidade de peixes-batata capturados, em kg.}

\item{\code{cpue}}{Captura por unidade de esforço, calculada como
    razão da quantidade de peixes-batata capturados (\code{captura})
    pelo número de dias de pesca \code{dias}.}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP. (Exemplo 2.4.2 pág. 127)
}
\description{
Dados parciais de um estudo sobre a atividade de frotas
    pesqueiras de espinhel de fundo baseadas no litoral paulista,
    Santos e Ubatuba. Neste estudo uma amostra de 156 embarcações
    pesqueiras, cujo peixe-batata é considerado para pesca, foi
    analisada no período de 1995 a 1999. Para cada embarcação foram
    obtidas variáveis sobre a frota (Santos ou Ubatuba), ano,
    timestre, latitude, longitude, dias de pesca, quantidade de
    peixes capturados e a captura por unidade de esforço (definida
    como divisão da quantidade de peixe captuurado pelos dias de
    pesca).
}
\examples{
data(PaulaEg2.4.2)

str(PaulaEg2.4.2)

# Separando as covariáveis numéricas
index <- sapply(PaulaEg2.4.2, is.numeric)

# Número de observações em cada combinação das covariáveis
# não numéricas
ftable(table(PaulaEg2.4.2[, !index]))

library(lattice)

# Relação marginal da variável resposta com as covariáveis
# não numéricas
bwplot(cpue ~ frota, data = PaulaEg2.4.2)
bwplot(cpue ~ ano, data = PaulaEg2.4.2)
bwplot(cpue ~ trim, data = PaulaEg2.4.2)

# Relação da variável respota com as combinações das
# covariáveis não numéricas
ue <- with(PaulaEg2.4.2, paste(ano, trim, sep = "-"))
bwplot(cpue ~ ue | frota, data = PaulaEg2.4.2,
       scales = list(x = list(rot = 90)))

# Verificando a suposição de coeficiente de variação constante,
# desconsiderando as covariáveis tri e numéricas
resumo <- aggregate(cpue ~ frota + ano, data = PaulaEg2.4.2,
                    FUN = function(x) {
                        c("Média" = mean(x),
                          "D.Padrão" = sd(x),
                          "C.Variação" = sd(x)/mean(x),
                          "n" = length(x))
                    })
ftable(xtabs(cpue ~ frota + ano, data = resumo))

# Relação das covariáveis numéricas
splom(~PaulaEg2.4.2[, index], groups = frota,
      data = PaulaEg2.4.2,
      auto.key = list(column = 2))

}
\references{
Paula, G. A., Oshiro, C. H. (2001). Relatório de Análise
    Estatística sobre o Projeto: Análise de Captura por Unidade de
    Esforço de Peixe-Batata na Frota Paulista. RAE-CEA0102, IME-USP.
}
\keyword{positivo-assimétrico}

