% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaEg2.8.1.R
\name{PaulaEg2.8.1}
\alias{PaulaEg2.8.1}
\title{Comparação de Tipos de Snack}
\format{Um \code{data.frame} com 750 observações e 3 variáveis.
    \describe{

\item{\code{semana}}{Inteiro com dez valores únicos que indentifica a
    semana da avaliação.}

\item{\code{tipo}}{Fator com cinco níveis que identifica o tipo de
    snack avaliado. Os níveis representam as seguintes configurações:
    \code{A}: 22\% de gordura, 0\% de óleo de canola, \code{B}: 0\% de
    gordura, 22\% de óleo de canola, \code{C}: 17\% de gordura, 5\% de
    óleo de canola, \code{D}: 11\% de gordura, 11\% de óleo de canola e
    \code{E}: 5\% de gordura, 17\% de óleo de canola.}

\item{\code{fnpc}}{Força necessária para o cisalhamento, valor para
    avaliação da textura dos produtos.}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP. (Exemplo 2.8.1 pág. 150)
}
\description{
Dados de um experimento desenvolvido pelo Departamento
    de Nutrição da Faculdade de Saúde Pública da USP em que 5 tipos
    diferentes de um novo snack, com baixo teor de gordura e de
    ácidos graxos, foram comparados ao longo de 20 semanas. Neste
    novo produto a gordura vegetal hidrogenada, responsável pela
    fixação do aroma do produto, foi substituída, totalmente ou
    parcialmente por óleo de canola. Ao todo foram 750 observações
    providas de 15 avaliações para cada tipo de snack a cada 2
    semanas.
}
\examples{
data(PaulaEg2.8.1)

str(PaulaEg2.8.1)

# Experimento balanceado, 15 observações para cada tipo em cada
# semana
ftable(PaulaEg2.8.1[, c("tipo", "semana")])
xtabs(fnpc ~ tipo + semana, data = PaulaEg2.8.1)

library(lattice)
bwplot(fnpc ~ tipo | factor(semana),
       data =  PaulaEg2.8.1,
       as.table = TRUE,
       strip = strip.custom(strip.names = TRUE,
                            var.name = "semana"))

# Estatísticas descritivas
resumo <- aggregate(fnpc ~ tipo + semana,
                    data = PaulaEg2.8.1,
                    FUN = function(x) {
                        c("Média" = mean(x),
                          "D.Padrão" = sd(x),
                          "C.Variação" = sd(x)/mean(x),
                          "n" = length(x))
                    })
ftable(xtabs(fnpc ~ tipo + semana, data = resumo))

xyplot(fnpc[, "Média"] ~ semana,
       groups = tipo,
       data = resumo,
       type = c("l", "g"),
       auto.key = list(
           points = FALSE,
           lines = TRUE,
           corner = c(0.1, 0.9)))

}
\keyword{positivo-assimétrico}

