% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEg3.6.9c}
\alias{PaulaEg3.6.9c}
\title{Prefer\enc{ê}{e}ncia de Consumidores}
\format{Um \code{data.frame} com 263 observações e 4 variáveis, em
    que

\describe{

\item{pref}{Preferência do comprador por um tipo de automóvel (1 =
   americano, 0 = japonês).}

\item{idade}{Idade do comprador (em anos).}

\item{sexo}{Sexo do comprador (0 =  masculino; 1 = feminino).}

\item{est}{Estado civil do comprador (0 = casado, 1 = solteiro).}

}}
\source{
PAULA (2004), Eg 3.6.9c, pág. 231.
}
\description{
Dados sobre a preferência de consumidores americanos com
    relação a automóveis. Uma amostra aleatória de 263 consumidores
    foi considerada. As seguintes variáveis foram observadas para
    cada comprador: preferência quanto ao tipo de automóvel, idade,
    sexo e estado civil.
}
\examples{

library(lattice)

data(PaulaEg3.6.9c)
str(PaulaEg3.6.9c)

bwplot(idade ~ pref,  data = PaulaEg3.6.9c,
       type = "p", pch = "|",
       xlab = "Preferência - Japonês e Americano",
       ylab = "Idade do Comprador")

}
\keyword{MLG}
\keyword{binarios}

