% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEg5.2.8a}
\alias{PaulaEg5.2.8a}
\title{N\enc{ú}{u}mero de \enc{Á}{A}caros em Placas de Esterco de
    Gado}
\format{Um \code{data.frame} com 102 observações e 8 variáveis.

\describe{

\item{\code{esp2}}{Número de ácaros coletados da espécie 2.}

\item{\code{esp3}}{Número de ácaros coletados da espécie 3.}

\item{\code{esp6}}{Número de ácaros coletados da espécie 6.}

\item{\code{esp14}}{Número de ácaros coletados da espécie 14.}

\item{\code{placa}}{Número de partes da placa de esterco onde foram
    coletados os ácaros. (1 ou 6)}

\item{\code{posic}}{Posição na placa de esterco onde foram coletados
    os ácaros (central ou lateral).}

\item{\code{reg}}{Região onde a placa de esterco foi coletada (São
     Roque, Pindamonhangaba, Nova Odessa ou Ribeirão Preto).}

\item{\code{temp}}{Temperatura no local da coleta, medida
    \eqn{C^{\circ}}.}

}}
\source{
PAULA (2004), (Exemplo 5.2.8a, pág. 359)
}
\description{
Dados de um experimento desenvolvido para estudar a
    distribuição do número de ácaros em placas de esterco de gado
    bovino no estado de São Paulo, obtidos por Paula e Tavares, 1992.
    Essas placas são depósitos de ovos da mosca do chifre
    (\emph{Haematobia irritans}), uma das pragas mais importantes da
    pecuária brasileira. Os ácaros são inimigos naturais da mosca do
    chifre, uma vez que se alimentam de ovos e larvas dessas moscas.
}
\examples{

data(PaulaEg5.2.8a)
str(PaulaEg5.2.8a)

library(lattice)

index <- sapply(PaulaEg5.2.8a, is.numeric)
splom(PaulaEg5.2.8a[, index],
      type = c("p", "g"),
      lwd = 2, col.line = 1)

}
\references{
Paula, G. A. e Tavares, H. R. (1992). Relatório de
    Análise Estatística sobre o Projeto: Ácaros Associados ao Esterco
    Bovino. Subsídios para Controle Biológico da Mosca do Chifre.
    RAECEA 9206, IME-USP
}
\keyword{contagem}
