% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx1.13.19}
\alias{PaulaEx1.13.19}
\title{Estudo Demogr\enc{á}{a}fico dos Estados Norte-Americanos}
\format{Um \code{data.frame} com 50 observações e 9 variáveis.

\describe{

\item{\code{estado}}{Nome do estado.}

\item{\code{pop}}{População estimada em julho de 1975.}

\item{\code{renda}}{Renda per capita em 1974 (em dólares).}

\item{\code{analf}}{Proporção de analfabetos em 1970.}

\item{\code{crime}}{Taxa de criminalidade por cem mil habitantes em
    1976.}

\item{\code{estud}}{Porcentagem de estudantes que concluem o segundo
    grau em 1970.}

\item{\code{ndias}}{Número de dias do ano com temperatura abaixo de
    0\eqn{^\circ C} na cidade mais importante do estado.}

\item{\code{area}}{Área do estado (em milhas quadradas).}

\item{\code{expvi}}{Expectativa de vida nos anos de 1969-1970.}

}}
\source{
PAULA (2004), Exercício 1.13.19, pág. 109.
}
\description{
Dados referentes a um estudo demográfico sobre os 50
    estados norte-americanos. Neste estudo foram registradas 8
    variáveis que contém informações sobre características da
    população e do estado. Dentre elas temos a variável expectativa
    de vida, havendo interesse em explicá-la utilizando as demais
    informações.
}
\examples{

data(PaulaEx1.13.19)
str(PaulaEx1.13.19)

library(car)

PaulaEx1.13.19$dens <- PaulaEx1.13.19$pop/PaulaEx1.13.19$area
scatterplotMatrix(~expvi + analf + crime + estud + ndias + dens,
                  data = PaulaEx1.13.19)


}
\keyword{RM}
