% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx2.10.16}
\alias{PaulaEx2.10.16}
\title{Rela\enc{çã}{ca}o entre Publicidade e Faturamento em
    Restaurantes}
\format{Um \code{data.frame} com 30 observações e 2 variáveis.

\describe{

\item{\code{fatura}}{Faturamento anual do restaurante, em mil
    dólares.}

\item{\code{gastos}}{Gastos do restaurante com publicidade, em mil
    dólares.}

}}
\source{
PAULA (2004), Exercício 2.10.16, pág. 179.
}
\description{
Dados referentes a faturamentos anuais e gastos com
    publicidade de uma amostra de 30 restaurantes. O objetivo
    principal é relacionar o faturamento médio com os gastos com
    publicidade.
}
\examples{

data(PaulaEx2.10.16)
str(PaulaEx2.10.16)

library(lattice)

xyplot(fatura ~ gastos,
       data = PaulaEx2.10.16,
       type = c("p", "g", "smooth"))

}
\references{
Montgomery, D. C., Peck, E. A., Vining,
    G. G. (2001). Introduction to Linear Regression Analysis (3rd
    Ed.). John Wiley, New York.
}
\keyword{RS}
