% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx2.10.19}
\alias{PaulaEx2.10.19}
\title{Estudo Sobre Leucemia e Caracter\enc{í}{i}stica
    Morfol\enc{ó}{o}gica nas C\enc{é}{e}lulas Brancas}
\format{Um \code{data.frame} com 32 observações e 3 variáveis.

\describe{

\item{\code{ncel}}{Número de células brancas na amostra do paciente.}

\item{\code{carac}}{Fator com dois níveis que representa a presença
    (\code{AG positivo}) ou ausência (\code{AG negativo}) da
    característica morfológica.}

\item{\code{tempo}}{Tempo de sobrevivência do paciente, em semanas.}

}}
\source{
PAULA (2004), Exemplo 2.10.19, pág. 180.
}
\description{
Dados provenientes de um estudo em que pacientes com
    leucemia foram classificados segundo a ausência ou presença de
    uma característica morfológica nas células brancas. O objetivo do
    estudo foi avaliar essa característica morfológica a partir do
    tempo de sobrevivência dos pacientes.
}
\examples{

data(PaulaEx2.10.19)
str(PaulaEx2.10.19)

library(latticeExtra)

xyplot(tempo ~ ncel,
       groups = carac,
       data = PaulaEx2.10.19,
       type = c("p", "g", "spline"),
       scales = list(x = list(log = 10)),
       xscale.components = xscale.components.logpower,
       auto.key = list(
           corner = c(0.95, 0.95),
           lines = TRUE,
           cex.title = 1,
           title = "Característica morfológica"))

}
\references{
Feigl, P., Zelen, M. (1965). Estimation of exponential
    survival probabilities with concomitant information. Biometrics
    21, 826-838.
}
\keyword{RS}
\keyword{dummy}
\keyword{heterovar}
