% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx3.7.20}
\alias{PaulaEx3.7.20}
\title{Pacientes com Leucemia}
\format{Um \code{data.frame} com 51 pacientes e 9 variáveis.

\describe{

\item{\code{idade}}{Idade do paciente (em anos).}

\item{\code{mdd}}{Mancha diferencial da doença (em porcentagem).}

\item{\code{if}}{Infiltração na medula (em porcentagem).}

\item{\code{cl}}{Células com leucemia (em porcentagem).}

\item{\code{md}}{Malignidade da doença (*10^3).}

\item{\code{tmax}}{Temperatura máxima antes do tratamento (*10 F°).}

\item{\code{trat}}{Tratamento (1 = satisfatório, 0 = não
    satisfatório).}

\item{\code{tsobre}}{Tempo de sobrevivência  (em meses).}

\item{\code{sit}}{Situação (1 = sobrevivente, 0 = não sobrevivente).}

}}
\source{
PAULA (2004), Exercício 3.7.19, página 276.
}
\description{
Estudo com 51 pacientes adultos, previamente
    diagnosticados com um tipo agudo de leucemia, que receberam um
    tipo de tratamento sendo verificada, após certo período, a
    eficiência ou não do tratamento.
}
\examples{

data(PaulaEx3.7.20)
str(PaulaEx3.7.20)

# Transformar variáveis.
PaulaEx3.7.20 <- transform(PaulaEx3.7.20,
                           trat = as.factor(PaulaEx3.7.20$trat),
                           sit = as.factor(PaulaEx3.7.20$sit))

library(car)

scatterplotMatrix(~idade + mdd + im + cl + md + tmax + trat +
                      tsobre + sit,
                  spread = FALSE,
                  pch = 20,
                  lwd = 2,
                  smooth = TRUE,
                  data = PaulaEx3.7.20,
                  cex = 1.5)

}
\references{
Everitt, B. S. (1994).  A Handbook of Statistical
    Analysis using S-Plus.  Chapman and Hall, London. (Página 253)
}
\keyword{binário}
\keyword{sobrevivência}

