% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaEx3.7.25.R
\name{PaulaEx3.7.25}
\alias{PaulaEx3.7.25}
\title{Pulsação}
\format{Um \code{data.frame} com 92 observações e 3 variáveis.
\describe{

\item{\code{pulsa}}{ Fator de 2 níveis que representam a pulsação
    em repouso (normal, alta). }

\item{\code{fuma}}{ Fator de 2 níveis referente ao hábito de
    fumar (sim, não). }

\item{\code{peso}}{ Peso medido em quilogramas (Kg). }

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
computacional. São Paulo, SP: IME-USP. (Exercício 3.7.25, pág. 280)
}
\description{
Os dados provém de uma amostra de 92 homens adultos.
    Eles foram questionados sobre o hábito de fumar e foi medido
    os respectivos peso e pulsação (classificada como normal e alta).
}
\examples{

library(lattice)

data(PaulaEx3.7.25)

xyplot(peso ~ fuma | pulsa, data = PaulaEx3.7.25,
    main = "Pulsação")
    
model <- glm(pulsa ~ peso + fuma, data = PaulaEx3.7.25, 
    family = "binomial")
summary(model)

}
\keyword{TODO}

