% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaEx3.7.7c.R
\name{PaulaEx3.7.7c}
\alias{PaulaEx3.7.7c}
\title{Influ\enc{ê}{e}ncia de Extrato Vegetal e Qu\enc{í}{i}mico}
\format{Um \code{data.frame} com 7 observações e 3 variáveis.
\describe{

\item{\code{dose}}{Dose.}

\item{\code{cexp}}{Caramujos expostos.}

\item{\code{cmort}}{Caramujos mortos.}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP. (Ex 3.7.7c, pág. 269)
}
\description{
Experimento de dose-resposta conduzido para avaliar a 
    influência dos extratos vegetais "aquoso frio de folhas", "aquoso
    frio de frutos" e de um extrato químico, respectivamente, na morte 
    de um determinado tipo de caramujo.
}
\examples{

data(PaulaEx3.7.7c)

str(PaulaEx3.7.7c)

library(lattice)

xyplot(PaulaEx3.7.7c$cmort/PaulaEx3.7.7c$cexp ~ dose,
     data = PaulaEx3.7.7c,
     xlab = "Dose",
     type = c("o"),
     ylab = "Proporção de mortos",
     auto.key = list(space="top", columns=2, 
                  title="Caramujos", cex.title=1,
                  lines=TRUE, points=FALSE))



}
\keyword{MLG}

