% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaTb2.6}
\alias{PaulaTb2.6}
\title{Proje\enc{çã}{ca}o de Vendas}
\format{Um \code{data.frame} com 20 observações e 2 variáveis.

\describe{

\item{\code{proj}}{Valor projetado de venda.}

\item{\code{real}}{Valor real de venda.}

}}
\source{
PAULA (2004), Tabela 2.6, pág. 159.
}
\description{
Dados referentes a 20 valores projetados para vendas de
    produtos confrontados com os verdadeiros valores de venda
    obtidos.
}
\examples{

data(PaulaTb2.6)
str(PaulaTb2.6)

library(lattice)

xyplot(real ~ proj, data = PaulaTb2.6,
       grid = TRUE, pch = 19, cex = 1.2,
       panel = function(x, y, ...) {
           panel.xyplot(x, y, ...)
           panel.abline(0, 1, lty = 2)
       })

}
\keyword{RS}

