% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaTb3.21.R
\name{PaulaTb3.21}
\alias{PaulaTb3.21}
\title{Distribuição de Rotifers das Duas Espécies}
\format{Um \code{data.frame} com 40 observações e 4 variáveis.
\describe{

\item{\code{dens}}{Densidade.}

\item{\code{susp}}{Rotifers suspensos.}

\item{\code{exp}}{Rotifers expostos.}

\item{\code{esp}}{Espécie (Polyarthra, Keratella).}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP. (Tb 3.21, pág. 257)
}
\description{
Experimentos com duas espécies de *rotifers*, um tipo  
    microscópio de invertebrado aquático. São apresentados pra cada 
    espécie a densidade relativa da substância, o número de *rotifers*
    expostos e o número de *rotifers* em suspensão.
}
\examples{

data(PaulaTb3.21)

str(PaulaTb3.21)

library(lattice)

xyplot(PaulaTb3.21$susp/PaulaTb3.21$exp ~ dens,
       groups = esp,
       data = PaulaTb3.21,
       xlab = "Densidade",
       ylab = "Proporção de rotifers suspensos",
       type = c("o"),
       auto.key = list(space="top", columns=2, 
                       title="Espécie", cex.title=1,
                       lines=TRUE, points=FALSE))

}
\keyword{MLG}

