% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaTb4.14.R
\name{PaulaTb4.14}
\alias{PaulaTb4.14}
\title{Ocorrência de Doença das Coronárias}
\format{Um \code{data.frame} com 32 observações e 4 variáveis.
    \describe{

\item{\code{doenca}}{Fator com dois níveis que indica a ocorrência
    (\code{sim}) ou não ocorrência (\code{não}) de doença das
    coronárias.}

\item{\code{colest}}{Fator com quatro níveis que indica o nível de
    colesterol do paciente. A unidade de medida adotada é
    \eqn{mg/100cm^3} com classes \code{<200}, \code{200-219},
    \code{220-259} e \code{>259}.}

\item{\code{pa}}{Fator com quatro níveis que indica a pressão
    arterial do paciente. A unidade de medida adotada é mm Hg
    (milímetro-mercúrio) com classes \code{<127}, \code{127-146},
    \code{147-166} e \code{>166}.}

\item{\code{nind}}{Número de indivíduos na combinação das variáveis
    \code{doenca}, \code{colest} e \code{pa}.}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP. (Tabela 4.14, pág. 334)
}
\description{
Os dados são referentes à classificação de 1330
    pacientes segundo três fatores: ocorrência da doença das
    coronárias, nível de colesterol e pressão arterial. O interesse
    está em verificar a associação destes fatores com base no número
    de indivíduos classificados em cada combinação.
}
\examples{

data(PaulaTb4.14)

str(PaulaTb4.14)

xt <- xtabs(nind ~ ., data = PaulaTb4.14)
ftable(xt)
plot(xt)

library(lattice)
xyplot(nind ~ colest | doenca,
       groups = pa,
       data = PaulaTb4.14,
       type = c("p", "a", "g"),
       scales = list(x = list(rot = 45)),
       auto.key = list(
           space = "right", cex.title = 1,
           title = "Pressão arterial\\nem mm Hg"))

}
\references{
Everitt, B. S. (1977). The Analysis of Contingency
    Tables. Chapman anda Hall, London.
}
\keyword{contagem}

