% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulinoEx9.31.R
\name{PaulinoEx9.31}
\alias{PaulinoEx9.31}
\title{Efeito de Dois Agentes Sexualmente Transmissíveis em Dois
    Grupos de Mulheres}
\format{Um \code{data.frame} com 24 observações e 5 variáveis, em que

\describe{

\item{\code{parc}}{Fator sócio-comportamental definido pelo número
    total de parceiros sexuais, categorizado em 3 níveis (1-3, 4-10 e
    > 10).}

\item{\code{herpes}}{Vírus do tipo 2 do herpes simples (1 para maior
    prevalência e 2 para menor).}

\item{\code{calmid}}{Vírus do tipo \emph{Clamídia trachomatis} (1
    para maior prevalência e 2 para menor).}

\item{\code{grupo}}{Classificação das mulheres (casos ou
    controles).}

\item{\code{freq}}{Frequências observadas.}

}}
\source{
PAULINO; SINGER (2006), Exercício 9.31, pág. 345.
}
\description{
Estudo envolvendo um grupo de mulheres com displasia
    historicamente confirmada (casos) e outro grupo de mulheres
    com ausência da mesma (controles). Cada uma das mulheres
    foi então analisada em termos da exposição a cada um dos agentes
    sexualmente transmissíveis (o vírus tipo 2 da herpes simples e a
    \emph{Clamídia trachomatis}, na etiologia da displasia cervical),
    categorizados em dois níveis (1 para uma maior prevalência de
    anticorpos e 2 para uma menor prevalência), analisando os
    possíveis efeitos desses agentes nessas mulheres.
}
\examples{

data(PaulinoEx9.31)

str(PaulinoEx9.31)

xt <- xtabs(freq ~., PaulinoEx9.31)
ftable(xt)

library(vcd)
doubledecker(xt)

}
\keyword{retrospetivo}
