% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelEg4.2}
\alias{PimentelEg4.2}
\title{Alimenta\enc{çã}{ca}o de Porcos}
\format{Um \code{data.frame} com 20 observações e 2 variáveis, em que

\describe{

\item{\code{racoes}}{Fator de 4 níveis qualitativos, usado para
    identificar o tipo de ração.}

\item{\code{ganhopeso}}{Aumento do peso dos porcos, observado em
    quilogramas.}

}}
\source{
PIMENTEL-GOMES (2009).
}
\description{
Experimento (fictício) de alimentação de porcos em que
    se usaram quatro rações (A, B, C, D), cada uma fornecida a cinco
    animais escolhidos ao acaso. A tabela apresenta os aumentos de
    peso observados.
}
\examples{

plot(PimentelEg4.2,
     xlab = "Rações",
     ylab = "Ganho de Peso")

}
\keyword{DIC}

