% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelEg7.4}
\alias{PimentelEg7.4}
\title{Experimento de Aduba\enc{çã}{ca}o com Torta de Filtro e Adubo
    Mineral}
\format{Um \code{data.frame} com 16 observações e 4 variáveis, em que

\describe{

\item{\code{bloco}}{Fator de 4 níveis qualitativos usado para
    controle local.}

\item{\code{mineral}}{Fator de 2 níveis métricos codificados que
    representa a aplicação de adubo mineral, sendo que a variável
    assume 1, quando houver adubo mineral na parcela e -1 caso
    contrário.}

\item{\code{torta}}{Fator de 2 níveis métricos codificados que
    representa a utilização de torta dos filtros de Oliver de usinas
    de açúcar, sendo que a variável assume 1, quando houver torta na
    parcela e -1 caso contrário.}

\item{\code{y}}{Respoata medida no ensaio.}

}}
\source{
PIMENTEL-GOMES (2009), Exemplo 7.4, pág. 120.
}
\description{
Experimento fatorial 2\eqn{^{2}} em que os fatores eram
    adubo mineral e torta dos filtros de Oliver de usinas de açúcar.
}
\examples{

library(lattice)

data(PimentelEg7.4)
str(PimentelEg7.4)

xtabs(~mineral + torta, data = PimentelEg7.4)

xyplot(y ~ mineral, groups = torta, data = PimentelEg7.4,
       type = c("p", "a"),
       xlab = "Adubo mineral",
       ylab = "Resposta",
       auto.key = list(columns = 2,
                       title = "Torta", cex.title = 1.1))

}
\keyword{FAT2K}

