% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelPg142.R
\name{PimentelPg142}
\alias{PimentelPg142}
\title{Grupo de Experimentos de Adubação de Algodão}
\format{Um \code{data.frame} com 100 observações e 6 variáveis, em
    que

\describe{

\item{\code{exper}}{Fator de 5 níveis qualitativos, que refere-se aos
    diferentes experimentos contidos no grupo de experimentos.}

\item{\code{rept}}{Fator de 4 níveis qualitativos, que refere-se às
    diferentes repetições de cada experimento.}

\item{\code{N}}{Fator de 3 níveis qualitativos, que refere-se às
    diferentes dosagens de Nitrogênio na composição do adubo, sendo
    que a variável assume 0 quando não houver dosagem alguma de
    Nitrogênio, 1 quando houver uma dosagem e 2 quando houverem duas
    dosagens.}
    
\item{\code{P}}{Fator de 2 níveis qualitativos, que refere-se às
    diferentes dosagens de Fósforo na composição do adubo, sendo
    que a variável assume 0 quando não houver dosagem alguma de
    Fósforo e 1 quando houver uma dosagem.}
    
\item{\code{K}}{Fator de 3 níveis qualitativos, que refere-se às
    diferentes dosagens de Potássio na composição do adubo, sendo
    que a variável assume 0 quando não houver dosagem alguma de
    Potássio, 1 quando houver uma dosagem e 2 quando houverem duas
    dosagens.}
    
\item{\code{Y}}{Variável que mede a produção de algodão em função
    dos tratamentos utilizados.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística
    Experimental (15th ed.). Piracicaba, SP: FEALQ. (Página 142)
}
\description{
Grupo de experimentos que são comparáveis, onde todos
    são experimentos de adubação de algodão com 4 repetições cada.
}
\examples{

library(lattice)

grafico1 <- xyplot(Y ~ exper,
                   groups = rept,
                   jitter.x = TRUE,
                   data = PimentelPg142,
                   xlab = "Experimento",
                   ylab = "Produção de Algodão")

grafico2 <- bwplot(Y ~ N,
                   data = PimentelPg142,
                   xlab = "Dosagem de Nitrogênio",
                   ylab = "Produção de Algodão")

grafico3 <- bwplot(Y ~ P,
                   data = PimentelPg142,
                   xlab = "Presença de Fósforo",
                   ylab = "Produção de Algodão")

grafico4 <- bwplot(Y ~ K,
                   data = PimentelPg142,
                   xlab = "Dosagem de Potássio",
                   ylab = "Produção de Algodão")

print(grafico1, position = c(0, 0.5, 0.5, 1), more = TRUE)

print(grafico2, position = c(0.5, 0.5, 1, 1), more = TRUE)

print(grafico3, position = c(0, 0, 0.5, 0.5), more = TRUE)

print(grafico4, position = c(0.5, 0, 1, 0.5))

}
\keyword{GE}

