% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelPg91}
\alias{PimentelPg91}
\title{Preparo com do solo na Produ\enc{çã}{ca}o de Milho}
\format{Um \code{data.frame} com 24 observações e 3 variáveis, em que

\describe{

\item{\code{aradura}}{Fator de 2 níveis qualitativos, que são os
    diferentes tipos de aradura.}

\item{\code{bloco}}{Fator de 6 níveis qualitativos, usado para
    controle local.}

\item{\code{prod}}{Produção de milho, em kg por parcela de
    200\eqn{m^{2}}.}

}}
\source{
PIMENTEL-GOMES (2009).
}
\description{
Experimento em blocos casualizados realizado pelo
    Engenheiro Agrônomo Duvilio Ometto, com 2 tratamentos (aradura
    profunda e aradura superficial) e 6 blocos com 2 repetições cada.
}
\examples{

library(lattice)

data(PimentelPg91)
str(PimentelPg91)

xtabs(~aradura + bloco, data = PimentelPg91)

xyplot(prod ~ aradura,
       jitter.x = TRUE,
       groups = bloco,
       type = "o",
       data = PimentelPg91,
       xlab = "Aradura",
       ylab = "Produção")

}
\keyword{DBC}
\keyword{replicata}

