% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb11.3.1.R
\name{PimentelTb11.3.1}
\alias{PimentelTb11.3.1}
\title{Experimento de Produ\enc{çã}{ca}o de Milho em L\enc{á}{a}tice Quadrado}
\format{Um \code{data.frame} com 48 observações e 4 variáveis, em que

\describe{

\item{\code{rept}}{Fator de 3 níveis qualitativos, que são as 3
    repetições do quadrado triplo. Cada repetição tem 4 blocos de
    tamanho 4 e 1 repetição de cada tratamento que não ocorrem juntas
    nos blocos de uma mesma repetição.}

\item{\code{bloc}}{Fator de 4 níveis qualitativos, usado para
    controle local. Cada bloco tem tamanho 4.}

\item{\code{hibr}}{Fator de 16 níveis qualitativos, que são os
    híbridos de milho.}

\item{prod}{Produção de milho, em kg/parcela.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística
    Experimental (15th ed.). Piracicaba, SP: FEALQ. (Tabela 11.3.1,
    pág. 215)
}
\description{
Produção de milho em um experimento em delineamento
    reticulado quadrado triplo de 4 \eqn{times} 4 avaliando a 16
    híbridos de milho, em que \eqn{m = 3} repetições ortogonais e
    \eqn{k = 4} blocos de tamanho 4.
}
\examples{

data(PimentelTb11.3.1)
str(PimentelTb11.3.1)

xtabs(~hibr + rept, data = PimentelTb11.3.1)
ftable(xtabs(~rept + bloc + hibr, data = PimentelTb11.3.1))

library(lattice)

xyplot(prod ~ hibr,
       groups = rept,
       type = c("p", "a"),
       auto.key = list(title = "Repetições",
                       cex.title = 1.1, columns = 3),
       data = PimentelTb11.3.1,
       xlab = "Híbridos de milho",
       ylab = "Produção (kg/parcela)")

}
\keyword{LAT}

