% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelTb12.2.1}
\alias{PimentelTb12.2.1}
\title{Experimento de Aduba\enc{çã}{ca}o com P\eqn{_2}O\eqn{_5} em
    Milho}
\format{Um \code{data.frame} com 20 observações e 3 variáveis, em que

\describe{

\item{\code{bloco}}{Fator de 4 níveis qualitativos, usado para
    controle local.}

\item{\code{P2O5}}{Fator de 5 níveis métricos que são as doses de
    \eqn{P_2O_5} em \eqn{kg ha^{-1}}.}

\item{\code{prod}}{Produção de milho, em kg/parcela.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística
    Experimental (15th ed.). Piracicaba, SP: FEALQ. (Tabela 12.2.1,
    pág. 232)
}
\description{
Experimento de adubação de milho feito pelos engenheiros
    agrônomos Glauco Pinto Viegas e Erik Smith, em blocos ao acaso,
    para estudar o efeito da adubação na produtividade da cultura.
}
\examples{

library(lattice)

data(PimentelTb12.2.1)
str(PimentelTb12.2.1)

xyplot(prod ~ P2O5,
       groups = bloco,
       data = PimentelTb12.2.1,
       auto.key = list(title = "Blocos", cex.title = 1,
                       columns = 4),
       ylab = "Produção (kg/parcela)",
       xlab = expression(Dosagem~de~P[2]*O[5]~(kg~ha^{-1})),
       panel = panel.superpose,
       degree = 2,
       panel.groups = function(x, y, col, col.symbol, degree, ...) {
           panel.xyplot(x, y, col = col.symbol, ...)
           xnew <- seq(min(x), max(x), length.out = 30)
           m0 <- lm(y ~ poly(x, degree = degree))
           ynew <- predict(m0, newdata = list(x = xnew))
           panel.lines(x = xnew, y = ynew, col = col.symbol)
       })


}
\keyword{RP}

