% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb14.4.1.R
\name{PimentelTb14.4.1}
\alias{PimentelTb14.4.1}
\title{Ensaio de Alimentação de Vacas}
\format{Um \code{data.frame} com 36 observações e 5 variáveis, em que

\describe{

\item{\code{grupo}}{Fator de 4 níveis qualitativos, usado para
    controle local.}

\item{\code{vaca}}{Fator de 3 níveis qualitativos, que são as raças
    das vacas, utilizado como linhas do quadrado latino de 3x3.}

\item{\code{periodo}}{Fator de 3 níveis qualitativos, que são os
    períodos sucessivos de produção de leite, utilizado como colunas
    do quadrado latino de 3x3.}

\item{\code{alimento}}{Fator de 3 níveis qualitativos, que são os
    tipos de alimentação utilizados.}

\item{\code{prod}}{Produções de leite, em kg}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 14.4.1, pág. 272)
}
\description{
Experimento de nutrição de vacas, períodos e grupos de
    vacas referente a produção de leite em função de níveis de
    alimentação. Em cada grupo tem-se um quadrado latino de 3x3.
}
\examples{

library(lattice)

data(PimentelTb14.4.1)
str(PimentelTb14.4.1)

xyplot(prod ~ alimento,
       groups = grupo,
       type = c("a", "p"),
       auto.key = list(title = "Grupos de Vaca", cex.title = 1, columns = 4),
       data = PimentelTb14.4.1,
       ylab = "Produção de Leite (em kg)",
       xlab = "Tipo de Alimentação Utilizado")

}
\keyword{SQL}

