% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb14.5.1.R
\name{PimentelTb14.5.1}
\alias{PimentelTb14.5.1}
\title{Ensaio de Reversão na Produção de Leite}
\format{Um \code{data.frame} com 36 observações e 5 variáveis, em que

\describe{

\item{\code{periodo}}{Fator de 3 níveis qualitativos, que são os
    períodos sucessivos de produção de leite.}

\item{\code{bloc}}{Fator de 3 níveis qualitativos, usado para
    controle local.}

\item{\code{vaca}}{Fator de 12 níveis qualitativos, que são as raças
    das vacas.}

\item{\code{trat}}{Fator de 3 níveis qualitativos, que são os
    tratamentos utilizados no experimento.}

\item{\code{prod}}{Produções de leite, em kg.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 14.5.1, pág. 278)
}
\description{
Experimento de reversão (switch-back) avaliando a
    produção de leite em função do nível de nutrição. Os fatores
    período, bloco e vaca são níveis de blocagem.
}
\examples{

library(lattice)

data(PimentelTb14.5.1)
str(PimentelTb14.5.1)

xyplot(prod ~ trat,
       groups = bloc,
       type = c("a", "p"),
       auto.key = list(title = "Bloco", cex.title = 1, columns = 3),
       data = PimentelTb14.5.1,
       ylab = "Produção de Leite (em kg)",
       xlab = "Tratamento Utilizado")

}
\keyword{ER}

