% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelTb16.2.1}
\alias{PimentelTb16.2.1}
\title{Ensaio de Controle de Pragas do Feijoeiro}
\format{Um \code{data.frame} com 40 observações e 4 variáveis, em que

\describe{

\item{\code{bloc}}{Fator de 8 níveis qualitativos, usado para
    controle local.}

\item{\code{inset}}{Fator de 5 níveis qualitativos, que são os
    diferentes tipos de inseticidas, sendo um a testemunha e os
    demais produtos comerciais.}

\item{\code{stand}}{Variável quantitativa, que refere-se ao número de
    plantas de cada parcela.}

\item{\code{prod}}{Produção de feijão, em g/parcela.}

}}
\source{
PIMENTEL-GOMES (2009), Tabela 16.2.1, pág. 300.
}
\description{
Ensaio de controle de pragas do feijoeiro, realizado
    pelo engenheiro agrônomo João Ferreira do Amaral. Nesse
    experimento, em 8 blocos casualizados, foram utilizados 5
    tratamentos. Além da produção do feijoeiro (g/parcela)
    determinou-se o número de plantas de cada parcela para que fosse
    corrigida a variação da produção final pela quantidade de plantas
    na parcela.
}
\examples{

library(lattice)
library(latticeExtra)

data(PimentelTb16.2.1)
str(PimentelTb16.2.1)

cex <- with(PimentelTb16.2.1, {
    std <- stand - min(stand)
    cex <- std/max(std)
    cex <- 0.5 + 1 * cex
    nums <- function(x) seq(min(x), max(x), length.out = 5)
    key <- list(text = list(formatC(nums(stand), digits = 1)),
                points = list(cex = nums(cex), pch = 1),
                title = "Plantas por parcela",
                cex.title = 1.1,
                columns = 5)
    return(list(cex = cex, key = key))
})

xyplot(prod ~ inset, data = PimentelTb16.2.1,
       cex = cex$cex, key = cex$key,
       ylab = "Produção de feijão (g/parcela)",
       xlab = "Inseticidas")

xyplot(prod ~ stand, data = PimentelTb16.2.1,
       groups = inset, type = c("p", "r"),
       auto.key = list(title = "Inseticidas", cex.title = 1.1,
                       columns = 2),
       ylab = "Produção de feijão (g/parcela)",
       xlab = "Plantas por parcela")

}
\keyword{COV}
\keyword{DBC}

