% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb16.3.1.R
\name{PimentelTb16.3.1}
\alias{PimentelTb16.3.1}
\title{Suscetibilidade de Variedades de Soja ao Ataque de Nematóides}
\format{Um \code{data.frame} com 60 observações e 4 variáveis, em que

\describe{

\item{\code{bloc}}{Fator de 4 níveis qualitativos, usado para
    controle local.}

\item{\code{varied}}{Fator de 15 níveis qualitativos, que são as
    variedades de soja.}

\item{\code{X}}{Média da notas da infestação por conta das plantas da
    variedade Abura, que é bastante suscetível aos nematóides.}

\item{\code{Y}}{Média das notas da infestação por conta das
    variedades de soja.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 16.3.1, pág. 306)
}
\description{
Experimento de Arruda (1952), em blocos casualizados, em
    que se compararam 21 variedades de soja para estudar sua
    suscetibilidade ao ataque de nematóides. Em cada parcela foram
    colocadas plantas da variedade Abura, bastante suscetível a
    nematóides.
}
\examples{

library(lattice)

data(PimentelTb16.3.1)
str(PimentelTb16.3.1)

xyplot(X ~ varied,
       groups = bloc,
       type = c("a", "p"),
       auto.key = list(title = "Bloco",
                       cex.title = 1,
                       columns = 4),
       data = PimentelTb16.3.1,
       ylab = "Valores da Infestação em Abura",
       xlab = "Variedades de Soja")

xyplot(Y ~ varied,
       groups = bloc,
       type = c("a", "p"),
       auto.key = list(title = "Bloco",
                       cex.title = 1,
                       columns = 4),
       data = PimentelTb16.3.1,
       ylab = "Valores da infestação nas Variedades de Soja",
       xlab = "Variedades de Soja")

}
\keyword{DBC}

