% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb17.3.1.R
\name{PimentelTb17.3.1}
\alias{PimentelTb17.3.1}
\title{Produção de Lenha de Espécies de Eucaliptos}
\format{Um \code{data.frame} com 100 observações e 4 variáveis, em que

\describe{

\item{\code{exper}}{Fator de 2 níveis qualitativos, usado para
    diferenciar o grupo de experimentos.}

\item{\code{bloc}}{Fator de 5 níveis qualitativos, usado para
    controle local.}

\item{\code{especie}}{Fator de 19 níveis qualitativos, que são as
    espécies de eucalipto.}

\item{\code{prod}}{Produção de lenha, em metros cúbicos por parcela,
    por corte aos 8 anos de idade.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 17.3.1, pág. 317)
}
\description{
Grupo de dois experimentos com espécies de eucaliptos
    realizados pela Companhia Paulista de Estradas de Ferro
    (Pimentel-Gomes e Guimarães, 1958). Ambos os ensaios, localizados
    lado a lado no campo, tinham 10 espécies, em 5 blocos
    casualizados. Uma única espécie (Eucalyptus saligna) ocorria em
    ambos os experimentos.
}
\examples{

library(lattice)

data(PimentelTb17.3.1)
str(PimentelTb17.3.1)

xyplot(prod ~ especie,
       groups = bloc,
       type = c("a", "p"),
       auto.key = list(title = "Bloco",
                       cex.title = 1,
                       columns = 5),
       data = PimentelTb17.3.1,
       ylab = "Produção de Lenha (em metros cúbicos por parcela)",
       xlab = "Espécies de Eucalipto")

}
\keyword{DBC}

