% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb17.4.1.R
\name{PimentelTb17.4.1}
\alias{PimentelTb17.4.1}
\title{Grupo de Ensaios de Variedades de Cana-de-açúcar}
\format{Um \code{data.frame} com 53 observações e 4 variáveis, em que

\describe{

\item{\code{exper}}{Fator de 2 níveis qualitativos, usado para
    diferenciar o grupo de experimentos.}

\item{\code{bloc}}{Fator de 4 níveis qualitativos, usado para
    controle local.}

\item{\code{varied}}{Fator de 11 níveis qualitativos, que são as
    variedades de cana-de-açúcar.}

\item{\code{y}}{Variável resposta medida nas parcelas.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 17.4.1, pág. 322)
}
\description{
Grupo de dois ensaios com variedades de cana-de-açúcar,
    realizado na Argentina por G. Kenning Voss. Os dados referem-se a
    um ensaio com 7 variedades de cana, com 3 repetições, e outro com
    8 variedades e 4 repetições.
}
\examples{

library(lattice)

data(PimentelTb17.4.1)
str(PimentelTb17.4.1)

xyplot(y ~ varied,
       groups = bloc,
       type = c("a", "p"),
       auto.key = list(title = "Bloco",
                       cex.title = 1,
                       columns = 4),
       data = PimentelTb17.4.1,
       ylab = "Variável Resposta",
       xlab = "Variedade de Cana-de-açúcar")

}
\keyword{GE}

