% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb18.2.1.R
\name{PimentelTb18.2.1}
\alias{PimentelTb18.2.1}
\title{Ensaio de Adubação de Milho}
\format{Um \code{data.frame} com 42 observações e 3 variáveis, em que

\describe{

\item{\code{bloc}}{Fator de 3 níveis qualitativos, usado para
    controle local.}

\item{\code{trat}}{Fator de 30 níveis qualitativos, que são os
    tratamentos aplicados em cada parcela, sendo que cada algarismo
    possui um significado diferente conforme sua posição: A posição
    1 indica os níveis de Nitrogênio; A posição 2 indica os níveis de
    Fósforo; A posição 3 indica  os níveis de Potássio; A posição 4
    (após o hífen) indica uma possível adição de Calcário; E a
    posição 5 (após o hífen) indica uma possível adição de
    micronutrientes.}

\item{\code{prod}}{Produção de milho, em \eqn{kg.ha^{-1}}.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 18.2.1, pág. 330)
}
\description{
Ensaio de adubação de milho, fatorial de \eqn{3^{3}} com
    N, P e K, com confundimento de 2 graus de liberdade da interação
    tripla. A cada três blocos de 9 parcelas juntaram-se 5
    tratamentos adicionais.
}
\examples{

library(lattice)

data(PimentelTb18.2.1)
str(PimentelTb18.2.1)

xyplot(prod ~ trat,
       groups = bloc,
       auto.key = list(title = "Blocos",
                       cex.title = 1,
                       columns = 3),
       data = PimentelTb18.2.1,
       type = c("a", "p"),
       ylab = "Produção de Milho (em kg/ha)",
       xlab = "Tratamento")

}
\keyword{FAT3}

