% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb5.3.1.R
\name{PimentelTb5.3.1}
\alias{PimentelTb5.3.1}
\title{Teor de colesterol no sangue}
\format{data.frame com 14 observações e 3 variáveis, em que

\describe{

\item{periodo}{Fator de 2 níveis qualitativos, usado para comparar o
    teor de colesterol no sangue antes e depois do uso do óleo de 
    milho.}

\item{paciente}{Variável utilizada para identificar os 7 pacientes.}

\item{colesterol}{Teor de colesterol no sangue em mg por 100g.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística
    Experimental (15th ed.). Piracicaba: FEALQ.
}
\description{
Pesquisa sobre o efeito do óleo de milho no teor de
    colesterol do sangue, realizada em sete pacientes tomados
    como blocos, cujos dados foram obtidos pelo médico Dr. Ben 
    Hur C. Paiva.
}
\examples{

library(lattice)

xyplot(colesterol ~ periodo,
       groups = paciente,
       data = PimentelTb5.3.1,
       type = "p",
       xlab = "Período",
       ylab = "Teor de Colesterol")

}
\keyword{DBC}
\keyword{colesterol}

