% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelTb7.6.1}
\alias{PimentelTb7.6.1}
\title{Experimento Fatorial Com Confunfimento de Aduba\enc{çã}{ca}o
    em Cana-de-a\enc{çú}{cu}car}
\format{Um \code{data.frame} com 54 observações e 6 variáveis, em que

\describe{

\item{\code{bloco}}{Fator de 3 níveis qualitativos, usado para
    controle local e acomodar o confundimento de dois graus de
    liberdade.}

\item{\code{rept}}{Inteiro com 2 níveis que indica as duas repetições
    realizadas de cada tratamento dentro de cada bloco.}

\item{\code{N}}{Fator de 3 níveis codificados que representa a dose
    de nitrogênio usada na adubação.}

\item{\code{P}}{Fator de 3 níveis codificados que representa a dose
    de fośforo usada na adubação.}

\item{\code{K}}{Fator de 3 níveis codificados que representa a dose
    de potássio usada na adubação.}

\item{\code{prod}}{Produção de cana-de-açúcar, em t.ha\eqn{^{-1}}.}

}}
\source{
PIMENTEL-GOMES (2009), Tabela 7.6.1, pág. 126.

Straus, F. Esperimentos de adubação na zona canavieira de
    Pernambuco. In: Terceira Reunião Brasileira de Ciência do
    Solo. Anais... t.1 p.336-443, 1951.
}
\description{
Experimento fatorial, de 3\eqn{^{3}}, de adubação de
    cana com NPK, feito por Strauss (1951), com confundimento (grupo
    W) de dois graus de liberdade da interação tripla N \eqn{\times}
    P \eqn{\times} K. Foi usado o confundimento correspondente aos
    blocos, e foram feitas duas repetições para os 27 tratamentos.
}
\details{
Foi constatado que o livro tem um erro de tipografia na
    tabela com os dados pois no bloco W1 existem duas ocorrências do
    tratamento 202 sendo que a última deveria ser 220. Foi feita a
    inclusão desses dados no pacote com essa correção.
}
\examples{

library(lattice)

data(PimentelTb7.6.1)
str(PimentelTb7.6.1)

xtabs(~N + P + K, data = PimentelTb7.6.1)

xyplot(prod ~ N | factor(P), groups = K, data = PimentelTb7.6.1,
       type = c("p", "a"), as.table = TRUE,
       xlab = "Nível codificado de nitrogênio",
       ylab = "Produção de cana-de-açúcar (t/ha)",
       auto.key = list(title = "Potássio", cex.title = 1.1,
                       columns = 3),
       strip = strip.custom(strip.names = TRUE,
                            var.name = "Fósforo"))

}
\keyword{FAT3K}
\keyword{confundimento}
