% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelTb7.8.1}
\alias{PimentelTb7.8.1}
\title{Acidez de Variedades de Mangas em Fun\enc{çã}{ca}o da \enc{É}{E}poca}
\format{Um \code{data.frame} com 54 observações e 4 variáveis, em que

\describe{

\item{\code{mes}}{Fator de 3 níveis qualitativos ordenados que
    representam 3 meses do ano: Novembro (N), Dezembro (D) e Janeiro
    (J).}

\item{\code{ano}}{Fator de 3 níveis métricos, que representam 3
    anos consecutivos: 1957, 1958 e 1959.}

\item{\code{varied}}{Fator de 6 níveis qualitativos que são as 6
    variedades de manga.}

\item{\code{acidez}}{Acidez das mangas. A escala de medida não foi
    informada.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística
    Experimental (15th ed.). Piracicaba, SP: FEALQ. (Tabela 7.8.1,
    pág 132)

Simão, S. Estudo da planta e dos frutos da mangueira (\emph{Magnifera
    indica} L.). Piracicaba, 1960. Tese.
}
\description{
Experimento fatorial, de 6 \eqn{\times} 3 \eqn{\times}
    3, referente a acidez de 6 variedades de mangas, em 3 meses do
    ano e em 3 anos agrícolas sucessivos reproduzidos de um trabalho
    de Simão (1960).
}
\examples{

library(lattice)

data(PimentelTb7.8.1)
str(PimentelTb7.8.1)

xyplot(acidez ~ mes | factor(ano), groups = varied,
       data = PimentelTb7.8.1,
       as.table = TRUE, type = c("p", "a"),
       auto.key = list(title = "Variedades", cex.title = 1.1,
                       columns = 3),
       xlab = "Meses do ano",
       ylab = "Acidez do fruto")

}
\keyword{FAT3}
\keyword{PS}

