% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb8.3.1.R
\name{PimentelTb8.3.1}
\alias{PimentelTb8.3.1}
\title{Grupo de Ensaios de Batatinha}
\format{Um \code{data.frame} com 56 observações e 3 variáveis, em que

\describe{

\item{\code{varied}}{Fator de 8 níveis qualitativos, que são as
    diferentes variedades de batatinha.}

\item{\code{exper}}{Fator de 7 níveis qualitativos, que refere-se aos
    diferentes experimentos.}
    
\item{\code{totais}}{Totais de variedades e quadrados médios
    residuais.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística
    Experimental (15th ed.). Piracicaba, SP: FEALQ. (Tabela 8.3.1)
}
\description{
Grupo de ensaios de competição de variedades de
    batatinha, realizados pelo engenheiro agrônomo Oscar A. Garay, da
    Estação Experimental de Balcarce, Argentina, instalados em
    localidades da região batateira da Província de Buenos Aires.
    Cada um deles tinha 4 repetições e 8 variedades em blocos
    casualizados.
}
\examples{

library(lattice)

xyplot(totais ~ varied,
       groups = exper,
       data = PimentelTb8.3.1,
       xlab = "Variedades de Batatinha",
       ylab = "Totais de variedades")

}
\keyword{GE}

