% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoEg12.10.R
\name{RamalhoEg12.10}
\alias{RamalhoEg12.10}
\title{Ensaio Dialelo para o Tempo Até o Florescimento do Feijoeiro}
\format{Um \code{data.frame} com 9 observações e 3 variáveis, em que

\describe{

\item{\code{prog1}}{Inteiro que indica o progenitor 1 do cruzamento.}

\item{\code{prog2}}{Inteiro que indica o progenitor 2 do cruzamento.}

\item{\code{dias}}{Número médio de dias para o florescimento do
    feijoeiro.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (Exemplo 12.10, pág. 227)

    Arriel, E., Ramalho, M. A. P., Pacheco, C. A. P. Expected and
    Realized Gains in the CMS-39 Maize Population after three Cycles
    of Selection. Revista Brasileira de Genética. Ribeirão Preto.
    v.16, n.4, p.1013. dez. 1993
}
\description{
Dados referentes ao cruzamento dialelo realizado por
    Arriel et al. (1993), em que foi avaliado o número de dias para o
    florescimento do feijoeiro para cada progenie.
}
\examples{

str(RamalhoEg12.10)

RamalhoEg12.10

# Número de genitores.
n <- with(RamalhoEg12.10, unique(c(prog1, prog2)))
RamalhoEg12.10 <- transform(RamalhoEg12.10,
                            prog1 = factor(prog1, levels = n),
                            prog2 = factor(prog2, levels = n))

# Matriz do cruzamento dialelo.
X1 <- model.matrix(~0 + prog1, data = RamalhoEg12.10)
X2 <- model.matrix(~0 + prog2, data = RamalhoEg12.10)
X <- 0.5 * (X1 + X2)

}
\keyword{Dialelo}

