% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoEg7.8.R
\name{RamalhoEg7.8}
\alias{RamalhoEg7.8}
\title{Produção de grãos de feijão}
\format{Um \code{data.frame} com 240 observações e 4 variáveis, em 
    que

\describe{

\item{\code{fam}}{Família \eqn{F_{5}} do cruzamento Jalo x Small 
    White.}

\item{\code{rept}}{Fator que indica se a produção veio da primeira ou 
    segunda repetição.}

\item{\code{plant}}{Inteiro que representa os grãos de feijão no 
    experimento.}

\item{\code{prod}}{Produção de grãos de feijão, em g/planta.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras, MG: UFLA. (pg 110).
}
\description{
Experimento referente à produção de grãos de feijão 
    obtida na avaliação de 24 famílias \eqn{F_{5}} do cruzamento 
    Jalo x Small White, cujos dados foram obtidos por Souza (1991).
    Experimento realizado em blocos incompletos.
}
\examples{

library(lattice)

boxplot(prod ~ plant, data = RamalhoEg7.8,
    xlab = "Planta",
    ylab = "Produção de grãos")

}
\keyword{DBC}

