% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoEx12.2.R
\name{RamalhoEx12.2}
\alias{RamalhoEx12.2}
\title{Dados Genéricos para Regressão Múltipla}
\format{Um \code{data.frame} com ?? observações e 5 variáveis, em que

\describe{

\item{\code{cult}}{Fator que indica as cultivares.}

\item{\code{prod}}{Produtividade de grãos, medidos em kg/ha.}

\item{\code{alt}}{Altura da planta, em centímetros.}

\item{\code{perf}}{Número de perfilhos.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (Exercício 12.2, pág. 231)
}
\description{
Conjunto de dados apresentados para exercício. Objetivo
do exercício é estimar a equação de regressão múltipla.
}
\examples{

data(RamalhoEx12.2)

plot(RamalhoEx12.2)

}
\keyword{RM}

