% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoEx13.2.R
\name{RamalhoEx13.2}
\alias{RamalhoEx13.2}
\title{Peso de Espigas de Milho}
\format{Um \code{data.frame} com 24 observações e 4 variáveis, em que

\describe{

\item{\code{cult}}{Fator de 8 níveis que indica a cultivar de milho.}

\item{\code{bloc}}{Fator de 3 níveis que são os blocos do
    experimento, utilizado para fazer controle local.}

\item{\code{plant}}{Número de plantas por parcela.}

\item{\code{peso}}{Peso de espigas despalhadas.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (Exercício 13.2, pág. 256)
}
\description{
Experimento envolvendo o número de plantas por parcela
    e peso de espigas despalhadas obtida na avaliação de cultivares
    de milho, utilizando o delineamento de blocos casualizados com 3
    repetições.
}
\examples{

data(RamalhoEx13.2)

str(RamalhoEx13.2)

library(lattice)

xyplot(peso ~ cult | bloc, data = RamalhoEx13.2,
       xlab = "Peso de espigas",
       ylab = "Plantas por parcela")

aggregate(plant ~ cult,  data = RamalhoEx13.2,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

xyplot(peso ~ plant, groups = bloc, data = RamalhoEx13.2,
       auto.key = list(space = "right", title = "Cultivar"),
       xlab = "Peso de espigas",
       ylab = "Plantas por parcela")

}
\keyword{COV}
\keyword{DBC}

