% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RamalhoEx3.1.R
\name{RamalhoEx3.1}
\alias{RamalhoEx3.1}
\title{Número de plantas de milho após inoculação}
\format{data.frame com 20 observações e 4 variáveis, em que

\describe{

\item{fugic}{Tipo de fungicida aplicado.}

\item{rept}{Repetições de inoculação do fungicida.}

\item{plant}{Número de plantas de milho observadas.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (pg 50)
}
\description{
Experimento referente ao número de plantas de milho,
    por parcela, após a inoculação com Diploidia. Experimento para
    avaliação de fungicidas para tratamento de sementes.
}
\examples{
library(lattice)

data(RamalhoEx3.1)

xyplot(jitter(plant) ~ fungic, data = RamalhoEx3.1,
       groups = repet, auto.key = TRUE,
       xlab = "Fungicida", ylab = "Número de plantas")
}
\keyword{contagem}

