% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoEx4.2.R
\name{RamalhoEx4.2}
\alias{RamalhoEx4.2}
\title{Número de perfilhos de arroz}
\format{data.frame com 32 observações e 3 variáveis, em que

\describe{

\item{linh}{Fator de níveis nominais, indicando a linhagem de arroz.}

\item{bloc}{Blocos onde é aplicado o tratamento (linhagem).}

\item{perf}{Número de perfilhos de arroz obervado.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005). 
    Experimentação em Genética e Melhoramento de Plantas (2th ed.). 
    Lavras: UFLA. (pg 66)
}
\description{
Experimento do número de perfilhos (dados médios por 
    plantsas) de 8 linhagens de arroz, avaliadas no delineamento de 
    blocos casualizados.
}
\examples{

library(lattice)

data(RamalhoEx4.2)

aggregate(perf ~ linh,  data = RamalhoEx4.2, 
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

xyplot(perf ~ linh, data = RamalhoEx4.2, 
       groups = bloc, 
       jitter.x = TRUE,
       xlab = "Linhagem",
       ylab = "Número de perfilho")

}
\keyword{DBC}

