% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramalho.R
\name{RamalhoTb13.15}
\alias{RamalhoTb13.15}
\title{Produ\enc{çã}{ca}o de Ramas de Clones de Mandioca}
\format{Um \code{data.frame} com 20 observações e 3 variáveis, em que

\describe{

\item{\code{clone}}{Fator categórico de 20 níveis que indica o clone
    de mandioca.}

\item{\code{mae}}{Produção de ramas da planta "mãe".}

\item{\code{filha}}{Produção de ramas da planta "filha".}

}}
\source{
RAMALHO et al. (2005), Tabela 13.15, pág. 252.
}
\description{
Experimento envolvendo a produção de ramas (parte aérea)
    de 20 genótipos parentais (clones) de mandioca e de seus
    respectivos decendentes.
}
\examples{

data(RamalhoTb13.15)

str(RamalhoTb13.15)

library(lattice)
library(latticeExtra)

xyplot(filha ~ mae, data = RamalhoTb13.15,
       type = c("p", "smooth"), grid = TRUE, aspect = "iso",
       xlab = "Produção de ramas na planta mãe",
       ylab = "Produção de ramas na planta filha")
    layer(panel.abline(a = 0, b = 1, lty = 2))

}
\keyword{COV}
\keyword{RS}

