% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramalho.R
\name{RamalhoTb13.6}
\alias{RamalhoTb13.6}
\title{Produ\enc{çã}{ca}o de Gr\enc{ã}{a}os de Feij\enc{ã}{a}o e Teor
    de \eqn{P_{2}O_{5}}}
\format{Um \code{data.frame} com 30 observações e 4 variáveis, em que

\describe{

\item{\code{cult}}{Fator de 10 níveis que representa a cultivar de
    feijão utilizada no experimento.}

\item{\code{bloc}}{Fator de 3 níveis utilizado para fazer controle
    local.}

\item{\code{teor}}{Teor de \eqn{P_{2}O_{5}} no solo, medidos em ppm.}

\item{\code{prod}}{Produção de grãos de feijão, medida em g/parcela.}

}}
\source{
RAMALHO et al. (2005), Tabela 13.6, pág. 243 e tabela 13.9,
    pág. 246.

    Pereira, A. F. Emprego de Covariância visando reduzir o efeito da
    heterogeneidade Ambiental nos Experimentos com a Cultura do
    Fejoeiro (Phaseolus vulgaris L.) Lavras: UFLA, 1993. 58p.
    (Dissertação-Mestrado Fitotecnia).
}
\description{
Experimento apresentado por Pereira (1993) referente a
    avaliação da produção de feijão e teor de \eqn{P_{2}O_{5}} em
    cada parcela, cujos dados foram obtidos de 10 cultivares de
    feijão. Para este experimento, um estudo por regressão deve
    considerar a produtividade de grãos como variável dependente e o
    teor como variável independente.
}
\examples{

data(RamalhoTb13.6)

str(RamalhoTb13.6)

library(lattice)

xyplot(prod ~ teor, groups = cult, data = RamalhoTb13.6,
       auto.key = list(space = "right", title = "Cultivar"),
       xlab = "Teor de P_{2}O_{5}",
       ylab = "Produção de grãos de feijão")

xyplot(prod ~ teor, groups = cult, data = RamalhoTb13.6,
       auto.key = list(space = "right",
                       title = "Cultivar", cex.title = 1.1),
       xlab = expression("Teor de"~P[2]*O[5]),
       ylab = "Produção de grãos de feijão (g/parcela)")

unit01 <- function(x) {
    x <- x - min(x)
    x <- x/max(x)
    return(x)
}

cex <- 0.5 + unit01(RamalhoTb13.6$teor)

key <- with(RamalhoTb13.6, {
    v <- round(seq(min(teor), max(teor), length.out = 4), digits = 0)
    cex <- 0.5 + unit01(v)
    list(title = expression(Teor~de~P[2]*O[5]),
         cex.title = 1.1,
         columns = length(v),
         text = list(as.character(v)),
         points = list(cex = cex, pch = 1))
})

xyplot(prod ~ cult, data = RamalhoTb13.6,
       cex = cex, key = key,
       xlab = expression("Teor de"~P[2]*O[5]),
       ylab = "Produção de grãos de feijão (g/parcela)")

}
\keyword{COV}
\keyword{DBC}

