% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoTb8.12.R
\name{RamalhoTb8.12}
\alias{RamalhoTb8.12}
\title{Produção Média de Grãos de Café-cereja}
\format{Um \code{data.frame} com 120 observações e 4 variáveis, em
    que

\describe{

\item{\code{prog}}{Fator que distingue as progênies do cultivar de
    café Icatu.}

\item{\code{ano}}{Fator de 3 níveis, colheitas realizadas em 
    anos sucessivos.}

\item{\code{bloc}}{Fator de 4 níveis do experimento, usado para 
    controle local.}

\item{\code{prod}}{Produção média de grãos de café-cereja, medido em
    kg/parcela nas unidades experimentais.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (pág. 128)
    
    Mendes, A. N. G. Avaliação de Metodologias Empregadas na Seleção
    de Progênies do Cafeeiro (Coffea arabica L.) no estado de Minas
    Gerais. Lavras: UFLA, 1994. 167p.
}
\description{
Experimento da produção média de grãos de café-cereja,
    dados obtidos por Mendes (1994). Delineamento em blocos
    casualizados.
}
\examples{

library(lattice)

data(RamalhoTb8.12)

xyplot(prod ~ prog | ano, data = RamalhoTb8.12,
       jitter.x = TRUE,
       xlab = "Progênie",
       ylab = "Produção")

aggregate(prod ~ prog,  data = RamalhoTb8.12,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

}
\keyword{DBC}

