% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosAnC4}
\alias{RamosAnC4}
\title{Processo Produtivo de Canetas}
\format{Um \code{data.frame} com 34 observações e 3 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{tamamostra}}{Tamanho da amostra.}

\item{\code{naoconf}}{Número de canetas não conformes na amostra.}

}}
\source{
RAMOS et al. (2013), pág. 135.
}
\description{
Conjunto de dados de um processo produtivo
    de canetas, com 34 amostras de tamanhos diferentes.
}
\examples{

data(RamosAnC4)

library(qcc)

qcc(RamosAnC4$naoconf,
    sizes = RamosAnC4$tamamostra,
    type = "p",
    xlab = "Amostra",
    ylab = "Número de canetas não conformes",
    title = "")

}
\keyword{CEQ}

