% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosAnC6}
\alias{RamosAnC6}
\title{Processo de Fabrica\enc{çã}{ca}o de Panelas}
\format{Um \code{data.frame} com 30 observações e 3 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{tamamostra}}{Tamanho da amostra de panelas avaliadas.}

\item{\code{naoconf}}{Número de não conformidades nas panelas.}

}}
\source{
RAMOS et al. (2013), pág. 136.
}
\description{
Conjunto de dados de um processo de fabricação de
    panelas, com 30 amostras com tamanhos diferentes.
}
\examples{

data(RamosAnC6)

library(qcc)

qcc(RamosAnC6$naoconf, sizes = RamosAnC6$tamamostra, type = "u",
    ylim = c(0,2.5), xlab = "Amostra", ylab = "Número de não
    conformidades em panelas", title = "")


}
\keyword{CEQ}

