% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosTb2.6}
\alias{RamosTb2.6}
\title{Resist\enc{ê}{e}ncia \enc{à}{a} Ruptura e Peso de Garrafas}
\format{Um \code{data.frame} com 25 observações e 2 variáveis, em que

\describe{

\item{\code{forca}}{Força de resistência à ruptura.}

\item{\code{peso}}{Peso da garrafa (em gramas).}

}}
\source{
RAMOS et al. (2013), pág. 30.
}
\description{
Resistência e peso de 25 garrafas.
}
\examples{

data(RamosTb2.6)
str(RamosTb2.6)

library(lattice)

xyplot(forca ~ peso, pch = 20,
       data = RamosTb2.6, type = c("g", "p"))

}
\keyword{CEQ}
\keyword{RS}

