% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosTb3.1}
\alias{RamosTb3.1}
\title{Teor de fl\enc{ú}{u}or de um processo qu\enc{í}{i}mico}
\format{Um \code{data.frame} com 75 observações e 2 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{fluor}}{Teor de flúor.}
    
}}
\source{
Ramos, E. M. L. S., et al.(2013). 
    Controle Estatístico da Qualidade (1th ed.). 
    Porto Alegre, RS: Bookman. (pg 38).
}
\description{
Conjunto de dados sobre teor de flúor com 15 
    amostras de tamanho 5 de um processo químico.
}
\examples{

data(RamosTb3.1)

library(qcc)

obj <- qcc.groups(RamosTb3.1$fluor, RamosTb3.1$amostra)

qcc(obj, type = "xbar", nsigmas = 3, xlab = 
     "Amostra", ylab = "Teor médio de fluor", title = "")

}
\keyword{CEQ}

