% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosTb4.1}
\alias{RamosTb4.1}
\title{Temperatura de Eletrodos de Carbono}
\format{Um \code{data.frame} com 75 observações e 2 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{temperat}}{Temperatura dos eletrodos de carbono (em
    \eqn{^\circ}C).}

}}
\source{
RAMOS et al. (2013), pág. 56.
}
\description{
Conjunto de dados referente às temperaturas de eletrodos
    de carbono, com 25 amostras de tamanho 8, em um processo de
    fabricação de alumínio.
}
\examples{

data(RamosTb4.1)

library(qcc)

obj <- qcc.groups(RamosTb4.1$temperat, RamosTb4.1$amostra)

qcc(obj, type = "xbar", nsigmas = 3, xlab = "Amostra", ylab =
     "Temperatura média", title = " ")

}
\keyword{CEQ}

