% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamosTb5.2.R
\name{RamosTb5.2}
\alias{RamosTb5.2}
\title{Itens não conformes no processo produtivo de ovos de galinha}
\format{Um \code{data.frame} com 30 observações e 3 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{subamostra}}{Número de subamostras conforme a amostra.}

\item{\code{naoconf}}{Número de unidades não conformes na amostra.}
    
}}
\source{
Ramos, E. M. L. S., et al.(2013). 
    Controle Estatístico da Qualidade (1th ed.). 
    Porto Alegre, RS: Bookman. (pg 79).
}
\description{
Número de itens não conformes de um processo de  
    produtivo de ovos de galinha com 15 amostras e subamostras de 
    tamanho diferentes.
}
\examples{

data(RamosTb5.2)

library(qcc)

obj <- qcc.groups(RamosTb5.2$naoconf, RamosTb5.2$amostra)

qcc(obj, type = "c", nsigmas = 3)

}
\keyword{CEQ}

