% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosTb5.8}
\alias{RamosTb5.8}
\title{Taxa de Defeitos na Fabrica\enc{çã}{ca}o de Copos de Cristal}
\format{Um \code{data.frame} com 26 observações e 3 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{tamamostra}}{Número de unidades amostradas.}

\item{\code{naoconf}}{Número de não conformidades na amostra.}

}}
\source{
RAMOS et al. (2013), pág. 96.
}
\description{
Número de não conformidades em um processo de fabricação
    de copos de cristal, com amostras de tamanhos diferentes.
}
\examples{

data(RamosTb5.8)
str(RamosTb5.8)

library(qcc)

qcc(RamosTb5.8$naoconf, sizes = RamosTb5.8$tamamostra, type = "u",
    xlab = "Amostra", ylab = "Taxa de não conformidades",
    title = "")

}
\keyword{CEQ}
