% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosTb6.1}
\alias{RamosTb6.1}
\title{Densidade Aparente de um Processo de Eletrodos}
\format{Um \code{data.frame} com 30 observações e 2 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{densi}}{Densidade aparente (em g/cm^{3} do eletrodo).}

}}
\source{
RAMOS et al. (2013), pág. 106.
}
\description{
Dados da densidade aparente de um processo de produtivo
    de eletrodos de carbono.
}
\examples{

data(RamosTb6.1)

library(qcc)

qcc(RamosTb6.1$densi, type = "xbar.one",
    std.dev = "SD", nsigmas = 3,
    xlab = "Amostra", ylab = "Densidade",
    title = "")

}
\keyword{CEQ}

