% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StorckTb2.R
\name{StorckTb2}
\alias{StorckTb2}
\title{Tabela referente ao peso das mudas}
\format{data.frame com 20 observações e 3 variáveis, em que
\describe{

\item{trat}{Fator de níveis nominais. Tratamento aplicado para
    controle do peso das plantas.}

\item{rep}{Número inteiro que identifica as repetições de cada
    tratamento.}

\item{peso}{Peso das plantas 30 dias após a semeadura.}

}}
\source{
Storck, L., & Garcia, B. C., & Lopes, S. J., & Estefanel, 
V. (2011). Experimentação Vegetal (3th ed.). Santa Maria, RS: ufsm.
(pg 21)
}
\description{
Experimento referente ao peso das plantas aos 30 dias
após a semeadura. Houve repetição de tratamentos.
}
\examples{

library(lattice)

data(StorckTb2)

xyplot(peso~rep|trat, data=StorckTb2, ylab='Peso das plantas',
xlab='Repetições', main='Dispersão dos pesos das plantas por 
tratamento')
}
\keyword{contínuo}
\keyword{peso}
\keyword{planta}

