% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storck.R
\name{StorckTb56}
\alias{StorckTb56}
\title{Experimento Bifatorial em Delineamento de Blocos Casualizados}
\format{Um \code{data.frame} com 48 observações e 4 variáveis, em que

\describe{

\item{\code{espec}}{Fator de 3 níveis que representa espécies.}

\item{\code{manejo}}{Fator de 4 níveis que representa formas de
    manejo.}

\item{\code{bloc}}{Fator de 4 níveis que são os blocos.}

\item{\code{result}}{Resultados de produção (kg/parcela).}

}}
\source{
STORCK et al. (2011), Tabela 56, pg 134.
}
\description{
Resultados (kg/parcela) em um experimento bifatorial 3
    \eqn{\times} 4 no delineamento de blocos ao acaso.
}
\examples{

library(lattice)

data(StorckTb56)
str(StorckTb56)

xyplot(prod ~ manejo, groups = espec, data = StorckTb56,
       type = c("p", "a"),
       xlab = "Tipos de manejo",
       ylab = "Produção (kg/parcela)",
       auto.key = list(title = "Espécie", cex.title = 1.1,
                       columns = 3))

}
\keyword{DBC}
\keyword{FAT2}

