% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StorckTb60.R
\name{StorckTb60}
\alias{StorckTb60}
\title{Resultados do experimento Bifatorial e Totais dos Tratamentos
    e Blocos.}
\format{Um \code{data.frame} com 48 observações e 4 variáveis.
\describe{

\item{\code{trat.A}}{ Os níveis do fator A são constituídos por 4 
    cultivares de milho. }

\item{\code{trat.D}}{ Os níveis do fator D são constituídos por 3
    foras de manejo. }

\item{\code{blocos}}{ Fator de 4 níveis constituídos por 4 blocos de 
    repetições  }

\item{\code{result}}{ Resultados (Kg/parcela) dos tratamentos.  }

}}
\source{
Storck, L., & Garcia, B. C., & Lopes, S. J., & Estefanel, 
V. (2011). Experimentaçãoo Vegetal (3th ed.). Santa Maria, RS: UFSM.
(Tabela 60, pág. 138)
}
\description{
Resultados fictícios (kg/parcela) de um experimento
    bifatorial 4x3, no delineamento de blocos ao acaso, com 4 
    repetições.
}
\examples{

library(lattice)

data(StorckTb101)

xyplot(result~trat.D|trat.A , data = StorckTb60, ylab = "Tratamento D", 
    xlab = "Tratamento A")
    
    
}
\keyword{DEA}

