% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storck.R
\name{StorckTb8}
\alias{StorckTb8}
\title{Experimento no Delimeamento Inteiramente Casualizado Com Um
    Fator}
\format{Um \code{data.frame} com 24 observações e 3 variáveis, em que

\describe{

\item{\code{trat}}{Fator com seis níveis que representam os
    tratamentos aplicados.}

\item{\code{rept}}{Inteiro que representa as repetições de cada
    tratamento.}

\item{\code{res}}{Representa a variável resposta observada.}

}}
\source{
STORCK et al. (2011), Tabela 8, pág. 33.
}
\description{
Dados referentes a um experimento conduzido no
    delineamento inteiramente casualizado. Não há expicação prática
    para as variáveis do experimento.
}
\examples{

library(lattice)

data(StorckTb8)
str(StorckTb8)

xyplot(res ~ trat, data = StorckTb8, jitter.x = TRUE,
       xlab = "Tratamentos", ylab = "Resposta")

}
\keyword{DIC}

