% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VieiraEx8.3.R
\name{VieiraEx8.3}
\alias{VieiraEx8.3}
\title{Experimento Em Blocos Casualizados Com Repetições Nos Blocos}
\format{Um \code{data.frame} com 24 observações e 3 variáveis, em que

\describe{

\item{\code{trat}}{Fator categóricos de 4 níveis que são os
    tratamentos aplicados.}

\item{\code{bloco}}{Fator categórico com 2 níveis que são os blocos
    do experimento.}

\item{\code{valor}}{Valor observado da resposta de cada unidade
    experimental.}

}}
\source{
Vieira, S. (1999). Estatística experimental (2th ed.).
    São Paulo, SP: Atlas. (Exercício 8.3, pág. 102).
}
\description{
Dados de um experimento em delineamento em blocos
    casualizados com repetições dos tratamentos dentro dos blocos.
}
\examples{

library(lattice)

data(VieiraEx8.3)
str(VieiraEx8.3)

xtabs(~bloco + trat, data = VieiraEx8.3)

xyplot(valor ~ trat, group = bloco, data = VieiraEx8.3,
       type = c("p", "a"),
       ylab = "Valor", xlab = "Tratamento")

}
\keyword{DBC}

