% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VieiraPg50.1.R
\name{VieiraPg50.1}
\alias{VieiraPg50.1}
\title{Experimento inteiramente casualizado}
\format{Um \code{data.frame} com 30 observações e 2 variáveis, em que

\describe{

\item{\code{trat}}{Fator de 5 níveis de tratamento.}

\item{\code{resp}}{Resposta em números inteiros.}

}}
\source{
Vieira, S.(1999). 
    Estatística experimental (2th ed.). 
    São Paulo, SP: Atlas. (pg 50 - Ex 2).
}
\description{
Experimento inteiramente casualizado com 
     5 tratamentos e resposta em inteiro.
}
\examples{

data(VieiraPg50.1)

library(lattice)

xyplot(resp ~ trat, data = VieiraPg50.1,
       xlab = "Tratamento",
       ylab = "Resposta")

}
\keyword{DIC}

