% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vieira.R
\name{VieiraPg57.2}
\alias{VieiraPg57.2}
\title{Peso de Ratos em Fun\enc{çã}{ca}o da Idade}
\format{Um \code{data.frame} com 15 observações e 3 variáveis, em que

\describe{

\item{\code{rato}}{Número do rato.}

\item{\code{idade}}{Fator de 5 níveis, da idade dos ratos em dias.}

\item{\code{peso}}{Peso em gramas.}

}}
\source{
VIEIRA (1999), pág. 57, exercício 4 .
}
\description{
Os dados referem-se ao peso de 3 ratos medidos 5 vezes
    dos 30 aos 46 dias de idade. Para fazer análise dos dados,
    pode-se considerar que os ratos são os blocos e que a idade é o
    fator de interesse.
}
\examples{

data(VieiraPg57.2)

str(VieiraPg57.2)

library(lattice)

xyplot(peso ~ idade, groups = rato,
      data = VieiraPg57.2, type = "b",
      xlab = "Idade (dias)",
      ylab = "Peso (g)")

}
\keyword{DBC}

