% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vieira.R
\name{VieiraTb4.1}
\alias{VieiraTb4.1}
\title{Produ\enc{çã}{ca}o de Variedades de Milho}
\format{Um \code{data.frame} com 20 observações e 2 variáveis, em que

\describe{

\item{\code{varied}}{Fator categórico de 4 níveis que indica as
    variedades de milho.}

\item{\code{prod}}{Produção de milho, medida em kg/100 m\eqn{^{2}}).}

}}
\source{
VIEIRA (1999), pág 44, tabela 4.1.
}
\description{
Experimento que mediu a produção, em kg/100
    m\eqn{^{-2}}), de 4 variedades de milho em um delineamento
    inteiramente casualizado.
}
\examples{

data(VieiraTb4.1)

str(VieiraTb4.1)

library(lattice)

xyplot(prod ~ varied, data = VieiraTb4.1,
      xlab = "Variedades",
      ylab = "Produção de milho")

}
\keyword{DIC}

