% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vieira.R
\name{VieiraTb5.3}
\alias{VieiraTb5.3}
\title{Produ\enc{çã}{ca}o de Variedades de Milho}
\format{Um \code{data.frame} com 20 observações e 3 variáveis, em que

\describe{

\item{\code{varied}}{Fator de níveis nominais que indicam a variedade
    do milho.}

\item{\code{bloc}}{Fator de 5 níveis, usado para controle local.}

\item{\code{prod}}{Produção de milho em kg/100 m\eqn{^{2}}.}

}}
\source{
VIEIRA (1999), pág. 53, tabela 5.3 .
}
\description{
Experimento da produção de milho de 4 diferentes
    variedades em um delineamento de blocos casualizados.
}
\examples{

data(VieiraTb5.3)

str(VieiraTb5.3)

library(lattice)

xyplot(prod ~ varied, groups = bloc,
      data = VieiraTb5.3,
      xlab = "Variedade",
      ylab = "Produção de milho")

}
\keyword{DBC}

