% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VieiraTb7.2.R
\name{VieiraTb7.2}
\alias{VieiraTb7.2}
\title{Números de ovos por poedeira}
\format{Um \code{data.frame} com 16 observações e 2 variáveis, em que

\describe{

\item{\code{tratamento}}{Tratamento ao qual a poedeira foi submetida.}

\item{\code{valor}}{Número de ovos da poedeira.}

}}
\source{
Vieira, S.(1999). Estatística experimental (2th ed.). 
    São Paulo, SP: Atlas. (Tabela 7.2 pág. 74).
}
\description{
Refere-se ao número de ovos por poedeira 35 dias após
    o início do experimento.
    Foram aplicados três tratamentos: música sertaneja (a), 
    música clássica (b) e música popular (c).
}
\examples{

library(lattice)

data(VieiraTb7.2)

xyplot(valor ~ tratamento, data = VieiraTb7.2,
       ylab = "Número de Ovos",
       xlab = "Tratamento")

}
\keyword{TODO}

