% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vieira.R
\name{VieiraTb8.5}
\alias{VieiraTb8.5}
\title{N\enc{ú}{u}mero de Itens Produzidos Em Fun\enc{çã}{ca}o do
    Tipo de M\enc{á}{a}quina}
\format{Um \code{data.frame} com 60 observações e 3 variáveis, em que

\describe{

\item{\code{maquina}}{Fator categórico com 5 níveis que representa os
    tipos de máquinas.}

\item{\code{bloco}}{Fator categóricos com 3 níveis que representa os
    operadores.}

\item{\code{qtd}}{Quantidade de itens produzidos por dia.}

}}
\source{
VIEIRA (1999), Tabela 8.5, pág. 98.
}
\description{
Os dados advém de um experimento no qual foram
    comparados cinco tipos de máquinas, operadas por três diferentes
    operadores (blocos).  O objetivo era verificar a suspeita de que
    o tipo de máquina usada na fabricação de determinado item tem
    efeito sobre a quantidade de itens produzidos por dia em uma
    fábrica.

    Os operadores foram tomados como blocos, pois já se sabia que
    existia diferença entre eles. Cada um deles trabalhou quatro dias
    em cada máquina (sorteadas aleatóriamente) e ao final de cada dia
    obteve-se a quantidade de itens produzidos por operador.
}
\examples{

library(lattice)

data(VieiraTb8.5)
str(VieiraTb8.5)

xtabs(~bloco + maquina, data = VieiraTb8.5)

xyplot(qtd ~ maquina, group = bloco, data = VieiraTb8.5,
       type = c("p", "a"),
       ylab = "Quantidade de itens produzidos por dia",
       xlab = "Tipo de máquina")

}
\keyword{DBC}
