% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb10.6}
\alias{ZimmermannTb10.6}
\title{Doses de Fungicida para Brusone na Cultura do Arroz}
\format{Um \code{data.frame} com 27 observações e 4 variáveis

\describe{

\item{\code{cult}}{Fator categórico que indica a cultivar do arroz.}

\item{\code{bloco}}{Fator categórico que identifica o bloco ao qual a
    observação pertence.}

\item{\code{dose}}{Fator métrico que representa a dose usada do
    fungicida, em gramas por hectare.}

\item{\code{ascpd}}{Logarítimo da área sob a curva de progresso da
    doença.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 10.6, pág. 201)
}
\description{
Dados de um experimento com dois fatores, em
    delineamento de parcelas divididas em blocos ao acaso. O
    experimento avaliou o efeito de 3 doses do fungicida Fongorene
    (0, 400 e 800 gramas por 100 kg) em cultivares de arroz. Mediu-se
    a percentagem de área foliar lesionada por brusone em cinco
    datas, e depois calculou-se a área sob a curva do progresso da
    doença (ASCPD). Os dados estão transformados em logaritmo.
}
\examples{

library(lattice)

data(ZimmermannTb10.6)

str(ZimmermannTb10.6)

xyplot(ascpd ~ dose, groups = cult, data = ZimmermannTb10.6,
       type = c("p", "a"),
       xlab = expression("Doses de inseticida"~(g~100~kg^{-1})),
       ylab = "Logaritmo da área sob a curva de progresso da doença")

}
\keyword{DBC}
\keyword{PS}

