% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb11.19}
\alias{ZimmermannTb11.19}
\title{Produtividade de Arroz Irrigado em Ensaio Fatorial com
    Confundimento}
\format{Um \code{data.frame} com 32 observações e 7 variáveis, em que

\describe{

\item{\code{bloc}}{Fator categórico que identifica o bloco.}

\item{\code{irri}}{Fator codificado que representa a irrigação:
    permanente e interminentente.}

\item{\code{adub}}{Fator codificado que representa a adubação: 30 e
    60 kg de N ha\eqn{^{-1}}.}

\item{\code{cult}}{Fator codificado que representa a cultivar de
    arroz: IAC 435 e IR 841-63-5-I-9-33.}

\item{\code{doen}}{Fator codificado que representa o controle das
    doenças: sem controle e controle com aplicação de Manzate.}

\item{\code{inva}}{Fator codificado que representa o controle das
    plantas invasoras: aplicação de Stan F-34 e Ronstar.}

\item{\code{prod}}{Produtividade de arroz irrigado (kg
    ha\eqn{^{-1}}).}

}}
\source{
ZIMMERMANN (2004), (Tabela 11.19, pág. 237)
}
\description{
Dados de um experimento fatorial \eqn{2^5}, com
    confundimento da interação de quinta ordem.  Estudou-se o efeito
    dos cinco principais problemas da cultura: controle de invasoras,
    adubação, irrigação, controle de doenças e cultivar. A interação
    de quarta ordem (5 fatores) foi confundida com blocos.  Os dados
    da produtividade do arroz estão em kg/ha.
}
\examples{

library(lattice)
library(latticeExtra)

data(ZimmermannTb11.19)
str(ZimmermannTb11.19)

ftable(xtabs(~irri + cult + adub + doen + inva,
             data = ZimmermannTb11.19))

useOuterStrips(
    xyplot(prod ~ factor(irri) | factor(inva) + factor(doen),
           data = ZimmermannTb11.19,
           groups = interaction(cult, adub, sep = ":"),
           type = c("p", "a"),
           xlab = "Irrigação",
           ylab = expression("Produtividade"~(kg~ha^{-1})),
           auto.key = list(title = "Cultivar:Adubação",
                           cex.title = 1.1,
                           columns = 4)),
    strip = strip.custom(strip.names = TRUE,
                         var.name = "Invasora"),
    strip.left = strip.custom(strip.names = TRUE,
                              var.name = "Doença"))

}
\keyword{DBC}
\keyword{FAT2K}
\keyword{confundimento}
