% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb11.7.R
\name{ZimmermannTb11.7}
\alias{ZimmermannTb11.7}
\title{Espaçamento, Densidade e Nitrogênio na Produção de Arroz}
\format{Um \code{data.frame} com 32 observações e 7 variáveis

\describe{

\item{\code{rept}}{Fator categórico que identifica a repetição. Cada
    repetição contém quatro blocos de tamanho 2.}

\item{\code{bloco}}{Fator de níveis categóricos que identifica o
    bloco em cada repetição.}

\item{\code{dens}}{Fator codificado que representa a densidade de
    plantio (50 e 90).}

\item{\code{espac}}{Fator codificado que representa o espaçamento
    entre as linhas de semeadura (35 e 50 cm).}

\item{\code{adub}}{Fator codificado que indica a presença ou não de
    adubação nitrogenada (1 e 0).}

\item{\code{prod}}{Produção de arroz, em kg ha\eqn{^{-1}}.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 11.7, pág 226)
}
\description{
Dados de um experimento fatorial \eqn{2^3}, com
    confundimento de duas das interações simples. Referem-se à
    produtividade de grãos, em kg ha\eqn{^{-1}}. O ensaio foi
    conduzido em 4 blocos ao acaso de tamanho 2 dentro de quatro
    repetições, em que cada uma continha 4 blocos (8 parcelas). Os
    fatores foram: densidades de plantio de 50 e 90 sementes por
    metro, espaçamentos entre linhas de 35cm e 50cm e fator presença
    ou ausência de adubação nitrogenada em cobertura.
}
\examples{

library(lattice)

data(ZimmermannTb11.7)

str(ZimmermannTb11.7)

ftable(xtabs(~adub + espac + dens + bloco, data = ZimmermannTb11.7))

xyplot(prod ~ factor(dens) | factor(adub), groups = espac,
       data = ZimmermannTb11.7,
       type = c("p", "a"),
       xlab = "Densidade de semeadura",
       ylab = expression ("Produção de Arroz"~(kg~ha^{-1})),
       auto.key = list(title = "Espaçamento", cex.title = 1.1,
                       columns = 2),
       strip = strip.custom(strip.names = TRUE,
                            var.name = "Adubação"))

}
\seealso{
\code{\link{ZimmermannTb11.1}}. Os ensaios
    \code{ZimmermannTb11.1} e \code{ZimmermannTb11.7} possuem os
    mesmos valores observados da resposta e fatores estudados, o que
    muda é a estrutura de confundimento utilizada, com modificação do
    tamanho dos blocos para que isso fosse ilustrado. O primeiro tem
    2 blocos de tamanho 4 em cada repetição e o segundo tem 4 blocos
    de tamanho 2. Essa modificação dos dados é artificial e foi feita
    para fins didáticos. Não se deve alterar o delineamento de dados
    reais para conduzir as análises.
}
\keyword{DBC}
\keyword{FAT}
\keyword{confundimento}

