% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb12.20}
\alias{ZimmermannTb12.20}
\title{Produtividade de Gr\enc{ã}{a}os de Cultivares de
    Feij\enc{ã}{a}o}
\format{Um \code{data.frame} com 36 observações e 3 variáveis

\describe{

\item{\code{cult}}{Fator de níveis nominais. dentifica a cultivar de
    feijão.}

\item{\code{bloco}}{Número inteiro que identifica o bloco da
    observação.}

\item{\code{prod}}{Produtividade de grãos em kg ha\eqn{^{-1}}.}

}}
\source{
ZIMMERMANN (2004), Tabela 12.20, pág. 258.
}
\description{
Dados do ensaio 2 de um experimento em DBC de competição
    de cultivares de feijão do grupo preto. Este ensaio foi realizado
    no município de Santo Antonio de Goiás (GO). Houveram três
    cultivares testemunhas e mais seis em cada ensaio.  Cada
    experimento foi conduzido no período de inverno, sob irrigação
    por aspersão. Os dados são relativos à produtividade de grãos, em
    kg/ha.
}
\examples{

library(lattice)

data(ZimmermannTb12.20)

xyplot(prod ~ cult , groups = bloco, data = ZimmermannTb12.20,
       type = c("p", "a"),
       xlab = "Tratamentos",
       ylab = "Produtividade de grãos")

aggregate(prod ~ cult, data = ZimmermannTb12.20,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })
}
\keyword{DBC}

