% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb12.32}
\alias{ZimmermannTb12.32}
\title{Competi\enc{çã}{ca}o de Cultivares de Feij\enc{ã}{a}o}
\format{Um \code{data.frame} com 13 observações e 4 variáveis

\describe{

\item{\code{geno}}{Fator de níveis nominais. Identifica o genótipo
    repetição da observação.}

\item{\code{irrigados}}{Médias nos ensaios irrigados.}

\item{\code{sequeiro}}{Médias nos ensaios de sequeiro.}

\item{\code{geral}}{Média geral.}

}}
\source{
ZIMMERMANN (2004), Tabela 12.32, pág. 264.
}
\description{
Dados de um estudo em delineamento de blocos completos
    com quatro repetições. O experimento foi um ensaio de competição
    de cultivares e linhagens de feijão, em cultivos irrigado e de
    sequeiro.  São apresentadas na tabela a média nos ensaios
    irrigados, nos de sequeiro e geral.
}
\examples{

library(lattice)

data(ZimmermannTb12.32)

str(ZimmermannTb12.32)

xyplot(geral ~ geno, data = ZimmermannTb12.32,
       type = "o", jitter.x = TRUE,
       xlab = "Genótipo",
       ylab = "Média Geral",
       main = "Competição de Cultivares de Feijão",
       scales = list(x = list(rot = 90)))
}
\keyword{DBC}

