% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb12.33.R
\name{ZimmermannTb12.33}
\alias{ZimmermannTb12.33}
\title{Competi\enc{çã}{ca}o de Cultivares de Feij\enc{ã}{a}o}
\format{Um \code{data.frame} com 13 observações e 4 variáveis

\describe{

\item{\code{geno}}{Fator de níveis nominais. Identifica o genótipo repetição 
    da observação.}

\item{\code{irrigados}}{Totais nos ensaios irrigados.}

\item{\code{sequeiro}}{Totais nos ensaios de sequeiro.}

\item{\code{geral}}{Total geral.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 12.33, pág 265)
}
\description{
Dados de um estudo em delineamento de blocos completos com 
    quatro repetições. O experimento foi um ensaio de competição de
    cultivares e linhagens de feijão, em cultivos irrigado e de sequeiro.
    São apresentadas na tabela o total nos ensaios irrigados, nos de sequeiro
    e geral.
}
\examples{

library(lattice)

data(ZimmermannTb12.33)

str(ZimmermannTb12.33)

xyplot(geral ~ geno, data = ZimmermannTb12.33,
       type = "o", jitter.x = TRUE,
       xlab = "Genótipo",
       ylab = "Total Geral",
       main = "Competição de Cultivares de Feijão",
       scales = list(x = list(rot = 90)))
}
\keyword{DBC}

