% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb12.7}
\alias{ZimmermannTb12.7}
\title{\enc{Á}{A}rea Sob a Curva do Progresso de Brusone}
\format{Um \code{data.frame} com 18 observações e 3 variáveis

\describe{

\item{\code{cult}}{Fator de níveis nominais. Indica a cultivar.}

\item{\code{bloco}}{Número inteiro que identifica o bloco da
    observação.}

\item{\code{aacpd}}{Logaritmo natural da área sob a curva de
    progresso da doença. A unidade de medida não é conhecida.}

}}
\source{
ZIMMERMANN (2004), Tabela 12.7, pág. 251.
}
\description{
Dados do ensaio 1 de um experimento que avaliou as
    cultivares para a área foliar atacada por brusone
    (\emph{Pyricularia Orizae} L.) em diferentes datas e se calculou
    área sob a curva do progresso da doença. Este primeiro
    experimento foi semeado na densidade de oitenta sementes por
    metro. Os dados foram transformados por logaritmo natural,
    procurando-se uma maior homogeneização das variâncias.
}
\examples{

library(lattice)

data(ZimmermannTb12.7)
str(ZimmermannTb12.7)

xyplot(aacpd ~ cult , groups = bloco, data = ZimmermannTb12.7,
       type = c("p", "a"),
       xlab = "Tratamentos",
       ylab = "Logaritmo da área sob a curva de progresso da doença")

aggregate(aacpd ~ cult, data = ZimmermannTb12.7,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

}
\keyword{DBC}

