% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb13.1.R
\name{ZimmermannTb13.1}
\alias{ZimmermannTb13.1}
\title{\enc{Í}{I}ndice de \enc{Á}{A}rea de Folhar e Idade das Plantas}
\format{Um \code{data.frame} com 8 observações e 5 variáveis

\describe{

\item{\code{idade}}{Variável métrica que a idade das plantas.}

\item{\code{IAF}}{Índice de área foliar.}

\item{\code{x2}}{IAF elevada ao quadrado.}

\item{\code{y2}}{Idade elevada ao quadrado.}

\item{\code{xy}}{Produto entre idade e IAF.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 13.1, pág 272)
}
\description{
Em um plantio de feijão com a cultivar Jalo Precoce
    foram coletadas plantas em dias sequenciais (Idades), e o índice
    de área foliar (IAF) foi medido.
}
\examples{

library(lattice)

data(ZimmermannTb13.1)
str(ZimmermannTb13.1)

splom(ZimmermannTb13.1, type = c("p", "smooth"))
}
\keyword{REG}

