% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb15.10}
\alias{ZimmermannTb15.10}
\title{Aduba\enc{çã}{ca}o NPK na Produ\enc{çã}{ca}o de Arroz}
\format{Um \code{data.frame} com 25 observações e 3 variáveis, em que

\describe{

\item{\code{N}}{Fator codificado que representa a dose de
    nitrogênio.}

\item{\code{P}}{Fator codificado que representa a dose de fósforo.}
\item{\code{K}}{Fator codificado que representa a dose de potássio.}

\item{\code{prod}}{Produção de grãos, kg ha\eqn{^{-1}}.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 15.10, pág. 314)
}
\description{
Dados de produção de grãos em terras altas, em kg
    ha\eqn{^{-1}}.  Este é um ensaio contendo apenas uma repetição de
    um fatorial fracionado correspondente a 1/5 de um \eqn{5^3}.
}
\examples{

library(lattice)

data(ZimmermannTb15.10)
str(ZimmermannTb15.10)

# Arranjo de fatores como a estrutura de quadrado latino.
reshape::cast(data = ZimmermannTb15.10,
              formula = N ~ P, value = "K")

xyplot(prod ~ N + P + K, data = ZimmermannTb15.10,
       outer = TRUE, as.table = TRUE, type = c("p", "a"),
       xlab = "Níveis de NPK",
       ylab = "Produção de grãos")

}
\keyword{FRAC}
\keyword{superficie}

