% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb15.4.R
\name{ZimmermannTb15.4}
\alias{ZimmermannTb15.4}
\title{Fatorial Fracionado para Estudo da Adubação em Arroz}
\format{Um \code{data.frame} com 27 observações e 3 variáveis

\describe{

\item{\code{bloco}}{Fator de níveis categóricos que identifica os
    blocos.}

\item{\code{zinco}}{Fator de níveis codificados que indica o nível de
    zinco aplicado: 0, 5 e 10 kg ha\eqn{^{-1}}.}

\item{\code{fosf}}{Fator de níveis codificados que indica o nível de
    fósforo aplicado: 0, 50 e 100 kg ha\eqn{^{-1}}.}

\item{\code{calc}}{Fator de níveis codificados que indica o nível de
    calcário aplicado: 0, 1.5 e 3 kg ton ha\eqn{^{-1}}.}

\item{\code{prod}}{Produtividade de grãos, em kg ha\eqn{^{-1}}.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 15.4, pág. 309)
}
\description{
Dados de um ensaio com fração 1/3 de um fatorial
    \eqn{3^3}, conduzido em blocos ao acaso. O estudo é sobre a
    absorção e translocação de zinco em arroz de terras altas sob a
    influência de fósforo e calcário. Os níveis dos fatores eram de
    0, 5 e 10 kg ha\eqn{^{-1}} para o zinco, 0, 50 e 100 kg
    ha\eqn{^{-1}} para o fósforo e 0, 1.5 e 3 ton ha\eqn{^{-1}} para
    o calcário. Os dados são de produção de grãos, em kg
    ha\eqn{^{-1}}.
}
\examples{

library(lattice)

data(ZimmermannTb15.4)
str(ZimmermannTb15.4)

xyplot(prod ~ zinco | factor(fosf), groups = calc,
       data = ZimmermannTb15.4, as.table = TRUE,
       xlab = "Níveis codificados de zinco",
       ylab = expression("Produção de grãos"~(kg~ha^{-1})),
       auto.key = list(title = "Calcário", cex.title = 1.1,
                       columns = 3),
       strip = strip.custom(strip.names = TRUE,
                            var.name = "Fósforo"))

}
\seealso{
Estes dados são na realiadade uma adaptação dos dados em
    \code{\link{ZimmermannTb11.13}} pois referem-se ao conjunto dos 3
    blocos (do total de 9) continham a combinação 000 de zinco,
    fósforo e cálcio.
}
\keyword{DBC}
\keyword{FRAC}

