% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb16.1.R
\name{ZimmermannTb16.1}
\alias{ZimmermannTb16.1}
\alias{ZimmermannTb16.2}
\title{An\enc{á}{a}lise de Composi\enc{çã}{ca}o do Solo Aluviais}
\format{Um \code{data.frame} com 7 observações e 7 variáveis, em que

\describe{

\item{\code{origem}}{Fator de níveis nominais que identifica o local
    de origem da amostra de solo.}

\item{\code{mo}}{Matéria orgânica no solo (\%).}

\item{\code{fosf}}{Disponibilidade de fósforo no solo (mg
    kg\eqn{^{-1}}).}

\item{\code{pmo}}{Posto da váriavel \code{mo}.}

\item{\code{pfosf}}{Posto da váriavel \code{fosf}.}

\item{\code{Nc}}{Indica o número de pares de postos concordantes.}

\item{\code{Nd}}{Indica o número de pares de postos discordantes.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 16.1, pág. 327)
}
\description{
Dados de análise de solos aluviais, que apresentaram pH
    abaixo de sete, referentes à associação existente entre teor de
    matéria orgânica no solo (porcentagem) com disponibilidade de
    fósforo (mg kg\eqn{^{-1}}). A amostra tem apenas sete
    observações.
}
\examples{

data(ZimmermannTb16.1)
str(ZimmermannTb16.1)

mean(ZimmermannTb16.1$mo)
mean(ZimmermannTb16.1$fosf)

with(ZimmermannTb16.1, {
    cbind(pmo = rank(mo), pfosf = rank(fosf))
})

}
\keyword{correlacao}

