% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb16.3.R
\name{ZimmermannTb16.3}
\alias{ZimmermannTb16.3}
\title{Teores de Mat\enc{é}{e}ria Org\enc{â}{a}nica de Solos de V\enc{á}{a}rzea}
\format{Um \code{data.frame} com 14 observações e 2 variáveis, em que

\describe{

\item{\code{solo}}{Tipo de solo, sendo A - aluviais e B - gley
    húmicos.}

\item{\code{mo}}{Teor de matéria orgânica do solo (\%).}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 16.3, pág. 337)
}
\description{
Teores de matéria orgânica de solos aluviais e gely
    húmicos.
}
\examples{

# Postos.
rank(ZimmermannTb16.3$mo)

aggregate(mo ~ solo, data = ZimmermannTb16.3, FUN = mean)

}
\keyword{TODO}

