% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb16.4}
\alias{ZimmermannTb16.4}
\title{Hastes Mortas de Arroz por \emph{Elasmopalpus lignosellus} L.}
\format{Um \code{data.frame} com 9 observações e 3 variáveis, em que

\describe{

\item{\code{cult}}{Fator categórico que identifica as cultivares de
    arroz.}

\item{\code{nmort}}{Número de hastes mortas por \emph{Elasmopalpus
    lignosellus} por parcela.}

\item{\code{posto}}{Posto da váriavel \code{nmort}.}

}}
\source{
ZIMMERMANN (2004), Tabela 16.4, pág. 341.
}
\description{
Dados de um experimento em delineamento inteiramente
    casualizado que estudou cultivares de arroz, em casa de
    vegetação. Os dados são referentes ao número de hastes mortas por
    \emph{Elasmopalpus lignosellus} por parcela. Cada cultivar foi
    repetida três vezes.
}
\examples{

library(lattice)

data(ZimmermannTb16.4)
str(ZimmermannTb16.4)

xyplot(nmort ~ cult, data = ZimmermannTb16.4,
       xlab = "Cultivar",
       ylab = "Número de hastes mortas")

rank(ZimmermannTb16.4$nmort)

}
\keyword{DIC}
\keyword{contagem}

