% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb16.5}
\alias{ZimmermannTb16.5}
\title{N\enc{ú}{u}mero de Dias para Flora\enc{çã}{ca}o de Cultivares
    de Arroz}
\format{Um \code{data.frame} com 24 observações e 4 variáveis, em que

\describe{

\item{bloco}{Fator categórico que indica o bloco.}

\item{cult}{Fator categórico que indica a cultivar.}

\item{dias}{Número total de dias até a floração das plantas.}

\item{posto}{Posto corresponden ao número de dias de cada cultivas
    dentro de cada bloco. O posto pode ser calculado com a função
    \code{\link[base]{rank}}.}

}}
\source{
ZIMMERMANN (2004), Tabela 16.5, pág 347.
}
\description{
Experimento realizado em delineamento em blocos
    completos ao acaso, utilizando oito cultivares de arroz
    irrigado. Os dados são referentes ao número de dias até a
    floração das plantas de arroz, uma variável do tipo tempo até o
    evento/desfecho. São apresentados na tabela também os respectivos
    postos de cada dado, a fim de facilitar a aplicação do teste de
    Friedman (teste não paramétrico).
}
\examples{

library(lattice)

data(ZimmermannTb16.5)
str(ZimmermannTb16.5)

# Como calcular o posto no R.
p <- by(data = ZimmermannTb16.5,
        INDICES = ZimmermannTb16.5$bloco,
        FUN = function(dataset) {
            dataset$posto <- rank(dataset$dias)
            return(dataset)
        })
ZimmermannTb16.5 <- do.call(rbind, p)
rownames(ZimmermannTb16.5) <- NULL

xyplot(dias + posto ~ cult, outer = TRUE, groups = bloco,
       data = ZimmermannTb16.5, jitter.x = TRUE,
       scales = list(y = list(relation = "free"),
                     x = list(rot = 90)),
       xlab = "Cultivares", ylab = "",
       strip = strip.custom(factor.levels = c("Dias para floração",
                                              "Posto")))

xyplot(posto ~ dias, data = ZimmermannTb16.5,
       xlab = "Dias para floração", ylab = "Posto",
       jitter.y = TRUE)

}
\keyword{DIC}
\keyword{sobrevivência}

