% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb4.11.R
\name{ZimmermannTb4.11}
\alias{ZimmermannTb4.11}
\title{Alturas Médias de Perfilhos}
\format{Um \code{data.frame} com 200 observações e 4 variáveis

\describe{

\item{\code{geno}}{Fator de níveis nominais. Identifica o genótipo
    da planta.}

\item{\code{bloco}}{Número inteiro que identifica o bloco da
    observação.}

\item{\code{amostra}}{Fator de níveis numéricos. Identifica à qual
    amostra pertence a observação.}

\item{\code{alt}}{Altura de perfilhos (cm).}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (Tabela 4.11, pág 79)
}
\description{
Dados de um ensaio com dez genótipos, quatro blocos e
    cinco amostras por parcela, tomadas ao acaso, das alturas dos
    perfilhos, medidos em cm.
}
\examples{

library(lattice)

data(ZimmermannTb4.11)

str(ZimmermannTb4.11)

xyplot(alt ~ geno, groups = bloco,
       data = ZimmermannTb4.11,
       type = c("p", "a"), jitter.x = TRUE,
       xlab = "Tratamentos",
       ylab = "Altura média de perfilhos (cm)",
       scales=list(x=list(rot=90)))

aggregate(alt ~ geno, data = ZimmermannTb4.11,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

}
\keyword{DBC}

