% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb7.1}
\alias{ZimmermannTb7.1}
\title{Ensaio de Competi\enc{çã}{ca}o de Cultivares em Reticulado
    Quadrado}
\format{Um \code{data.frame} com 75 observações e 4 variáveis, em que

\describe{

\item{\code{rept}}{Fator categórico que identifica os grupos de
    blocos que formam uma repetição. Cada repetição tem 5 blocos de
    tamanho 5.}

\item{\code{bloc}}{Fator categórico que identifica os blocos dentro
    das repetições.}

\item{\code{cult}}{Fator categórico que representa as cultivares de
    arroz.}

\item{\code{prod}}{Produção de arroz, kg ha\eqn{^{-1}}.}

}}
\source{
ZIMMERMANN (2004), Tabela 7.1, pág. 136.
}
\description{
Experimento feito pelos pesquisadores Orlando Peixoto
    Moraes, Emilio da Maia de Castro e Flavio Breseghello, da Embrapa
    Arroz e Feijão, na fazenda Capivara em Santo Antônio de Goiás. O
    delineamento reticulado quadrado 5 \eqn{\times} 5 acomodou 25
    cultivares, cada uma repetida 3 vezes. Foi medida a produção das
    cultivares.
}
\examples{

library(lattice)

data(ZimmermannTb7.1)
str(ZimmermannTb7.1)

ftable(xtabs(~rept + cult, data = ZimmermannTb7.1))
ftable(xtabs(~rept + bloc + cult, data = ZimmermannTb7.1))

xyplot(prod ~ reorder(cult, prod), groups = rept,
       data = ZimmermannTb7.1,
       type = c("p", "a"),
       xlab = "Cultivares",
       ylab = expression("Produção"~(kg~ha^{-1})))

}
\keyword{LAT}
