% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.27.R
\name{AndradeTb2.27}
\alias{AndradeTb2.27}
\title{Estudo da Eficiência de uma Vacina Antigripal}
\format{Um \code{data.frame} com 6 observações e 3 variáveis, em que

\describe{

\item{\code{saude}}{Estado de saúde dos habitantes (gripados ou não
    gripados).}

\item{\code{vacina}}{identifica em quantas doses a vacina antigripal
    foi aplicada (0, não houve aplicação; 1, em dose única; 2, em
    duas doses).}

\item{\code{freq}}{Número de habitantes no cruzamento de \code{saude}
    e \code{vacina}.}

}}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Tabela 2.27, pág. 108)
}
\description{
Estudo para avaliar a eficiência de uma nova vacina
    antigripal, a qual foi administrada aos membros de uma pequena
    comunidade, em duas doses, ao longo de duas semanas. Algumas
    pessoas tomaram as duas doses, outras tomaram apenas a primeira
    dose e outras não tomaram nenhuma dose. Os tratamentos foram
    aplicados em 1000 habitantes dessa comunidade.
}
\examples{

data(AndradeTb2.27)
str(AndradeTb2.27)

xt <- xtabs(freq ~ ., data = AndradeTb2.27)
xt

mosaicplot(xt, col = c("darkturquoise", "lawngreen", "blue"),
           cex.axis = 0.8, main = "")

library(lattice)
barchart(t(xt),
         horizontal = FALSE,
         stack = TRUE,
         auto.key = list(
             corner= c(0.9, 0.9),
             title = "Estado de saúde",
             cex.title = 1.1),
         xlab = "Número de doses da vacina",
         ylab = "Número de habitantes")

}
\keyword{contigência}

