% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndradeTb2.42.R
\name{AndradeTb2.42}
\alias{AndradeTb2.42}
\title{Peso de Carne de Mexilhões}
\format{Um \code{data.frame} com 30 observações e 2 variáveis, em que

\describe{

\item{\code{local}}{Fator de 2 níveis qualitativos, que são os locais
    habituais dos mexilhões.}

\item{\code{peso}}{Peso de carne de mexilhões, em gramas.}

}}
\source{
Andrade, D. F., Ogliari, P. J. (2010). Estatística para as
    ciências agrárias e biológicas com noções de experimentação (2nd
    ed.). Florianópolis, SC. Editora da UFSC. (Tabela 2.42, pág. 146)
}
\description{
Resultados da variável peso de carne, em gramas, de
    mexilhões de dois locais: 1) Sambaqui e 2) Manguezal.
}
\examples{

data(AndradeTb2.42)
str(AndradeTb2.42)

library(lattice)
boxplot(peso ~ local,
        data = AndradeTb2.42,
        xlab = "Local",
        ylab = "Peso de Carne (em gramas)")

}
\keyword{AAS}

