% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEg4.2}
\alias{CharnetEg4.2}
\title{Avalia\enc{çã}{ca}o de Vendedores pelos Clientes}
\format{Um \code{data.frame} com 2 colunas e 10 linhas.

\describe{

\item{\code{notas}}{Nota atribuída ao vendedor, em escala de 0 a
    100.}

\item{\code{vendas}}{Volume anual de vendas, em milhões de reais.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 4, exemplo 4.2,
    pág. 79, Capítulo 5, exercício 4, pág. 109, Capítulo 5, exercício
    7, pág. 111, Capítulo 6, exercício 1, pág. 142)
}
\description{
Análise das vendas através das notas atribuídas pelos
    clientes aos vendedores de uma empresa de seguro.
}
\examples{

data(CharnetEg4.2)

plot(CharnetEg4.2)

}
\keyword{RS}

