% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEg6.4}
\alias{CharnetEg6.4}
\title{Efeito de um Produto Qu\enc{í}{i}mico no Peso de
    Girass\enc{ó}{o}is}
\format{Um \code{data.frame} com 2 colunas e 20 linhas, em que

\describe{

\item{\code{peso}}{Peso da planta, em gramas.}

\item{\code{dose}}{Dose da aplicação, em \eqn{\log(g cm^{-2})}.}

}}
\source{
CHARNET et al. (2008), Capítulo 6, exemplo 4, pág. 136.
}
\description{
Medição do efeito da adição de uma solução de um certo
   produto químico na regação de girassóis.
}
\examples{

data(CharnetEg6.4)
str(CharnetEg6.4)

plot(peso ~ dose, data = CharnetEg6.4)

}
\keyword{RS}

