% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEg7.3}
\alias{CharnetEg7.3}
\title{Rela\enc{çã}{ca}o entre Sal\enc{á}{a}rio, Tempo de
    Experi\enc{ê}{e}ncia e Sexo}
\format{Um \code{data.frame} com 2 colunas e 15 linhas, em que

\describe{

\item{\code{salario}}{Salário, em mil reais.}

\item{\code{exp}}{Tempo de experiência, em anos completos.}

\item{\code{sexo}}{Sexo do gerente de agência bancária.}

}}
\source{
CHARNET et al. (2008), Capítulo 7, exemplo 3, pág. 152.
}
\description{
Dados de salário, tempo de experiência e sexo dos
    gerentes de agências bancárias de um grande banco.
}
\examples{

data(CharnetEg7.3)
str(CharnetEg7.3)

library(lattice)

with(CharnetEg7.3, plot(salario, exp, col = sexo))
xyplot(exp ~ salario, data = CharnetEg7.3, groups = sexo)

}
\keyword{RM}
\keyword{dummy}

