% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEg8.2.R
\name{CharnetEg8.2}
\alias{CharnetEg8.2}
\title{Conjunto de Dados Genérico}
\format{Um \code{data.frame} com 2 colunas e 15 linhas.

\describe{

\item{\code{x}}{Variável independente, sem interpretação.}

\item{\code{y}}{Variável dependente, sem interpretação.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 8, exer. 2, p. 191)
}
\description{
Um conjunto de dados qualquer para exercício de ajuste
    de modelos de regressão linear com diferentes preditores.
}
\examples{

data(CharnetEg8.2)

plot(CharnetEg8.2)

}
\keyword{RS}

