% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEg9.2.R
\name{CharnetEg9.2}
\alias{CharnetEg9.2}
\title{Relação entre Poluentes Atmosféricos}
\format{Um \code{data.frame} com 3 colunas e 7 linhas.

\describe{

\item{\code{CO}}{Monóxido de carbono, sem especificação.}

\item{\code{HC}}{Hidrocarbonetos, sem especificação.}

\item{\code{NO}}{Óxido de nitrogênio, sem especificação.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 9, exemplo 2,
    pág. 216)
}
\description{
Relação de poluentes atmosféricos emitidos por carros
    durante uma coleta realizada em sete períodos diferentes em um
    grande centro.
}
\examples{

data(CharnetEg9.2)

plot(CharnetEg9.2)
cor(CharnetEg9.2)

}
\keyword{RM}

