% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEg9.4}
\alias{CharnetEg9.4}
\title{Um Conjunto de Dados Gen\enc{é}{e}rico para Regress\enc{ã}{a}o Linear M\enc{ú}{u}ltipla}
\format{Um \code{data.frame} com 8 colunas e 36 linhas.

\describe{

\item{\code{y}}{Variável dependente.}

\item{\code{x1}}{Variável regressora.}

\item{\code{x2}}{Variável regressora.}

\item{\code{x3}}{Variável regressora.}

\item{\code{x4}}{Variável regressora.}

\item{\code{x5}}{Variável regressora.}

\item{\code{x6}}{Variável regressora.}

\item{\code{x7}}{Variável regressora.}


}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 9, exemplo 4,
    pág. 226)
}
\description{
Um conjunto de dados qualquer para exercício de ajuste
    de um modelo de regressão linear múltipla onde temos uma variável
    dependente relaciona a sete outras variáveis independentes.
}
\examples{

data(CharnetEg9.4)

plot(CharnetEg9.4)

}
\keyword{RM}

