% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx1.20}
\alias{CharnetEx1.20}
\title{Testes de Avalia\enc{çã}{ca}o de Personalidade}
\format{Um \code{data.frame} com 2 colunas e 20 linhas.

\describe{

\item{\code{t1}}{Resultados do teste I, em pontos.}

\item{\code{t2}}{Resultados do teste II, em pontos.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 1, exercício
    20, pág. 26; e Capítulo 3, exercício 8, pág. 67)
}
\description{
Dois testes de avaliação de personalidade aplicados a 20
    indivíduos.
}
\examples{

data(CharnetEx1.20)

plot(CharnetEx1.20)

}
\keyword{RS}

