% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx1.6}
\alias{CharnetEx1.6}
\title{Press\enc{ã}{a}o Sangu\enc{í}{i}nea Sist\enc{ó}{o}lica}
\format{Um \code{vetor} numérico com 16 observações.}
\source{
CHARNET et al. (2008), Capítulo 1, exercício 6, pág. 23.
}
\description{
Pressão sanguínea sistólica de um grupo de 16 pacientes
    de uma clínica.
}
\examples{

data(CharnetEx1.6)
str(CharnetEx1.6)

(m <- mean(CharnetEx1.6))
(s <- sd(CharnetEx1.6))
fivenum(CharnetEx1.6)

curve(dnorm(x, m, s),
      xlim = extendrange(CharnetEx1.6, f = 0.2),
      col = 4, lwd = 2)
hist(CharnetEx1.6, prob = TRUE, add = TRUE)
rug(CharnetEx1.6)

}
\keyword{AAS}
