% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx11.2.R
\name{CharnetEx11.2}
\alias{CharnetEx11.2}
\title{Distribuição de Trabalho em um Departamento de Contabilidade}
\format{Um \code{data.frame} com 7 colunas e 30 linhas.

\describe{

\item{\code{nhora}}{Número de horas trabalhadas por dia.}

\item{\code{ncheq}}{Número de cheques descontados.}

\item{\code{npag}}{Número de pagamentos recebidos pelos
    funcionários.}

\item{\code{ndoc}}{Número de documentos processados e enviados ao
    banco para compensação.}

\item{\code{nord}}{Número de ordens de pagamento, certificados e
    recibos de vendas emitidos pelos funcionários.}

\item{\code{ncor}}{Número de correspondências processadas e enviadas
    aos clientes.}

\item{\code{nout}}{Número de documentos processados e outras
    atividades.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 11, exercício
    2, pág. 273)
}
\description{
Estudo para determinar as atividades mais importantes
    dos funcionários do departamento de contabilidade em uma empresa,
    durante 30 dias.
}
\examples{

data(CharnetEx11.2)

plot(CharnetEx11.2)

}
\keyword{RM}

