% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx2.10.R
\name{CharnetEx2.10}
\alias{CharnetEx2.10}
\title{Estudo do Efeito do Carbono na Resistência Elétrica}
\format{Um \code{data.frame} com 2 colunas e 15 linhas.

\describe{

\item{\code{carb}}{Porcentagem de carbono nos fios de aço.}

\item{\code{res}}{Resistência elétrica, mensurada em \eqn{\mu} ohms
    cm a 20ºC.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Capítulo 2, exercício 10,
    página 47)
}
\description{
Estudo sobre o efeito do carbono contido, em fios de aço
    em resistência elétrica, na temperatura de 20ºC.
}
\examples{

data(CharnetEx2.10)

plot(CharnetEx2.10)
abline(lm(res ~ carb, data = CharnetEx2.10))

}
\keyword{TODO}

