% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx2.13}
\alias{CharnetEx2.13}
\title{Compara\enc{çã}{ca}o entre Alturas de Pais e Filhos}
\format{Um \code{data.frame} com 2 colunas e 7 linhas, em que

\describe{

\item{\code{pai}}{Altura do pai, em centímetros.}

\item{\code{filhos}}{Altura do filho, em centímetros.}

}}
\source{
CHARNET et al. (2008), Capítulo 2, exercício
    13, pág. 48.
}
\description{
Comparação entre alturas de 12 pais e respectivos 12
   filhos.
}
\examples{

data(CharnetEx2.13)
str(CharnetEx2.13)

plot(CharnetEx2.13)

}
\keyword{RS}

