% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx2.14.R
\name{CharnetEx2.14}
\alias{CharnetEx2.14}
\title{Produto Interno Bruto Trimestral do Brasil}
\format{Uma série temporal \code{ts}, com 32 observações, sendo 4
    observações por ano (trimestralmente) de 1990 a 1997.}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Capítulo 2, exercício 14,
    página 49)
}
\description{
Comparação do Produto Interno Bruto (PIB) trimestral do 
    Brasil em 8 anos, de 1990 a 1997.
}
\examples{

data(CharnetEx2.14)

CharnetEx2.14

plot(CharnetEx2.14, type = "p")

}
\keyword{TS}

