% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx2.15}
\alias{CharnetEx2.15}
\title{Notas M\enc{é}{e}dias de Candidatos ao Vestibular}
\format{Um \code{data.frame} com 9 colunas e 46 linhas.

\describe{

\item{\code{curso}}{Cursos ofertados no Vestibular.}

\item{\code{per}}{Período ofertado (Diurno / Noturno)}

\item{\code{cv}}{Relação Candidato/Vaga do curso.}

\item{\code{nq}}{Nota média nas provas de Química.}

\item{\code{qp}}{Disciplina Prioritária no curso (Química).}

\item{\code{nm}}{Nota média nas provas de Matemática.}

\item{\code{mp}}{Disciplina Prioritária no curso (Matemática).}

\item{\code{nh}}{Nota média nas provas de História.}

\item{\code{hp}}{Disciplina Prioritária no curso (História).}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 2, exercício
    15, pág. 50, Capítulo 7, exercício 3, pág. 164, Capítulo 8,
    exercício 6, pág. 197, Capítulo 9, exercício 8, pág. 231,
    Capítulo 10, exercício 3, pág. 254)
}
\description{
Notas médias de candidatos ao vestibular da Unicamp em
    1998.
}
\examples{

data(CharnetEx2.15)

library(ggplot2)

# Considerando a notas médias na prova de química
qplot(cv, nq, data = CharnetEx2.15, color = qp,
      xlab = "Relação Candidatos Vaga",
      ylab = "Notas na Prova de Química",
      main = "Notas na Prova de Química x Candidatos Vaga")
    geom_smooth(method = "lm", se = FALSE)

# Considerando a notas médias geral (soma das notas de química,
# matemática e história)
soma <- with(CharnetEx2.15, nq + nm + nh)
qplot(cv, soma, data = CharnetEx2.15, color = qp,
      xlab = "Relação Candidatos Vaga",
      ylab = "Soma das Notas")
    geom_smooth(method = "lm", se = FALSE)

# Correlação entre as variáveis numéricas (notas e relação
# candidato/vaga)
numcols <- sapply(CharnetEx2.15, is.numeric)
plot(CharnetEx2.15[, numcols])

}
\keyword{RM}

