% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx3.1}
\alias{CharnetEx3.1}
\title{Compara\enc{çã}{ca}o entre Notas}
\format{Um \code{data.frame} com 2 colunas e 50 linhas, em que

\describe{

\item{\code{teo}}{Notas da prova teórica.}

\item{\code{lab}}{Notas da prova de laboratório.}

}}
\source{
CHARNET et al. (2008), Capítulo 3, exercício 1, pág. 65.
}
\description{
Comparação entre as notas de uma prova teórica e de
    laboratório de 50 alunos, numa escala de 0 a 100 pontos.
}
\examples{

data(CharnetEx3.1)
str(CharnetEx3.1)

xtabs(~teo, data = CharnetEx3.1)

plot(CharnetEx3.1)

}
\keyword{RS}
