% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx3.9.R
\name{CharnetEx3.9}
\alias{CharnetEx3.9}
\title{Consumo de Combust\enc{í}{i}vel por Velocidade}
\format{Um \code{data.frame} com 2 colunas e 28 linhas.

\describe{

\item{\code{velo}}{Velocidade, em milhas por hora}

\item{\code{cons}}{Consumo de combustível, em milhas por galão}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 3, exercício
    9, pág. 68)
}
\description{
Teste feito com um certo tipo de carro, comparando o
    consumo de combustível (em milhas por galão) e sua velocidade
    (em milhas por hora).
}
\examples{

data(CharnetEx3.9)

plot(CharnetEx3.9)

}
\keyword{RS}

