% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx4.10}
\alias{CharnetEx4.10}
\title{Custo de Produ\enc{çã}{ca}o por Tamanho do Lote}
\format{Um \code{data.frame} com 2 colunas e 10 linhas.

\describe{

\item{\code{npecas}}{Quantidade de peças produzidas no lote, em
unidades.}

\item{\code{custo}}{Custo do total de peças do lote.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 4, exercício
    10, pág. 85, Capítulo 6, exercício 10, pág. 146)
}
\description{
Custo de produção pelo número de peças produzidas em uma
    amostra de 10 lotes.
}
\examples{

data(CharnetEx4.10)

plot(CharnetEx4.10)

}
\keyword{RS}

