% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx4.2}
\alias{CharnetEx4.2}
\title{Sal\enc{á}{a}rio M\enc{é}{e}dio Mensal de Ex-Alunos de
    Economia}
\format{Um \code{data.frame} com 2 colunas e 7 linhas, em que

\describe{

\item{\code{ano}}{Ano da coleta dos dados.}

\item{\code{ms}}{Valor da média salarial.}

}}
\source{
CHARNET et al. (2008), Capítulo 4, exercício 2, pág. 82.
}
\description{
Pesquisa de uma faculdade de economia sobre a evolução
    do salário médio mensal de 20 de seus ex-alunos no período de
    1987 a 1993.
}
\examples{

data(CharnetEx4.2)
str(CharnetEx4.2)

plot(CharnetEx4.2)

}
\keyword{RS}

