% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx4.8.R
\name{CharnetEx4.8}
\alias{CharnetEx4.8}
\title{Taxa de Acidentes de Trabalho}
\format{Um \code{data.frame} com 2 colunas e 9 linhas.

\describe{

\item{\code{ano}}{Ano em que foi registrado a taxa.}

\item{\code{taxa}}{Taxa de acidentes de trabalho, em milhão de
    horas/homem.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 4, exer. 8, p. 84,
    cap. 6, exer. 9, p. 146.)
}
\description{
Taxa de acidentes de trabalho por milhão de horas/homem
    de exposição ao risco, durante 9 anos.
}
\examples{

data(CharnetEx4.8)

plot(CharnetEx4.8)

}
\keyword{RS}

