% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx5.1}
\alias{CharnetEx5.1}
\title{Peso de Embri\enc{õ}{o}es de Galinha no Tempo}
\format{Um \code{data.frame} com 2 colunas e 11 linhas.

\describe{

\item{\code{idade}}{Idade do embrião, em dias.}

\item{\code{peso}}{Peso do embrião, em gramas.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 5, exercício
    1, pág. 108)
}
\description{
Peso de embriões de galinha no tempo.
}
\examples{

data(CharnetEx5.1)

plot(CharnetEx5.1)

}
\keyword{RS}

