% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx5.3.R
\name{CharnetEx5.3}
\alias{CharnetEx5.3}
\title{Precisão de um Velocímetro}
\format{Um \code{data.frame} com 2 colunas e 8 linhas.

\describe{

\item{\code{vr}}{Velocidade real, em km/h.}

\item{\code{va}}{Velocidade anotada no velocímetro, em km/h.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 5, exercício
    3, pág. 109)
}
\description{
Medição da precisão de um velocímetro de locomotiva.
}
\examples{

data(CharnetEx5.3)

plot(CharnetEx5.3)
abline(a = 0, b = 1)

}
\keyword{RS}

