% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx5.6.R
\name{CharnetEx5.6}
\alias{CharnetEx5.6}
\title{Consumo de Combustível e Peso do Automóvel}
\format{Um \code{data.frame} com 2 colunas e 14 linhas.

\describe{

\item{\code{peso}}{Peso do veículo, em kg.}

\item{\code{cons}}{Consumo do veículo num determinado trecho,
   em litros.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 5, exercício
    6, pág. 110)
}
\description{
Consumo de combustível para percorrer determinado trecho
    conforme o peso dos veículos de passeio, de mesmo ano,
    selecionados aleatoriamente de uma grande empresa.
}
\examples{

data(CharnetEx5.6)

plot(CharnetEx5.6)

}
\keyword{RS}

