% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx6.3.R
\name{CharnetEx6.3}
\alias{CharnetEx6.3}
\title{Salário Mensal de Formandos em Economia}
\format{Um \code{data.frame} com 2 colunas e 35 linhas.

\describe{

\item{\code{ano}}{Ano do registro do salário.}

\item{\code{salario}}{Salário mensal.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R. & Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed., p. 356). SP: Editora Unicamp (Cap. 6, exer. 3, p. 143)
}
\description{
Pesquisa de uma faculdade de economia sobre a evolução
    do salário mensal de 5 de seus formandos no período de 87 a 93.
}
\examples{

data(CharnetEx6.3)

plot(CharnetEx6.3)

}
\keyword{RS}

