% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharnetEx6.7.R
\name{CharnetEx6.7}
\alias{CharnetEx6.7}
\title{Peso de Correspondências}
\format{Um \code{data.frame} com 2 colunas e 11 linhas.

\describe{

\item{\code{peso}}{Peso total das correspondências no dia, em kg.}

\item{\code{cor}}{Total de correspondências no dia, em milhares.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 6, exercício
    7, pág. 145)
}
\description{
Relação do peso das correspondências levantados por uma
    agência de correios durante 11 dias.
}
\examples{

data(CharnetEx6.7)

plot(CharnetEx6.7)

}
\keyword{RS}

