% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx7.2}
\alias{CharnetEx7.2}
\title{Impacto de Impurezas em um Reator Qu\enc{í}{i}mico}
\format{Um \code{data.frame} com 4 colunas e 20 linhas, em que

\describe{

\item{\code{rea}}{Reator analisado, um fator com dois níveis \code{I}
    ou \code{II}.}

\item{\code{t}}{Tempo, em minutos.}

\item{\code{imp}}{Porcentagem de impurezas, em escala \eqn{\log}.}

}}
\source{
CHARNET et al. (2008), Capítulo 7, exercício 2, pág. 163;
    Capítulo 10, exercício 2, pág. 253.
}
\description{
Relação entre a porcentagem de impurezas dentro de um
    reator químico e o tempo de permanência da substância dentro
    deste reator.
}
\examples{

data(CharnetEx7.2)
str(CharnetEx7.2)

with(CharnetEx7.2, plot(tempo ~ imp, col = rea))

}
\keyword{RS}
\keyword{dummy}
