% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx7.7}
\alias{CharnetEx7.7}
\title{Rela\enc{çã}{ca}o entre Consumo de Combust\enc{í}{i}vel e
    Pot\enc{ê}{e}ncia do Motor}
\format{Um \code{data.frame} com 4 colunas e 20 linhas, em que

\describe{

\item{\code{marca}}{Marca do carro, fator com dois níveis \code{A} ou
    \code{B}.}

\item{\code{cons}}{Consumo de combustível, em km/l.}

\item{\code{pot}}{Potência do motor, em HP.}

}}
\source{
CHARNET et al. (2008), Capítulo 7, exercício 7, pág. 167,
    Capítulo 10, exercício 6, pág. 256.
}
\description{
Relação entre o consumo de combustível (km/l) e a
    potência do motor (HP) de duas marcas de carro (A e B).
}
\examples{

data(CharnetEx7.7)
str(CharnetEx7.7)

with(CharnetEx7.7, plot(cons, pot, col = marca))

}
\keyword{RS}
\keyword{dummy}
