% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx8.2}
\alias{CharnetEx8.2}
\title{Rela\enc{çã}{ca}o da Temperatura e Produ\enc{çã}{ca}o de um
    Produto Qu\enc{í}{i}mico}
\format{Um \code{data.frame} com 2 colunas e 11 linhas, em que

\describe{

\item{\code{temp}}{Temperatura codificada.}

\item{\code{peso}}{Peso do produto, em kg.}

}}
\source{
CHARNET et al. (2008), Capítulo 8, exercício 2, pág. 195,
    Capítulo 9, exercício 11, pág. 233.
}
\description{
Uma indústria química está interessada em maximizar a
    sua produção de um certo processo químico. Para isso, obteve os
    seguintes resultados da temperatura codificada, \code{temp}, e do
    peso do produto, \code{peso}, em kg.
}
\examples{

data(CharnetEx8.2)
str(CharnetEx8.2)

plot(CharnetEx8.2)

}
\keyword{RNL}
\keyword{RS}

