% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx8.3}
\alias{CharnetEx8.3}
\title{Tempo de Corros\enc{ã}{a}o do Metal}
\format{Um \code{data.frame} com 2 colunas e 10 linhas.

\describe{

\item{\code{tempo}}{Tempo de exposição do metal, em semanas.}

\item{\code{gc}}{Grau de corrosão do metal.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 8, exercício
    3, pág. 196, Capítulo 9, exercício 9, pág. 232)
}
\description{
Estudo da relação entre o grau de corrosão de um certo
   metal e o tempo de exposição deste metal à ação da acidez do solo.
}
\examples{

data(CharnetEx8.3)

plot(CharnetEx8.3)

}
\keyword{RP}

