% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charnet.R
\name{CharnetEx8.5}
\alias{CharnetEx8.5}
\title{Efeito da Temperatura e Concentra\enc{çã}{ca}o numa Rea\enc{çã}{ca}o Qu\enc{í}{i}mica}
\format{Um \code{data.frame} com 3 colunas e 20 linhas.

\describe{

\item{\code{prod}}{Produção de um certo composto, em litros.}

\item{\code{temp}}{Temperatura da reação, em ºC.}

\item{\code{conc}}{Percentual de concentração, em \%.}

}}
\source{
Charnet, R., de Luna Freire, C.A., Charnet, E.M.R., Bonvino,
    H. (2008). Análise de modelos de regressão linear com aplicações
    (2nd ed.). Campinas, SP: Editora Unicamp (Capítulo 8, exercício
    5, pág. 197.  Capítulo 9, exercício 10, pág. 233)
}
\description{
Efeito da temperatura (\code{temp}) e percentual de
    concentração (\code{conc}) na produção de um composto numa reação
    química (\code{prod}).
}
\examples{

data(CharnetEx8.5)

plot(CharnetEx8.5)

}
\keyword{RM}

