% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioEx1.4.2}
\alias{DemetrioEx1.4.2}
\title{Tempo de Irriga\enc{çã}{ca}o de Solo}
\format{Um \code{data.frame} de 15 linhas e 2 colunas, em que

\describe{

\item{\code{tempo}}{Tempo acumulado de irrigação, medido em
    minutos.}

\item{\code{infil}}{Infiltração acumulada de água no solo, medida
    em centímetros (cm).}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Exercício 1.4.2 pág. 16.
}
\description{
Neste estudo foram medidos os tempos acumulados de
    irrigação e as correspondentes medidas de infiltração acumulada
    de água no solo. O objetivo do experimento era estimar as
    equações de infiltração acumulada em relação ao tempo acumulado e
    de velocidade de infiltração em relação ao tempo acumulado e à
    velocidade básica de infiltração.  Essas equações são importantes
    para determinar o tempo de irrigação para atingir uma determinada
    lâmina de água, pois é anti-econômico irrigar a uma velocidade
    maior à de infiltração.
}
\examples{

data(DemetrioEx1.4.2)
str(DemetrioEx1.4.2)

library(lattice)
library(latticeExtra)

xyplot(infil ~ tempo, data = DemetrioEx1.4.2,
       xlab = "Tempo",
       ylab = "Infiltração",
       type = c("p", "r"), col.line = 3)

xyplot(infil ~ tempo, data = DemetrioEx1.4.2,
       xlab = "Tempo",
       ylab = "Infiltração",
       type = c("p", "r"), col.line = 3,
       scales = list(y = list(log = 10), x = list(log = 10)),
       yscale.components = yscale.components.log10ticks,
       xscale.components = xscale.components.log10ticks)

}
\keyword{RS}
