% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioEx2.12.5.R
\name{DemetrioEx2.12.5}
\alias{DemetrioEx2.12.5}
\title{Estimativa de Quadrados Mínimos}
\format{Um \code{data.frame} de 5 linhas e 2 colunas.

\describe{

    \item{\code{x}}{Sem interpretação.}
    
    \item{\code{y}}{Sem interpretação.}
    
}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Exercício 2.12.5 pág. 60)
}
\description{
Os dados estão presentes no exercício com o objetivo de
    estimar na prática o quadrado mínimo.
}
\examples{

data(DemetrioEx2.12.5)

library(lattice)

xyplot(y ~ x, data = DemetrioEx2.12.5,
     main = "x vs y",
     xlab = "x",
     ylab = "y",
     type = c("p", "r"), col.line = 3)
     
}
\keyword{TODO}

