% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioEx5.4.5}
\alias{DemetrioEx5.4.5}
\title{Estudo sobre a Avalia\enc{çã}{ca}o Visual do Grau da
    Infesta\enc{çã}{ca}o de Plantas por Doen\enc{ç}{c}as}
\format{Um \code{data.frame} de 10 linhas e 2 colunas, em que

\describe{

\item{\code{visual}}{Estimativas visuais do pesquisador.}

\item{\code{real}}{Valores reais do grau de infestação.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Exercício 5.4.5 pág. 165.
}
\description{
Dados referentes a avaliação visual (realizada por um
    pesquisador) e real do grau de infestação de folhas de amendoim.

    A avaliação exige um treinamento específico para a cultura e a
    doença em questão. Por esse motivo, foram desenvolvidos programas
    computacionais que geram imagens de folhas com diferentes
    porcentagens de infestação para o pesquisador estimar visualmente
    e em seguida compara-se com as porcentagens reais.
}
\examples{

data(DemetrioEx5.4.5)
str(DemetrioEx5.4.5)

library(lattice)

xyplot(visual ~ real, data = DemetrioEx5.4.5,
       xlab = "Real",
       ylab = "Visual",
       type = c("p", "r"), col.line = 3)

}
\keyword{RS}
