% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioEx5.4.5.R
\name{DemetrioEx5.4.5}
\alias{DemetrioEx5.4.5}
\title{Avaliação visual do grau de infestação}
\format{Um \code{data.frame} de 10 linhas e 2 colunas.

\describe{

    \item{\code{visual}}{Estimativas visuais do pesquisador.}
    
    \item{\code{real}}{Valores reais do grau de infestação.}
    
}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Exercício 5.4.5 pág. 165)
}
\description{
Dados referentes a avaliação visual (realizada por um
    pesquisador) e real do grau de infestação de folhas de amendoim.
    
    A avaliação exige um treinamento específico para a cultura e
    a doença em questão. Por esse motivo, foram desenvolvidos
    programas computacionais que geram imagens de folhas com 
    diferentes porcentagens de infestação para o pesquisador estimar
    visualmente e em seguida mostram as porcentagens reais.
}
\examples{

data(DemetrioEx5.4.5)

library(lattice)

xyplot(visual ~ real, data = DemetrioEx5.4.5,
     main = "Real vs Visual",
     xlab = "Real",
     ylab = "Visual",
     type = c("p", "r"), col.line = 3)
     
}
\keyword{RS}

