% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioTb1.3.R
\name{DemetrioTb1.3}
\alias{DemetrioTb1.3}
\title{Estudo da Construção de um Tensiômetro}
\format{Um \code{data.frame} com 9 linhas e 2 colunas.

\describe{

    \item{\code{altura}}{ Altura da câmara no tensiômetro medida
    em milímetros (mm). }
    
    \item{\code{tensao}}{ Tensão da água no solo medida em coloumb
    (mb) }

}}
\source{
Demétrio, C. G. B., & Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 1.3 pág. 10; 
    Ex 2.12.14 pág. 62)
}
\description{
Estudo da construção de um tensiômetro de leitura
    direta. Neste estudo obteve-se os resultados de alturas da câmara
    no tensiômetro e tensão da água no solo.
}
\examples{

library(lattice)

data(DemetrioTb1.3)

xyplot(tensao ~ altura, data = DemetrioTb1.3,
    main = "Altura VS Tensão",
    xlab = "Altura do Tensiômetro",
    ylab = "Tensão da Água")

xyplot(tensao ~ log(altura), data = DemetrioTb1.3,
    main = "Altura VS Tensão",
    xlab = expression(log~"(Altura do Tensiômetro)"),
    ylab = "Tensão da Água")

}
\keyword{RS}

