% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemetrioTb2.10.R
\name{DemetrioTb2.10}
\alias{DemetrioTb2.10}
\title{Absor\enc{çã}{ca}o de CO2 por Folhas de Trigo}
\format{Um \code{data.frame} de 17 linhas e 2 colunas.

\describe{

    \item{\code{co2}}{Concentração de \eqn{CO_2} aplicada sobre as
    folhas de trigo.}
    
    \item{\code{absorv}}{Quantia de \eqn{CO_2} absorvida pelas folhas
    de trigo, medida em \eqn{cm^3/dm^2/hora.}}
    
}}
\source{
Demétrio, C. G. B., Zocchi, S. S. (2011). Modelos de
    Regressão. Piracicaba: ESALQ. (Tabela 2.10 pág. 65, Exercício
    1.4.1.3 pág. 14)
}
\description{
Dados provenientes de um ensaio inteiramente casualizado
    onde aplicou-se \eqn{CO_2} sobre folhas de trigo a uma
    temperatura de 35°C e mediu-se a quantia de \eqn{CO_2} absorvido
    pelas folhas.
}
\examples{

data(DemetrioTb2.10)

library(lattice)

xyplot(absorv ~ co2, data = DemetrioTb2.10,
       main = "CO2 Aplicado VS Absorvido",
       xlab = "Aplicado",
       ylab = "Absorvido",
       type = c("p", "r"), col.line = 3)

# Subconjunto do exercício 1.4.1.3
obs <- c(1, 2, 3, 5, 6, 7, 8, 11, 12, 16, 17)
DemetrioEx1.4.1.3 <- DemetrioTb2.10[obs, ]

xyplot(absorv ~ co2, data = DemetrioEx1.4.1.3,
       type = c("p", "r"), col.line = 3)

}
\keyword{TODO}

