% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demetrio.R
\name{DemetrioTb4.5}
\alias{DemetrioTb4.5}
\title{Dados Simulados para Regress\enc{ã}{a}o Linear Simples e
    Polinomial}
\format{Um \code{data.frame} de 9 linhas e 5 colunas, em que

\describe{

\item{\code{x}}{Variável explicativa, sem interpretação.}

\item{\code{y1}}{Variável dependente, sem interpretação.}

\item{\code{y2}}{Variável dependente, sem interpretação.}

\item{\code{y3}}{Variável dependente, sem interpretação.}

\item{\code{y4}}{Variável dependente, sem interpretação.}

}}
\source{
DEMÉTRIO; ZOCCHI (2011), Tabela 4.5 pág. 137.
}
\description{
Dados simulados para exercício de ajuste de modelos de
    regressão linear simples e polinomial. São simulados, para uma
    mesma covariável (x), quatro diferentes variáveis de interesse
    (y).
}
\examples{

data(DemetrioTb4.5)
str(DemetrioTb4.5)

# Relação da covariável com cada uma das variáveis resposta
par(mfrow = c(1, ncol(DemetrioTb4.5) - 1))
for (i in 2:ncol(DemetrioTb4.5)) {
    plot(DemetrioTb4.5[, c(1, i)])
    lines(lowess(DemetrioTb4.5[, c(1, i)]), col = 2)
}

# Gráfico de dispersão por pares, apresenta também a relação entre as
# variáveis resposta
pairs(~ x + y1 + y2 + y3 + y4 , data = DemetrioTb4.5,
     main = "Dispersão duas a duas")

}
\keyword{RS}

