% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiasEg3.2.R
\name{DiasEg3.2}
\alias{DiasEg3.2}
\title{Produ\enc{çã}{ca}o de frutos em cacaueiro}
\format{Vetor com 43 observações que são o número de frutos
    produzidos por plantas de cacaueiro.}
\source{
Dias, L. A. S., Barros, W. S. (2009). Biometria
    Experimental. Viçosa, MG: UFV.
}
\description{
Amostra aleatória do número de frutos produzidos por
    plantas cacaueiro.
}
\examples{

stem(DiasEg3.2)

hist(DiasEg3.2)
rug(DiasEg3.2)

plot(density(DiasEg3.2))
rug(DiasEg3.2)

boxplot(DiasEg3.2)
rug(DiasEg3.2, side = 2)

}
\keyword{AAS}

