% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epprecht.R
\name{EpprechtTb6.12}
\alias{EpprechtTb6.12}
\title{Peso de um Produto}
\format{Um \code{data.frame} com 30 observações e 1 variável, em que

\describe{

\item{\code{peso}}{Peso do produto.}

}}
\source{
COSTA et al. (2010), pág. 184.
}
\description{
Trinta observações registradas referentes aos pesos de
    um produto.
}
\examples{

data(EpprechtTb6.12)
str(EpprechtTb6.12)

library(qcc)

qcc(EpprechtTb6.12$peso, type = "xbar.one", nsigmas = 3,
    xlab = "", ylab = "Peso", title = "")


}
\keyword{CEQ}

