% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epprecht.R
\name{EpprechtTb6.4}
\alias{EpprechtTb6.4}
\title{Temperatura do Banho Qu\enc{í}{i}mico}
\format{Um \code{data.frame} com 60 observações e 2 variáveis, 
    em que

\describe{

\item{\code{amostra}}{Identificação da amostra.}
 
\item{\code{temp}}{Temperaturas registradas.}
 
}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estatístico de Qualidade (2nd ed.). São Paulo - SP:
     Editora Atlas. (pg 167).
}
\description{
A cada 30 minutos registram-se três temperaturas do 
    banho, espaçadas de 3 minutos. Exemplo: na primeira amostra 
    efetua-se uma medida às 8h00, outra às 8h03 e outra às 8h06. Na 
    segunda amostra são registradas as temperaturas nos horários 
    8h30, 8h33 e 8h36, e assim por diante.
}
\examples{

data(EpprechtTb6.4)

library(qcc)

obj <- qcc.groups(EpprechtTb6.4$temp, EpprechtTb6.4$amostra)
qcc(obj, type="xbar", nsigmas=3, 
   xlab="Amostra", ylab="Temperatura", title=" ")
                  
}
\keyword{CEQ}

