% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epprecht.R
\name{EpprechtTb8.12}
\alias{EpprechtTb8.12}
\title{Controle de qualidade para a fra\enc{çã}{ca}o de n\enc{ã}{a}o conformes.}
\format{Um \code{data.frame} com 20 observações e 1 variável, 
    em que

\describe{ 

\item{\code{nconf}}{Número de peças não conformes nas amostras.}

}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estatístico de Qualidade (2nd ed.). São Paulo - SP:
     Editora Atlas. (pg 227).
}
\description{
Dados de 20 amostras, representando o número 
    de peças não-conformes em amostras de tamanho 100.
}
\examples{

data(EpprechtTb8.12) 

library(qcc)

qcc(EpprechtTb8.12, type="p", size=100,
    xlab="Amostra", ylab="Proporção de itens não conformes", title=" ")
                  
}
\keyword{CEQ}

