% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpprechtTb8.2.R
\name{EpprechtTb8.2}
\alias{EpprechtTb8.2}
\title{Número de Clientes Insatisfeitos com a Comida}
\format{Um \code{data.frame} com 30 observações e 1 variável, 
    em que

\describe{

\item{\code{cinsat}}{Número de clientes insatisfeitos.}

}}
\source{
Costa, A. F. B., Epprecht, E. K., Carpinetti, L. (2010).
     Controle Estatístico de Qualidade (2nd ed.). São Paulo - SP:
     Editora Atlas. (pg 204).
}
\description{
Dados referentes ao número de clientes insatisfeitos
    com a comida de um restaurante. Esses dados foram obtidos após o
    diagnóstico e a eliminação de causas especiais.
}
\examples{

data(EpprechtTb8.2) 

library(qcc)

qcc(EpprechtTb8.2, type="p", size=30,
    xblab=" ", ylab="Observações", title=" ")
                  
}
\keyword{CEQ}

