% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Faria.R
\name{FariaEg3.2.4}
\alias{FariaEg3.2.4}
\title{Produção de Amêndoas de Clones de Cacau}
\format{Um \code{data.frame} com 24 observações e 3 variáveis, em que

\describe{

\item{\code{clone}}{Fator categórico que representam os clones de
    cacau.}

\item{\code{rept}}{Inteiro que indica a repetição de cada clone.}

\item{\code{prod}}{Produção de amendoas, em kg 10 plantas\eqn{{-1}}
    ano\eqn{{-1}}.}

}}
\source{
FARIA (2009), Tabela 3.2.4, pág 46.
}
\description{
Experimento em delineamento inteiramente casualizado que
    estudou a produção de amendoas de clones de cacau.
}
\examples{

library(lattice)

data(FariaEg3.2.4)
str(FariaEg3.2.4)

unstack(x = FariaEg3.2.4, form = prod ~ clone)

xyplot(prod ~ clone, data = FariaEg3.2.4,
       ylab = expression("Produção"~
                             (kg~10~plantas^{-1}~ano^{-1})),
       xlab = "Clones de cacau")

}
\keyword{DIC}

