% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Faria.R
\name{FariaQd11.9}
\alias{FariaQd11.9}
\title{Qualidade de Mudas em Fun\enc{çã}{ca}o do Recipiente da
    Esp\enc{é}{e}cie}
\format{Um \code{data.frame} com 12 observações e 4 variáveis, em que

\describe{

\item{\code{recip}}{Fator categórico cujos níveis representam os
    recipientes utilizados para o cultivo das mudas.}

\item{\code{espécie}}{Fator categórico cujos níveis representam as
    espécies cultivadas.}

\item{\code{rept}}{Inteiro que indica as repetições das combinações
    dos fatores.}

\item{\code{quali}}{Resposta correspondente à qualidade das mudas.}

}}
\source{
FARIA (2009), Quadro 11.9, pág 145.
}
\description{
Experimento montado no delineamento inteiramente
    casualizado cujo resultado é a qualidade de mudas em função do
    recipiente utilizado e para cada espécie cultiavda.
}
\examples{

library(lattice)

data(FariaQd11.9)
str(FariaQd11.9)

xtabs(~recip + espec, data = FariaQd11.9)
with(FariaQd11.9, tapply(quali, list(recip, espec), FUN = sum))

xyplot(quali ~ recip, groups = espec, data = FariaQd11.9,
       type = c("p", "a"),
       xlab = "Recipientes", ylab = "Qualidade da muda",
       auto.key = list(title = "Espécie", cex.title = 1.1,
                       columns = 2))

}
\keyword{DIC}
\keyword{FAT2}

