% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Faria.R
\name{FariaQd6.1}
\alias{FariaQd6.1}
\title{Produ\enc{çã}{ca}o de Variedades de Milho}
\format{Um \code{data.frame} com 20 observações e 3 variáveis, em que

\describe{

\item{\code{varied}}{Fator categórico que representa as variedades de
    milho.}

\item{\code{rept}}{Inteiro que identifica as repetições de cada
    variedade.}

\item{\code{prod}}{Produção de milho, em kg por 100 m\eqn{^2}.}

}}
\source{
FARIA (2009), Tabela 6.1 pág 76.
}
\description{
Produção de milho, em kg 100 m\eqn{^2} em um experimento
    conduzido no delineamento inteiramente casualizado com 5
    repetições, onde foram avaliadas quatro variedades de milho.
}
\examples{

library(lattice)

data(FariaQd6.1)
str(FariaQd6.1)

xyplot(prod ~ varied, data = FariaQd6.1)

}
\keyword{DIC}
