% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FerreiraEg5.1.R
\name{FerreiraEg5.1}
\alias{FerreiraEg5.1}
\title{Teores de Areia e Argila em Solo na Amaz\enc{ô}{o}nia}
\format{Um \code{data.frame} com 30 observações e 2 variáveis, em que

\describe{

\item{\code{areia}}{Representa o teor de areia em cada parcela de
    solo.}

\item{\code{argila}}{Representa o teor de argila em cada parcela de
    solo.}

}}
\source{
Ferreira, D. F. (2011). Estatística Multivariada (2nd
    ed.). Lavras, MG: Editora UFLA. (Exemplo 5.1 pág. 193-194)
}
\description{
Conjunto de dados referente aos teores de areia e argila
    de uma amostra com 30 parcelas de solo de capoeira nova na
    Amazônia.
}
\examples{

data(FerreiraEg5.1)

str(FerreiraEg5.1)

pairs(FerreiraEg5.1)

bks <- apply(FerreiraEg5.1, 2,
             FUN = function(x, step = 5) {
                 r <- range(x)
                 r <- r + (r \%\% step * (-1)) + c(0, step)
                 seq(r[1], r[2], by = step)
             })
areiaPlot <- hist(FerreiraEg5.1$areia, plot = FALSE,
                  breaks = bks$areia)
argilaPlot <- hist(FerreiraEg5.1$argila, plot = FALSE,
                   breaks = bks$argila)

layout(matrix(c(2, 0, 1, 3), ncol = 2, byrow = TRUE),
       widths = c(4/5, 1/5), heights = c(1/5, 4/5))
par(mar = c(5, 5, 0.5, 0.5))
plot(argila ~ areia, data = FerreiraEg5.1, cex = 1.5,
     xlab = "Teor de areia do solo",
     ylab = "Teor de argila do solo",
     xlim = range(bks$areia),
     ylim = range(bks$argila))
grid()
rug(x = FerreiraEg5.1$areia, side = 3)
rug(x = FerreiraEg5.1$argila, side = 4)
par(mar = c(0, 5, 1, 0.5))
barplot(areiaPlot$counts, axes = FALSE, space = 0)
par(mar = c(5, 0, 0.5, 1))
barplot(argilaPlot$counts, axes = FALSE, space = 0, horiz = TRUE)

}
\keyword{TODO}

