% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ferreira.R
\name{FerreiraEx3.8.5}
\alias{FerreiraEx3.8.5}
\title{Avalia\enc{çã}{ca}o de Cultivar de Mel\enc{ã}{a}o}
\format{\code{data.frame} com 32 observações e 3 variáveis, em que

\describe{

\item{\code{nfrut}}{Número de frutos total por héctare.}

\item{\code{prod}}{Produção de melão mensurada em kg ha\eqn{^{-1}}.}

\item{\code{tam}}{Teor médio de açucar mensurados em graus brix.}

}}
\source{
FERREIRA (2011), Exercício 3.8.5, pág. 169.
}
\description{
Os dados referem-se à avaliação de uma cultivar de melão
    em 32 unidades experimentais.
}
\examples{

data(FerreiraEx3.8.5)
str(FerreiraEx3.8.5)

# Análise descritiva
panel.density <- function(x, ...) {
    usr <- par("usr")
    on.exit(par(usr))
    par(usr = c(usr[1:2], 0, 1.5))
    par(new = TRUE)
    plot(density(x), xlab = "", ylab = "", main = "", lwd = 2)
}

pairs(FerreiraEx3.8.5, pch = 20, diag.panel = panel.density)

# Análise de componentes principais (PCA)
(comp <- prcomp(FerreiraEx3.8.5, scale = TRUE))
screeplot(comp, type = "lines")
biplot(comp)

}
\keyword{AnaComPrin}
