% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labestData.R
\docType{package}
\name{labestData}
\alias{labestData}
\alias{labestData-package}
\title{Biblioteca de Dados para Ensino de Estatística}
\description{
O labestData é um projeto coletivo do PET Estatística
    UFPR para desenvolver um pacote R com conjuntos de dados para
    ensino de Estatística. O desenvolvimento do pacote visa
    contribuir com a comunidade científica ao reunir, organizar,
    manter e disponibilizar conjuntos de dados disponíveis em livros,
    páginas de internet e arquivos pessoais para o ensino-aprendizado
    de Estatística.  Nesta versão o pacote tem 451 conjuntos de
    dados.
}

