% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labestDataView.R
\name{labestDataView}
\alias{labestDataView}
\title{Shiny App para Visualiza\enc{çã}{ca}o e Download dos Datasets}
\usage{
labestDataView()
}
\value{
Abre uma aplicação web com shiny em seu navegador padrão.
}
\description{
Essa função abre uma interface \pkg{shiny} em seu
    navegador padrão para visualizar as tabelas de dados, consultar
    sua respectiva documentação e salvá-las, em txt, se preciso.
}
\section{Warning}{
 Para funcionar, é necessário ter o pacote shiny
    instalado. Sua visualização web depende dos recursos CSS3,
    Bootstrap, HTML5 e JavaScript5, portanto use navegadores com
    suporte para tais recursos (e.g Mozilla Firefox, Chromium Web
    Browser).
}
\examples{
\dontrun{

labestDataView()

}
}
\author{
Walmes Zeviani e Eduardo Ribeiro Jr.
}

