% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labestDataView.R
\name{labestDataView}
\alias{labestDataView}
\title{Shiny App para Visualiza\enc{çã}{ca}o e Download dos Datasets}
\usage{
labestDataView()
}
\value{
Abre uma aplicação web com shiny em seu navegador padrão.
}
\description{
Essa função abre uma interface \pkg{shiny} em seu
    navegador padrão para visualizar as tabelas de dados, consultar
    sua respectiva documentação e salvá-las, em txt, se preciso. A
    interfaze exibe uma tabela reativa a filtros para as obras e
    keywords que classificam as obras.
}
\section{Warning}{
 Para funcionar, é necessário ter o pacote shiny
    instalado. Sua visualização web depende dos recursos CSS3,
    Bootstrap, HTML5 e JavaScript5, portanto use navegadores com
    suporte para tais recursos (e.g Mozilla Firefox, Chromium Web
    Browser).
}

\examples{

# Objetos do pacote labesData.
ls("package:labestData")

# Dataframe com os pares keywords e datasets.
str(keywords)

# Cria uma coluna para a obra.
keywords$obra <-
    gsub(pattern = "^ *([A-Z]{1}[a-z]*)[A-Z]{1}[a-z]{1}.*$",
         replacement = "\\\\1",
         x = keywords$name)

# Mostra um pedaço do dataframe.
head(keywords)

library(lattice)

# Gráfico de Pareto das keywords.
barchart(sort(xtabs(~keyword, data = keywords)))

ob <- gsub(pattern = "^ *([A-Z]{1}[a-z]*)[A-Z]{1}[a-z]{1}.*$",
           replacement = "\\\\1",
           x = unique(keywords$name))

# Gráfico de pareto das obras.
barchart(sort(xtabs(~ob)))

\dontrun{

# Abre a interface no navegador.
labestDataView()

}
}
\author{
Walmes Zeviani e Eduardo Ribeiro Jr.
}
