% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb1.1}
\alias{ManlyTb1.1}
\title{Pardais sobreviventes da tempestade}
\format{Um \code{data.frame} com 49 registros e 6 variáveis.

\describe{

\item{\code{ct}}{Comprimento total do pardal (mm).}

\item{\code{ea}}{Extensão alar (mm).}

\item{\code{cbc}}{Comprimento do bico e cabeça (mm).}

\item{\code{cdu}}{Comprimento do úmero (mm).}

\item{\code{cqe}}{Comprimento da quilhado esterno (mm).}

\item{\code{sobrev}}{Se o animal sobreviveu: S = Sim e N = Não.}

}}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados:
    uma introdução. Porto Alegre, RS: Bookman (Tabela 1.1,  pág 14 e
    15)
}
\description{
Estudo em 1898, para a teoria da evolução de Darwin com
    pardais moribundos, com total de 49 dados.
}
\examples{

data(ManlyTb1.1)

pairs(~ct + ea + cbc + cdu + cqe + sobrev,
      data = ManlyTb1.1,
      main = "Gráfico de dispersão das variáveis nos pardais")

}
\keyword{TODO}

