% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb11.3}
\alias{ManlyTb11.3}
\title{Dist\enc{â}{a}ncias Rodovi\enc{á}{a}rias}
\format{Uma matriz (13x13) com as distâncias rodoviárias entre as
    cidades da ilha Sul da Nova Zelândia.}
\source{
MANLY (2005), pág. 182.
}
\description{
Distâncias rodoviárias (em milhas) entre cidades na ilha
    Sul da Nova Zelândia
}
\examples{

data(ManlyTb11.3)
str(ManlyTb11.3)

library(lattice)

levelplot(ManlyTb11.3,
          xlab = "",
          ylab = "",
          scales = list(x = list(rot = 90)))

}
\keyword{AnaAgrup}
