% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyTb11.5.R
\name{ManlyTb11.5}
\alias{ManlyTb11.5}
\title{Comportamento da votação de parlamentares}
\format{Um \code{matrix} com 15 parlamentares

\describe{

\item{\code{Colnames}}{Identificação dos parlamentares.}

\item{\code{Rownames}}{Identificação das parlamentares.}

}}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados - uma introdução.
     Porto Alegre, RS: Bookman (pg 185)
}
\description{
Número de votos discordantes entre os parlamentares de Nova Jersey em leis que concernem sobre problemas ambientais
}
\examples{

data(ManlyTb11.5)

levelplot(ManlyTb11.5, xlab = "", ylab = "", main = "Gráfico das distâncias",
 scales = list(x = list(rot=90)))
}
\keyword{EM}

