% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manly.R
\name{ManlyTb11.5}
\alias{ManlyTb11.5}
\title{Vota\enc{çõ}{co}es de parlamentares}
\format{Uma matriz com os números de votos discordantes entre 15
parlamentares de Nova Jersey.}
\source{
Manly, B. J. F. (2005). Métodos Estatísticos Multivariados
     - uma introdução. Porto Alegre, RS: Bookman (pg 185)
}
\description{
Número de votos discordantes entre os parlamentares de
Nova Jersey em leis referentes a problemas ambientais
}
\examples{

data(ManlyTb11.5)

require(lattice)

levelplot(ManlyTb11.5, xlab = "", ylab = "", main = "Gráfico das
distâncias entre parlamentares", scales = list(x = list(rot=90)))
}
\keyword{TODO}

