% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiAnA2}
\alias{MingotiAnA2}
\title{Desempenho de 44 Funcion\enc{á}{a}rios de uma Empresa}
\format{Um \code{data.frame} com 44 observações e 8 variáveis, em que

\describe{

\item{\code{funcio}}{Número de identificação do funcionário.}

\item{\code{venda}}{Nota atribuída ao funcionário referente ao
    desempenho nas vendas.}

\item{\code{lucro}}{Nota atribuída ao funcionário referente ao
    desempenho no lucro da empresa.}

\item{\code{clie}}{Nota atribuída ao funcionário referente ao
    desempenho na captação de novos clientes.}

\item{\code{escri}}{Nota atribuída à habilidade do funcionário na
    escrita.}

\item{\code{logica}}{Nota atribuída à habilidade do funcionário em
    lógica.}

\item{\code{social}}{Nota atribuída à habilidade social do
    funcionário.}

\item{\code{mate}}{Nota atribuída a habilidade do funcionário em
    matemática.}

}}
\source{
MINGOTI (2005), pág. 123, Ex. 4.2; pág. 126, Ex 4.3.
}
\description{
Dados referentes a 44 funcionários de uma empresa, aos
    quais foram atribuídas notas para desempenho nas vendas,
    desempenho nos lucros, captação de novos clientes, além dos
    resultados de quatro testes que medem a habilidade de escrita,
    lógica, social e matemática.
}
\examples{

data(MingotiAnA2)
str(MingotiAnA2)

pairs(MingotiAnA2)

}
\keyword{AnaFat}

