% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MingotiAnA5.R
\name{MingotiAnA5}
\alias{MingotiAnA5}
\title{Seleção de alunos para um programa de pós-graduação}
\format{Um \code{data.frame} com 62 observações e 4 variáveis, em que

\describe{

\item{\code{cand}}{Número de identificação do candidato.}
    
\item{\code{grupo}}{Grupo ao qual o candidato foi alocado, segundo 
    seu desempenho no processo seletivo (I, II ou III).}
    
\item{\code{nota}}{Nota da prova de conhecimento específico da área
    do programa.}
    
\item{\code{histor}}{Nota atribuída ao histórico escolar do 
    candidato.}
    
}}
\source{
Mingoti, S.A. (2005). Análise de dados através de métodos de 
    estatística multivariada - uma abordagem aplicada. 
    Belo Horizonte, MG: Editora UFMG. (pg 234)
}
\description{
Um programa de pós-graduação dividiu 62 candidatos do 
    ano anterior em 3 grupos: (I) candidatos aprovados, 
    (II) candidatos na lista de espera e (III) candidatos que não
    foram aprovados. Para a nova seleção, os responsáveis pensam em
    considerar a nota da prova de conhecimento específico e a nota 
    atribuída ao histórico escolar.
}
\examples{

data(MingotiAnA5)

library(lattice)

pairs(MingotiAnA5)

}
\keyword{AnaDisc}

