% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiTb2.1}
\alias{MingotiTb2.1}
\title{Rochas de uma determinada regi\enc{ã}{a}o}
\format{Um \code{data.frame} com 12 observações e 4 variáveis, em que

\describe{

\item{\code{rocha}}{Número da rocha amostrada.}

\item{\code{qrtz}}{Porcentagem de quartzo.}

\item{\code{fdsp}}{Porcentagem de feldspato.}

\item{\code{cor}}{Índice que mede a cor.}

}}
\source{
Mingoti, S. A. (2005). Análise de dados através de métodos de
    estatística multivariada: uma abordagem aplicada.  Belo
    Horizonte, MG: Editora UFMG. (pg 42).
}
\description{
Dados relativos a uma amostra de 12 rochas de uma certa
    região no qual tem as porcentagens de quartzo, feldspato e um
    índice que mede cor para cada uma das rochas.
}
\examples{

data(MingotiTb2.1)

pairs(~qrtz + cor + fdsp,
      data = MingotiTb2.1,
      main = "Matriz de dispersão das variáveis da rocha")

}
\keyword{TODO}

