% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mingoti.R
\name{MingotiTb3.5}
\alias{MingotiTb3.5}
\title{Impress\enc{õ}{o}es Sensoriais de Marcas de Coxinha de Galinha}
\format{Um \code{data.frame} com 8 observações e 5 variáveis, em que

\describe{

\item{\code{marca}}{Identificação da marca.}

\item{\code{sabor}}{Média das notas dos juízes para o sabor da
    coxinha de galinha.}

\item{\code{aroma}}{Média das notas dos juízes para o aroma da
    coxinha de galinha.}

\item{\code{massa}}{Média das notas dos juízes para a qualidade da
    massa.}

\item{\code{recheio}}{Média das notas dos juízes para a qualidade do
    recheio.}

}}
\source{
MINGOTI (2005), pág. 42.
}
\description{
Dados de 8 marcas de coxinha de galinha, avaliados em 4
    atributos (sabor, aroma, qualidade da massa e qualidade do
    recheio) avaliados por 5 julgadores numa escala de 1 a 5, quanto
    maior a nota melhor a classificação do atributo.  Os dados de
    entrada da tabela referem-se à média das notas dos 5 julgadores,
    para cada marca e cada atributo.
}
\examples{

data(MingotiTb3.5)
str(MingotiTb3.5)

library(car)

scatterplotMatrix(~sabor + aroma + massa + recheio,
                  data = MingotiTb3.5, smooth = FALSE)

}
\keyword{sensorial}

