% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MingotiTb8.1.R
\name{MingotiTb8.1}
\alias{MingotiTb8.1}
\title{Renda e número de filhos}
\format{data.frame com 3 observações e 5 variáveis, em que

\describe{

\item{renda}{Renda, medida em 3 categorias (menos de 2000, 2000 a 5000
         e 5000 ou mais).}

\item{f0}{Indivíduos que não tem filhos.}

\item{f1}{Indivíduos que tem 1 filho.}

\item{f2}{Indivíduos que tem 2 filhos.}

\item{f3}{Indivíduos que tem mais de 2 filhos.}
}}
\source{
MINGOTI, S.A. (2005). Análise de dados através de métodos de 
     estatística multivariada - uma abordagem aplicada. 
     Belo Horizonte, MG: Editora UFMG. (pg 258).
}
\description{
Dados relativos de 257
}
\examples{

data(MingotiTb8.1)

library(lattice)

xyplot(f0 + f1 + f2 + fmaior ~ renda, data = MingotiTb8.1)

}
\keyword{Porcentagem}
\keyword{rocha}

