% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEg3.6.11a}
\alias{PaulaEg3.6.11a}
\title{Exposi\enc{çã}{ca}o de Besouros}
\format{Um \code{data.frame} com 8 observações e 3 variáveis, em que

\describe{

\item{mortos}{Quantidade observada de besouros mortos.}

\item{exp}{Quantidade de besouros expostos a cada dose.}

\item{dose}{Dose de disulfeto de carbono gasoso à qual os besouros
   foram expostos.}

}}
\source{
PAULA (2004), Eg 3.6.11a pág. 237.
}
\description{
Dados de um estudo sobre o efeito da exposição de
    besouros adultos a diferentes doses de disulfeto de carbono
    gasoso \emph{(CS2)}, durante cinco horas. Foram registrados os
    números de besouros mortos.
}
\examples{

library(lattice)

data(PaulaEg3.6.11a)
str(PaulaEg3.6.11a)

xyplot(mortos/(sum(mortos)) ~ dose,
       data = PaulaEg3.6.11a,
       type = "o",
       xlab = "Dose de disulfeto de carbono gasoso",
       ylab = "Proporção de besouros mortos")

}
\keyword{MLG}
\keyword{binário}
