% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx1.13.22}
\alias{PaulaEx1.13.22}
\title{Vendas de Im\enc{ó}{o}veis}
\format{Um \code{data.frame} com 27 observações e 5 variáveis.

\describe{

\item{\code{impos}}{Valor cobrado de imposto, em 100 dólares.}

\item{\code{areat}}{Área do terreno, em 1000 pés quadrados.}

\item{\code{areac}}{Área construída, em 1000 pés quadrados.}

\item{\code{idade}}{Idade da residência, em anos.}

\item{\code{preco}}{Preço de venda do imóvel, em 1000 dólares.}

}}
\source{
PAULA (2004), Exercício 1.13.22, pág. 111.
}
\description{
Dados relativos a uma amostra de 27 imóveis vendidos. Os
    resgistros visam identificar as características que influenciam
    no preço de venda de um imóvel.
}
\examples{

data(PaulaEx1.13.22)
str(PaulaEx1.13.22)

library(lattice)

splom(PaulaEx1.13.22,
      type = c("p", "g", "smooth"),
      col.line = 1)

}
\keyword{RM}

