% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx2.10.15}
\alias{PaulaEx2.10.15}
\title{Consumo de Energia em Domic\enc{í}{i}lios}
\format{Um \code{data.frame} com 53 observações e 2 variáveis.

\describe{

\item{\code{consu}}{Consumo de energia num determinado mês, em
    quilowatts-hora.}

\item{\code{deman}}{Demanda de energia no horário de pico (unidade de
    medida não informada).}

}}
\source{
PAULA (2004), Exercício 2.10.15, pág. 178.
}
\description{
Dados referentes ao consumo de energia em 53 domicílios
    e demanda de energia no horário de pico.
}
\examples{

data(PaulaEx2.10.15)
str(PaulaEx2.10.15)

library(lattice)

xyplot(deman ~ consu, data = PaulaEx2.10.15,
       pch = 19, lwd = 2, type = c("p", "g", "smooth"))

}
\references{
Montgomery, D. C., Peck, E. A., Vining,
    G. G. (2001). Introduction to Linear Regression Analysis (3rd
    Ed.). John Wiley, New York.
}
\keyword{RS}
\keyword{heterovar}

