% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx3.7.14}
\alias{PaulaEx3.7.14}
\title{Confiabilidade de Equipamentos}
\format{Um \code{data.frame} com 4 variáveis.

\describe{

\item{\code{temp}}{Tempo de operação (1 a 5).}

\item{\code{equip}}{Tipo de equipamento (A, B ou C).}

\item{\code{nit}}{Número de equipamentos que não falharam até o tempo
    t, t = 1,2,3,4,5.}

\item{\code{yit}}{Número de falhas no intervalo entre os tempos t-1 e
    t.}

}}
\source{
PAULA (2004), Exercício 3.7.14, página 272.
}
\description{
Dados referentes aos tempos de falhas de equipamentos.
}
\examples{

library(lattice)

data(PaulaEx3.7.14)
str(PaulaEx3.7.14)

PaulaEx3.7.14$temp <- as.factor(PaulaEx3.7.14$temp)

xyplot(nit ~ temp,
       groups = equip,
       data = PaulaEx3.7.14,
       type = "o",
       auto.key = TRUE,
       xlab = "Tempos",
       ylab = "Número de equipamentos operantes")

}
\references{
Lawless, J. F. (1982). Statistical Models and Methods for
    Lifetime Data. John Wiley & Sons, New York. (Página 389)

Efron, B. (1988). Logistic regression, survival analysis,
    and the Kaplan-Meier curve. J. Amer. Stat. Assoc., 83.  (Páginas
    414-425)
}
\keyword{sobrevivência}

