% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaEx3.7.19}
\alias{PaulaEx3.7.19}
\title{Gestantes fumantes}
\format{Um \code{data.frame} com 5 variáveis.

\describe{

\item{\code{idade}}{Idade (<30 anos ou +30).}

\item{\code{ncigar}}{Número de cigarros consumidos por dia
(< 5 ou +5).}

\item{\code{tgest}}{Tempo de gestação (<= 260 dias ou >260).}

\item{\code{sobres}}{Número de crianças que sobreviveram.}

\item{\code{sobren}}{Número de crianças que não sobreviveram.}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP.
    (Exercício 3.7.19, página 276)
}
\description{
Estudo com gestantes fumantes, no qual as participantes
foram classificadas segundo os fatores de idade, número de cigarros
consumidos, tempo de gestação, e a condição (sobrevivência) da criança.
}
\examples{

data(PaulaEx3.7.19)

require(vcd)

# Paciente que sobreviveram
ss <- xtabs(sobres ~ idade + ncigar + tgest, PaulaEx3.7.19)

# Paciente que não sobreviveram
ns <- xtabs(sobren ~ idade + ncigar + tgest, PaulaEx3.7.19)

mosaic(ss,
       main = "Crianças que sobreviveram",
       labeling_args = list(
         set_varnames = c(ncigar = "Número de cigarros",
                          tgest = "Tempo de gestação")))

mosaic(ns,
       main = "Crianças que não sobreviveram",
       labeling_args = list(
         set_varnames = c(ncigar = "Número de cigarros",
                          tgest = "Tempo de gestação")))
}
\references{
Agresti A. (1990). Categorical Data Analysis. John Wiley,
New York. (página 253))
}
\keyword{ML}

