% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaEx3.7.22.R
\name{PaulaEx3.7.22}
\alias{PaulaEx3.7.22}
\title{Avalia\enc{çã}{ca}o de caduquice}
\format{Um \code{data.frame} com 55 observações e 2 variáveis.

\describe{

\item{\code{score}}{Escore do indivíduo no exame psicológico.}

\item{\code{resp}}{Resposta binária que representando a ocorrência
    (1) ou não ocorrência (0) do sintoma.}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP. (Exercício 3.7.22,
    pág. 278)
}
\description{
Os dados provém de um experimento com 54 indivíduos
    considerados idosos. Eles foram submetidos a um exame
    psiquiátrico para avaliar a ocorrência ou não de sintoma de
    caduquice.

    Acredita-se que o escore obtido em um exame feito previamente
    esteja associado com a ocorrência ou não do sintoma.
}
\examples{

library(lattice)

data(PaulaEx3.7.22)
str(PaulaEx3.7.22)

xyplot(resp ~ score, data = PaulaEx3.7.22,
       xlab = "Score do teste psicológico",
       ylab = "Sintoma de caduquice",
       jitter.y = TRUE, amount = 0.02)

}
\keyword{binomial}

