% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaEx3.7.23.R
\name{PaulaEx3.7.23}
\alias{PaulaEx3.7.23}
\title{Incidência de dengue}
\format{Um \code{data.frame} com 196 observações e 4 variáveis.
\describe{

\item{\code{idade}}{ Idade do entrevistado medida em anos. }

\item{\code{nivel}}{ Fator de 3 níveis representando o nível
    sócio-econômico do entrevistado (alto, médio, baixo). }

\item{\code{setor}}{ Fator de 2 níveis representando o
    setor da cidade que o entrevistado mora. }

\item{\code{caso}}{ Fator de 2 níveis que representa se o entrevistado
    contraiu (1) ou não contraiu (0) a doença recentemente. }

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
computacional. São Paulo, SP: IME-USP. (Exercício 3.7.23, pág. 279)
}
\description{
Os dados provém de um estudo para investigar a incidência
    de dengue numa determinanda cidade da costa mexicana.
    
    Foram escolhidos aleatóriamente 196 indivíduos de dois setores
    da cidade e cada um repondeu às seguintes perguntas: idade, 
    nível sócio-econômico, setor da cidade onde mora e se contraiu
    a doença recentemente.
    
    Um dos objetivos do estudo é tentar prever ou explicar a 
    probabilidade de um indivíduo contrair a doença dado as variáveis
    explicativas: \code{idade}, \code{nivel} e \code{setor}.
}
\examples{

library(lattice)

data(PaulaEx3.7.23)

histogram(caso ~ idade, data = PaulaEx3.7.23, xlab = "Idade", 
    ylab = "Percentual do Total")
histogram(caso ~ setor, data = PaulaEx3.7.23, xlab = "Setor", 
    ylab = "Percentual do Total")

}
\keyword{TODO}

