% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaulaEx5.6.14.R
\name{PaulaEx5.6.14}
\alias{PaulaEx5.6.14}
\title{Experimento em que ratos com leucemia induzida}
\format{Um \code{data.frame} com 120 observações e 5 variáveis.
\describe{

\item{\code{rato}}{Número de identificação do rato.}

\item{\code{period}}{Quatro diferentes períodos que foram avaliadas
    a quantidade de células.}

\item{\code{celubran}}{Quantidade de células brancas.}

\item{\code{celuverm}}{Quantidade de células vermelhas.}

\item{\code{celucanc}}{Número de colônias de células cancerosas.}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
computacional. São Paulo, SP: IME-USP. (Exercício 5.6.14, pág. 401)
}
\description{
Dados referente a experimento em que 30 ratos tiveram
    uma condição de leucemia induzida com três drogas quimioterápicas.
    Foram coletadas de cada animal a quantidade de células brancas, a 
    quantidade de células vermelhas e o número de colônias de células 
    cancerosas em quatro períodos diferentes.
}
\examples{

library(lattice)

data(PaulaEx5.6.14)

xyplot(celucanc ~ celubran + celuverm, 
       type = c("p", "g", "a"), data = PaulaEx5.6.14)
       
}
\references{
Myers, R.H.; Montgomery, D. C. e Vining, G. G. (2002).
Generalized Linear Models: With Applications in Engineering and the 
Sciences. John Wiley, New York.
}
\keyword{quase-verossimilhança}

