% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paula.R
\name{PaulaTb1.9}
\alias{PaulaTb1.9}
\title{Bact\enc{é}{e}rias Sobreviventes em Amostras}
\format{Um \code{data.frame} com 12 observações e 2 variáveis.
\describe{

\item{\code{num}}{Número de bactérias sobreviventes.}

\item{\code{temp}}{Tempo de exposição (em minutos).}

}}
\source{
Paula, G. A. (2004). Modelos de regressão: com apoio
    computacional. São Paulo, SP: IME-USP. (Tabela 1.9, pág. 88)
}
\description{
Número de bactérias sobreviventes em amostras de um
    produto alimentício segundo o tempo de exposição do produto a uma
    temperatura de 300°F.
}
\examples{

data(PaulaTb1.9)

str(PaulaTb1.9)

library(lattice)

xyplot(num ~ temp,
       ylab = "Número de bactérias",
       xlab = "Tempo de exposição",
       data = PaulaTb1.9,
       type = c("o"),
       main = "Número de Bactérias Sobreviventes por Tempo de Exposição")

}
\keyword{MLG}

