% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelEg5.2.R
\name{PimentelEg5.2}
\alias{PimentelEg5.2}
\title{Competição de Variedades de Batatinha}
\format{Um \code{data.frame} com 32 observações e 3 variáveis, em que

\describe{

\item{\code{bloco}}{Fator com 4 níveis.}

\item{\code{variedade}}{Fator de 8 correspondente às
    variedades de batatinha.}

\item{producao}{Produção de batatinha, em ton ha\eqn{^{-1}}.}

}}
\source{
PIMENTEL-GOMES (2009), Exemplo 5.2, pág. ??.
}
\description{
Dados de um experimento de competição de variedades de batatinha
    feito pelo Engenheiro Agrônomo Oscar A. Garay em Balcare,
    Argentina. O experimento foi realizado em blocos casualizados.
}
\examples{

library(lattice)

data(PimentelEg5.2)
str(PimentelEg5.2)

xtabs(~variedade + bloco, data = PimentelEg5.2)

xyplot(producao ~ variedade,
       groups = bloco,
       data = PimentelEg5.2,
       type = "p",
       ylab = expression(Produção ~ (t ~ ha^{-1})),
       xlab = "Variedades de batatinha")

}
\keyword{DBC}
