% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb10.3.1.R
\name{PimentelTb10.3.1}
\alias{PimentelTb10.3.1}
\title{Experimento em Blocos Imcompletos do Tipo I}
\format{Um \code{data.frame} com 56 observações e 4 variáveis, em que

\describe{

\item{\code{rept}}{Fator de 7 níveis qualitativos, que são as
    repetições do experimento. Cada repetição tem 4 blocos de tamanho
    2 e um parcela de cada variedade (8 parcelas).}

\item{\code{trat}}{Fator de 8 níveis qualitativos que são os
    tratamentos estudados.}

\item{\code{bloc}}{Fator de 4 níveis qualitativos que identifica os
    blocos dentro de uma repetição. Os blocos tem tamanho 2.}

\item{y}{Variável medida nas parcelas. Unidade de medida não
    fornecida.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 10.3.1, pág. 190)
}
\description{
Experimento em blocos imcompletos equilibrados, no qual
    temos \eqn{v = 8} tratamentos, \eqn{\lambda = 1}, \eqn{k = 2}
    parcelas por bloco, \eqn{r = 7} repetições, \eqn{b = 28} blocos e
    \eqn{E = 55\%}, em que os blocos podem ser agrupados em
    repetições.
}
\examples{

data(PimentelTb10.3.1)
str(PimentelTb10.3.1)

library(lattice)

xtabs(~trat + rept, data = PimentelTb10.3.1)
ftable(xtabs(~trat + bloc + rept, data = PimentelTb10.3.1))

xyplot(y ~ trat,
       groups = bloc,
       auto.key = list(title = "Blocos",
                       cex.title = 1.1, columns = 4),
       data = PimentelTb10.3.1,
       xlab = "Variedade",
       ylab = "Resposta")

}
\keyword{DBI}

