% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb12.3.1.R
\name{PimentelTb12.3.1}
\alias{PimentelTb12.3.1}
\title{Temperaturas M\enc{á}{a}ximas M\enc{é}{e}dias em Piracicaba}
\format{Um \code{data.frame} com 15 observações e 2 variáveis, em que

\describe{

\item{\code{ano}}{Variável que indica o ano em que foram registradas
    as temperaturas máximas médias do mês de Julho.}

\item{\code{temperatura}}{Temperatura máxima média de Julho em graus
    centígrados.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 12.3.1, pág. 236)
}
\description{
Série de 15 anos das temperaturas máximas médias de
    Julho em Piracicaba-SP, em graus centígrados.
}
\examples{

library(lattice)

data(PimentelTb12.3.1)
str(PimentelTb12.3.1)

xyplot(temperatura ~ ano,
       data = PimentelTb12.3.1, type = "o",
       ylab = expression("Temperatura"~(degree*C)),
       xlab = "Ano")

}
\keyword{RP}

