% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelTb17.3.1}
\alias{PimentelTb17.3.1}
\title{Produ\enc{çã}{ca}o de Lenha de Esp\enc{é}{e}cies de Eucaliptos}
\format{Um \code{data.frame} com 100 observações e 4 variáveis, em
    que

\describe{

\item{\code{exper}}{Fator de 2 níveis qualitativos, usado para
    diferenciar os experimentos que foram instalados lado a lado.}

\item{\code{bloc}}{Fator de 5 níveis qualitativos que identifica os
    blocos em cada experimento.}

\item{\code{espec}}{Fator de 19 níveis qualitativos, que são as
    espécies de eucalipto.}

\item{\code{prod}}{Produção de lenha, em metros cúbicos por parcela,
    com corte aos 8 anos de idade.}

}}
\source{
PIMENTEL-GOMES (2009), Tabela 17.3.1, pág. 317.
}
\description{
Grupo de dois experimentos com espécies de eucaliptos
    realizados pela Companhia Paulista de Estradas de Ferro
    (Pimentel-Gomes e Guimarães, 1958). Ambos os ensaios, localizados
    lado a lado no campo, tinham 10 espécies, em 5 blocos
    casualizados. Uma única espécie (\emph{Eucalyptus saligna})
    ocorria em ambos os experimentos.
}
\examples{

library(lattice)

data(PimentelTb17.3.1)
str(PimentelTb17.3.1)

ftable(xtabs(~exper + bloc + espec, data = PimentelTb17.3.1))

xyplot(prod ~ espec,
       data = PimentelTb17.3.1,
       ylab = expression("Produção de lenha"~(m^3 ~ parcela^{-1})),
       xlab = "Espécies de eucalipto")

}
\keyword{DBC}
\keyword{GE}

