% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb20.2.1.R
\name{PimentelTb20.2.1}
\alias{PimentelTb20.2.1}
\title{Ensaio Fatorial de Aduba\enc{çã}{ca}o de Cana-de-a\enc{çú}{cu}car}
\format{Um \code{data.frame} com 9 observações e 3 variáveis, em que

\describe{

\item{\code{P}}{Variável que indica os níveis de Fósforo (P) em cada
    parcela.}

\item{\code{K}}{Variável que indica os níveis de Potássio (K) em cada
    parcela.}

\item{\code{totais}}{Produção total nos 6 blocos, em ton
    ha\eqn{^{-1}}. Valores individuais não disponíveis.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 20.2.1, pág. 369)
}
\description{
Ensaio fatorial, de \eqn{3^{2}}, de adubação de
    cana-de-açúcar com P e K, em 6 blocos casualizados.
}
\examples{

library(lattice)

data(PimentelTb20.2.1)
str(PimentelTb20.2.1)

xyplot(totais ~ P, data = PimentelTb20.2.1,
       groups = K, type = "o",
       auto.key = list(title = "Níveis de Potássio (K)",
                       cex.title = 1.1, columns = 3),
       ylab = "Totais de tratamentos (ton/ha)",
       xlab = "Níveis de fósforo (P)")

}
\keyword{FAT2}

