% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb6.3.1.R
\name{PimentelTb6.3.1}
\alias{PimentelTb6.3.1}
\title{Experimento de Adubação de Cana-de-açúcar}
\format{Um \code{data.frame} com 36 observações e 4 variáveis, em que

\describe{

\item{\code{linhas}}{Fator de 6 níveis qualitativos, usado para
    identificar as linhas do quadrado latino de 6 x 6.}

\item{\code{colunas}}{Fator de 6 níveis qualitativos, usado para
    identificar as colunas do quadrado latino de 6 x 6.}

\item{\code{adub}}{Fator de 6 níveis qualitativos, que são os 6
    diferentes tipos de tratamento. A variável assume os valores 0 ou
    1, sendo que 0 representa ausência e 1 representa presença de
    determinada característica. A primeira posição representa a
    presença ou ausência de Calcário; A segurança posição representa
    a presença ou ausência de Crotalária; E a terceira posição, caso
    exista, representa a presença ou ausência de Adubo Mineral.}

\item{\code{prod}}{Totais de açúcar provável em kg por parcela.}

}}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 6.3.1)
}
\description{
Experimento de adubação de cana da Usina Monte Alegre,
    com fertilizantes minerais e adubos verdes. O delineamento
    utilizado foi um quadrado latino de 6 x 6.
}
\examples{

library(lattice)

grafico1 = bwplot(prod ~ adub,
                  data = PimentelTb6.3.1,
                  xlab = "Tipo de Adubação",
                  ylab = "Produção em kg por parcela")

grafico2 = xyplot(prod ~ linhas + colunas,
                  data = PimentelTb6.3.1,
                  xlab = "Linha (Azul) + Coluna (Rosa)",
                  ylab = "Produção em kg por parcela")

print(grafico1, position = c(0, 0, 0.5, 1), more = TRUE)

print(grafico2, position = c(0.5, 0, 1, 1))

}
\keyword{DQL}

