% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimentelTb8.3.1.R
\name{PimentelTb8.3.1}
\alias{PimentelTb8.3.1}
\title{Grupo de Ensaios de Batatinha na Província de Buenos Aires}
\format{Um \code{data.frame} com 56 observações e 3 variáveis, em que

\describe{

\item{\code{varied}}{Fator de 8 níveis qualitativos que são as
    variedades de batatinha.}

\item{\code{exper}}{Fator de 7 níveis qualitativos que refere-se aos
    experimentos conduzidos em diferentes localidades na região
    batateira da Província de Buenos Aires.}

\item{\code{totais}}{Totais de variedades em cada experimento,
    resultado da soma da produção (t/ha) das parcelas de cada um dos
    blocos.}

}

O atributo \code{qmr} é um vetor com os quadrados médios residuais da
    análise de variância de cada experimento, segundo um modelo para
    o delineamento de blocos casualizados. Um exemplo de como são dos
    dados individuais está em \code{\link{PimentelEg5.2}}, que
    refere-se ao experimento número 3 com todas as observações.}
\source{
Pimentel-Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, SP: FEALQ. (Tabela 8.3.1, pág. 147)
}
\description{
Grupo de ensaios de competição de variedades de
    batatinha, realizados pelo engenheiro agrônomo Oscar A. Garay, da
    Estação Experimental de Balcarce, Argentina, instalados em
    localidades da região batateira da Província de Buenos Aires.
    Cada um deles tinha 4 repetições e 8 variedades em blocos
    casualizados. Os dados, no entanto, não são os valores
    individuais mas os totais das produções das variedades em cada
    experimento, somando as parcelas de todos os blocos. Os quadrados
    médios dos resíduos de cada experimento estão disponíveis no
    atributo objeto.
}
\examples{

library(lattice)

data(PimentelTb8.3.1)
str(PimentelTb8.3.1)
attr(PimentelTb8.3.1, "qmr")

ord <- with(PimentelTb8.3.1, order(exper, varied))
PimentelTb8.3.1 <- PimentelTb8.3.1[ord, ]

xyplot(totais ~ varied,
       groups = exper, type = "o",
       data = PimentelTb8.3.1,
       xlab = "Variedades de batatinha",
       ylab = "Totais das variedades",
       auto.key = list(title = "Experimentos", cex.title = 1,
                       columns = 4))

}
\seealso{
\code{\link{PimentelEg5.2}}.
}
\keyword{GE}

