% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimentel.R
\name{PimentelTb9.2.1}
\alias{PimentelTb9.2.1}
\title{Produ\enc{çã}{ca}o de Adubos Verdes e Milho}
\format{Um \code{data.frame} com 64 observações e 4 variáveis, em que

\describe{

\item{\code{ano}}{Fator de 2 níveis qualitativos, que diferencia os
    dois anos sucessivos, tratados como subparcelas.}

\item{\code{bloco}}{Fator de 4 níveis qualitativos, usado para
    controle local.}

\item{\code{cultura}}{Fator de 8 níveis qualitativos, que são os
    diferentes tipos de cultura, variando entre adubos verdes e
    milho.}

\item{\code{prod}}{Produção de adubos verdes e milho medidos em kg de
    matéria verde por parcela.}

}}
\source{
PIMENTEL-GOMES (2009), Tabela 9.2.1, pág. 166.
}
\description{
Experimento com 8 tratamentos (7 adubos verdes e milho)
    em blocos ao acaso, com 4 repetições, realizado em dois anos
    sucessivos nas mesmas parcelas.
}
\examples{

library(lattice)

data(PimentelTb9.2.1)
str(PimentelTb9.2.1)

xyplot(prod ~ cultura,
       groups = ano, type = c("p", "a"),
       data = PimentelTb9.2.1,
       scales = list(x = list(rot = 90)),
       auto.key = list(title = "Ano", cex.title = 1.1,
                       columns = 2),
       xlab = "Culturas",
       ylab = "Produção de matéria verde (kg/parcela)")

}
\keyword{PSS}

