% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoEg7.8.R
\name{RamalhoEg7.8}
\alias{RamalhoEg7.8}
\title{Produção de Grãos de Cruzamentos Feijão}
\format{Um \code{data.frame} com 240 observações e 4 variáveis, em
    que

\describe{

\item{\code{fam}}{Família \eqn{F_{5}} do cruzamento Jalo x Small
    White.}

\item{\code{rept}}{Fator que indica a repetição da combinação entre 
    família \eqn{F_{5}} e planta.}

\item{\code{plant}}{Inteiro que indica a planta de feijão avaliada no
    experimento.}

\item{\code{prod}}{Produção de grãos de feijão, em g/planta.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras, MG: UFLA. (Exemplo 7.8, pág 110).
}
\description{
Experimento referente à produção de grãos de feijão
    obtida na avaliação de 24 famílias \eqn{F_{5}} do cruzamento
    Jalo x Small White, cujos dados foram obtidos por Souza (1991).
    Experimento realizado em blocos incompletos.
}
\examples{

library(lattice)

xtabs(~fam + rept, data = RamalhoEg7.8)

ps <- list(box.rectangle = list(col = 1, fill = c("gray90")),
           box.umbrella = list(col = 1, lty = 1),
           plot.symbol = list(col = 1, cex = 0.7))

bwplot(prod ~ fam, data = RamalhoEg7.8,
       horizontal = FALSE, pch = "|",
       xlab = "Família", ylab = "Produção de grãos",
       par.settings = ps)

}
\keyword{DBC}

