% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramalho.R
\name{RamalhoEg8.8}
\alias{RamalhoEg8.8}
\title{Produ\enc{çã}{ca}o de Gr\enc{ã}{a}os de Feij\enc{ã}{a}o de
    Popula\enc{çõ}{co}es Segregantes}
\format{Um \code{data.frame} com 72 observações e 4 variáveis, em que

\describe{

\item{\code{ger}}{Fator categórico de 3 níveis que representa o
    experimento para cada geração segregante de feijoeiro.}

\item{\code{bloc}}{Fator categórico usado para fazer controle local
    dentro de cada experimento (geração).}

\item{\code{pop}}{Fator categórico de 6 níveis que representa as
    populações segregates de feijoeiro.}

\item{\code{prod}}{Produção de grãos de feijão, em kg ha\eqn{^{-1}}.}

}}
\source{
RAMALHO et al. (2005), Tabela 8.8, pág. 123.

    Corte, H. R., Ramalho, M. A. P., Gonçalves, F. M. A., Abreu. A de
    F. B. Natural Selection for Grain Yield in Dry Bean Population
    Bred by the Bulk Method. Euphytica. Wageningen. v.123, n.3. 2002
}
\description{
Dados da produção de grãos de feijão obtidos por Corte
    et al. (2002) relativos à avaliação de 6 populações segregantes
    avaliadas nas gerações F\eqn{_{2}}, F\eqn{_{3}} e F\eqn{_{4}} com
    diferente número de blocos. O delineamento foi em blocos
    casualizados e cada geração corresponde a um experimento
    separado. Sendo assim, a identificação dos blocos é para cada
    experimento (geração).
}
\examples{

library(lattice)

data(RamalhoEg8.8)

str(RamalhoEg8.8)

xtabs(~pop + ger, data = RamalhoEg8.8)

ftable(with(RamalhoEg8.8,
            tapply(prod,
                   list(ger = ger, pop = pop, bloc = bloc),
                   FUN = identity)))

xyplot(prod ~ pop | ger, data = RamalhoEg8.8,
       groups = bloc, type = c("p", "a"),
       as.table = TRUE, layout = c(NA, 1),
       xlab = "População",
       ylab = expression("Produção"~(kg~ha^{-1})))

}
\keyword{GE}
