% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramalho.R
\name{RamalhoEx12.2}
\alias{RamalhoEx12.2}
\title{Dados Gen\enc{é}{e}ricos para Regress\enc{ã}{a}o
    M\enc{ú}{u}ltipla}
\format{Um \code{data.frame} com 8 observações e 4 variáveis, em que

\describe{

\item{\code{cult}}{Fator que indica as cultivares.}

\item{\code{prod}}{Produtividade de grãos, medidos em kg/ha.}

\item{\code{alt}}{Altura da planta, em centímetros.}

\item{\code{perf}}{Número de perfilhos.}

}}
\source{
RAMALHO et al. (2005), Exercício 12.2, pág. 231.
}
\description{
Conjunto de dados apresentados para exercício. Objetivo
    do exercício é estimar a equação de regressão múltipla.
}
\examples{

data(RamalhoEx12.2)

str(RamalhoEx12.2)

pairs(RamalhoEx12.2[, -1])

library(lattice)

splom(RamalhoEx12.2[, -1], type = c("p", "r"))

}
\keyword{RM}
