% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramalho.R
\name{RamalhoEx3.1}
\alias{RamalhoEx3.1}
\title{N\enc{ú}{u}mero de plantas de milho ap\enc{ó}{o}s
    inocula\enc{çã}{ca}o}
\format{data.frame com 20 observações e 4 variáveis, em que

\describe{

\item{\code{fugic}}{Fator que indica o tipo de tratamento aplicado.
    Fungicida (A, B, C, D, E) ou controle (Fc).}

\item{\code{rept}}{Repetições de inoculação do fungicida.}

\item{\code{plant}}{Número de plantas de milho observadas.}

}}
\source{
RAMALHO et al. (2005), pág. 50.
}
\description{
Experimento referente ao número de plantas de milho,
    por parcela, após a inoculação com Diploidia. Experimento para
    avaliação de fungicidas para tratamento de sementes.
}
\examples{

library(lattice)

data(RamalhoEx3.1)

xyplot(jitter(plant) ~ fungic, data = RamalhoEx3.1,
       groups = repet, auto.key = TRUE,
       xlab = "Fungicida", ylab = "Número de plantas")

}
\keyword{contagem}

