% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RamalhoTb1.2.R
\name{RamalhoTb1.2}
\alias{RamalhoTb1.2}
\title{Produtividade Média de Espiga}
\format{data.frame com 28 observações e 2 variáveis, em que

\describe{

\item{varied}{Fator de níveis nominais das variedades de milho.}

\item{prod}{Produtividade média de espiga (ton ha\eqn{^{-1}}).}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (pg 15)
}
\description{
Experimento da produtividade média de espigas de milho
    de diversas cultivares obtidas por Ferreira et al (1995).
}
\examples{
bk <- c(2.5, 4.25, 6, 7.75, 9.5)
ht <- hist(RamalhoTb1.2$prod,
           xlab = expression("Produtividade"~(ton~ha^{-1})),
           ylab = "Frequência absoluta",
           main = NULL,
           breaks = bk,
           axes = FALSE,
           labels = TRUE)
axis(side = 1, at = bk)
axis(side = 2)
rug(RamalhoTb1.2$prod)

plot(ecdf(RamalhoTb1.2$prod),
     xlab = expression("Produtividade"~(ton~ha^{-1})),
     ylab = "Frequência relativa acumulada",
     main = NULL)
}
\keyword{ASS}

