% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramalho.R
\name{RamalhoTb13.1}
\alias{RamalhoTb13.1}
\title{Produ\enc{çã}{ca}o de Gr\enc{ã}{a}os de Milho}
\format{Um \code{data.frame} com 10 observações e 2 variáveis, em que

\describe{

\item{\code{x}}{Altura da planta, medida em cm.}

\item{\code{y}}{Produção de grãos, medida em g/planta.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (Tabela 13.1, pág. 233)
}
\description{
Experimento referente a cultura de milho, cujos dados
    foram obtidos de dez plantas. Neste experimento, a produção de
    grãos é considerada a variável dependente e a altura da planta a
    independente.
}
\examples{

data(RamalhoTb13.1)

str(RamalhoTb13.1)

library(lattice)

xyplot(y ~ x, data = RamalhoTb13.1,
       type = c("p", "r"),
       xlab = "Altura (cm)",
       ylab = "Produção (g/planta)")

}
\keyword{COV}

