% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoTb13.15.R
\name{RamalhoTb13.15}
\alias{RamalhoTb13.15}
\title{Produção de Raízes de Mandioca}
\format{Um \code{data.frame} com 20 observações e 3 variáveis, em que

\describe{

\item{\code{clone}}{Fator de 20 níveis que indica o clone de
    mandioca.}

\item{\code{mae}}{Produção de ramas da planta mãe.}

\item{\code{filha}}{Produção de ramas da planta filha.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (Tabela 13.15, pág. 252)
}
\description{
Experimento envolvendo a produção de ramas (parte aérea)
    de 20 genótipos parentais (clones) de mandioca e de suas
    respectivas filhas.
}
\examples{

data(RamalhoTb13.15)

str(RamalhoTb13.15)

library(lattice)

xyplot(filha ~ mae, data = RamalhoTb13.15,
       type = c("p", "smooth", "g"), aspect = "iso",
       xlab = "Produção de ramas na planta mãe",
       ylab = "Produção de ramas na planta filha")

}
\keyword{COV}
\keyword{RS}

