% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamalhoTb3.1.R
\name{RamalhoTb3.1}
\alias{RamalhoTb3.1}
\title{Número de nematóides por vasos}
\format{data.frame com 16 observações e 3 variáveis, em que

\describe{

\item{\code{dias}}{Idade de inoculação, tratamento.}

\item{\code{rept}}{Número que indica as repetições de cada
    tratamento.}

\item{\code{nemat}}{Número de nematóides por vaso coletado.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras, MG: UFLA. (pg 43)
}
\description{
Experimento do número de nematóides (vermes que estão
    presentes no solo) por vasos infectando plantas de figo em
    diferentes idades de inoculação, experimento com 4 repetições.
}
\examples{

library(lattice)

data(RamalhoTb3.1)

# Dados originais.

aggregate(nemat ~ dias, data = RamalhoTb3.1,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

plot(nemat ~ dias, data = RamalhoTb3.1,
     ylab = "Número de nematóides",
     xlab = "Tempo (dias)")

xyplot(nemat ~ dias, data = RamalhoTb3.1,
       ylab = "Número de nematóides",
       xlab = "Tempo (dias)")

# Dados aplicando logaritmo.

aggregate(log(nemat) ~ dias, data = RamalhoTb3.1,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

plot(log10(nemat) ~ dias, data = RamalhoTb3.1,
     ylab = "log do número de nematóides",
     xlab = "Tempo (dias)")

plot(nemat ~ dias, data = RamalhoTb3.1,
     log = "y",
     ylab = "log do número de nematóides",
     xlab = "Tempo (dias)")

xyplot(nemat ~ dias, data = RamalhoTb3.1,
       scales = list(y = list(log = TRUE)),
       ylab = "Número de nematóides",
       xlab = "Tempo (dias)")

}
\keyword{DIC}
\keyword{contagem}

