% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RamalhoTb3.4.R
\name{RamalhoTb3.4}
\alias{RamalhoTb3.4}
\title{Incidênica de \emph{Colletotrichum} no feijoeiro}
\format{data.frame com 48 observações e 4 variáveis, em que

\describe{

\item{trat}{Tratamento.}

\item{bloc}{Blocos onde são aplicados os tratamentos.}

\item{cult}{Cultivar de feijão.}

\item{nota}{Variável reposta que é a nota atribuída pelos avaliadores
    para a incidência e \emph{Colletotrichum}.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (pg 45)
}
\description{
Neste experimento foram avaliados 16 tratamentos com 3
    observações em blocos casualizados referentes a incidência de
    Colletotrichum no feijoeiro, as notas foram dadas por 3
    avaliadores.
}
\examples{
library(lattice)

data(RamalhoTb3.1)

xyplot(nota ~ cult, data = RamalhoTb3.4,
       groups = bloc, auto.key = TRUE, jitter.x = TRUE,
       xlab = "Cultivares",
       ylab = "Notas")
}
\keyword{DBC}

