% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RamalhoTb3.6.R
\name{RamalhoTb3.6}
\alias{RamalhoTb3.6}
\title{Largura de ascos \emph{Colletrotrichum lindemuthianum}}
\format{data.frame com 90 observações e 3 variáveis, em que

\describe{

\item{amostra}{Número identificador da amostra.}

\item{isol}{Isolados do fungo (A, B e C).}

\item{larg}{Largura dos ascos dos isolados (TODO unidade de medida).}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras: UFLA. (pg 48)
}
\description{
Experimento referente a largura dos ascos de três
    isolados do fungo \emph{Colletotrichum lindemuthianum}.
}
\examples{
data(RamalhoTb3.6)

aggregate(larg ~ isol,  data = RamalhoTb3.6,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })
}
\keyword{TODO}

