% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramalho.R
\name{RamalhoTb3.6}
\alias{RamalhoTb3.6}
\title{Largura de ascos \emph{Colletrotrichum lindemuthianum}}
\format{data.frame com 90 observações e 3 variáveis, em que

\describe{

\item{\code{amostra}}{Número identificador da amostra.}

\item{\code{isol}}{Isolados do fungo (A, B e C).}

\item{\code{larg}}{Largura dos ascos dos isolados.}

}}
\source{
Ramalho, M. A. P., Ferreira, D. F., Oliveira, A. C. (2005).
    Experimentação em Genética e Melhoramento de Plantas (2th ed.).
    Lavras, MG: UFLA. (pg 48)
}
\description{
Experimento referente a largura dos ascos de três
    isolados do fungo \emph{Colletotrichum lindemuthianum}.
}
\examples{

data(RamalhoTb3.6)

aggregate(larg ~ isol,  data = RamalhoTb3.6,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

}
\keyword{DBC}

