% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamosTb2.5.R
\name{RamosTb2.5}
\alias{RamosTb2.5}
\title{Resistência à ruptura de garrafas}
\format{Um \code{data.frame} com 9 observações e 2 variáveis, em que

\describe{

\item{\code{forca}}{Resistência das garrafas.}

\item{\code{freq}}{Frequências observadas para as resistências.}

}}
\source{
Ramos, E. M. L. S., et al.(2013). 
    Controle Estatístico da Qualidade (1th ed.). 
    Porto Alegre, RS: Bookman. (pg 29).
}
\description{
Distribuição de frequências para as 
    resistências à  ruptura de 100 garrafas de 
    refrigerante de um litro.
}
\examples{

data(RamosTb2.5)

barplot(RamosTb2.5$freq, names=RamosTb2.5$forca, 
      xlab = 'Resistência', ylab = 'Frequência')


}
\keyword{CEQ}

