% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamosTb2.6.R
\name{RamosTb2.6}
\alias{RamosTb2.6}
\title{Resistência à ruptura e peso de garrafas}
\format{Um \code{data.frame} com 25 observações e 2 variáveis, em que

\describe{

\item{\code{forca}}{Força de resistência à ruptura.}

\item{\code{peso}}{Peso da garrafa (em gramas).}

}}
\source{
Ramos, E. M. L. S., et al.(2013). 
    Controle Estatístico da Qualidade (1th ed.). 
    Porto Alegre, RS: Bookman. (pág. 30).
}
\description{
Resistência e peso de 25 garrafas.
}
\examples{

data(RamosTb2.6)

library(lattice)

xyplot(forca ~ peso, pch = 20,
       data = RamosTb2.6, type = c("g", "p"))

}
\keyword{CEQ}

