% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramos.R
\name{RamosTb5.2}
\alias{RamosTb5.2}
\title{Itens n\enc{ã}{a}o Conformes no Processo Produtivo de Ovos de
    Galinha}
\format{Um \code{data.frame} com 30 observações e 3 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{tamamostra}}{Número de unidades amostradas.}

\item{\code{naoconf}}{Número de ovos não conformes na amostra.}

}}
\source{
RAMOS et al. (2013), pág. 79.
}
\description{
Número de ovos não conformes de um processo produtivo de
    ovos de galinha, com 30 amostras de tamanhos diferentes.
}
\examples{

data(RamosTb5.2)

library(qcc)

qcc(RamosTb5.2$naoconf, sizes = RamosTb5.2$tamamostra, type = "p",
    xlab = "Amostra", ylab = "Fração de ovos não conformes",
    title = "")

}
\keyword{CEQ}
