% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamosTb5.8.R
\name{RamosTb5.8}
\alias{RamosTb5.8}
\title{Número médio de defeitos na fabricação de copos de cristal}
\format{Um \code{data.frame} com 26 observações e 3 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{tamamostra}}{Número de unidades amostradas.}

\item{\code{naoconf}}{Número médio de unidades não conformes na amostra.}
    
}}
\source{
Ramos, E. M. L. S., et al.(2013). 
    Controle Estatístico da Qualidade (1th ed.). 
    Porto Alegre, RS: Bookman. (pg 96).
}
\description{
Número médio de defeitos por unidade e número médio
    de não conformidades estimados de um processo de fabricação
    de copos de cristal, com tamanhos diferentes.
}
\examples{

data(RamosTb5.8)

library(qcc)

qcc(RamosTb5.8$naoconf, sizes = RamosTb5.8$tamamostra, type = "u")

}
\keyword{CEQ}

