% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RamosTb6.1.R
\name{RamosTb6.1}
\alias{RamosTb6.1}
\title{Densidade aparente de um processo de eletrodos}
\format{Um \code{data.frame} com 30 observações e 2 variáveis, em que

\describe{

\item{\code{amostra}}{Número da amostra.}

\item{\code{densi}}{Densidade aparente (em g/cm^{3} do eletrodo).}
 
}}
\source{
Ramos, E. M. L. S., et al.(2013). 
    Controle Estatístico da Qualidade (1th ed.). 
    Porto Alegre, RS: Bookman. (pg 106).
}
\description{
Dados da densidade aparente de um processo de produtivo  
    de eletrodos de carbono.
}
\examples{

data(RamosTb6.1)

library(qcc)

qcc(RamosTb6.1$densi, type = "xbar.one", 
     std.dev = "SD", nsigmas = 3, xlab = "Amostra", ylab = "Densidade",
     title = "")

}
\keyword{CEQ}

