% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storck.R
\name{StorckEg2.3.5}
\alias{StorckEg2.3.5}
\title{Rendimento de Cultivares de Alho}
\format{Um \code{data.frame} com 16 observações e 4 variáveis, em que

\describe{

\item{\code{fila}}{Fator que indica a fila ou curva nível,
    considerada para blocar a fertilidade do solo que muda entre as
    curvas de nível.}

\item{\code{col}}{Fator que bloca o tamanho do bulbo de alho (florão,
    graudo, médio e miúdo) usado para o plantio.}

\item{\code{trat}}{Fator categórico que representa as cultivares de
    alho.}

\item{\code{rend}}{Rendimento, em toneladas por hectare.}

}}
\source{
Storck, L., Garcia, B. C., Lopes, S. J., Estefanel,
    V. (2011). Experimentação Vegetal (3th ed.). Santa Maria, RS:
    UFSM. (Tabela 2.3.5, pág 63)
}
\description{
Dados de um experimento no delineamento quadrado latino
    em que foram avaliados os rendimentos (t/ha) de quatro cultivares
    de alho. Na escolha do delineamento, o bloqueamento de linhas foi
    em razão da heterogeniedade da fertilidade entre as curvas de
    nível (cada curva igual a uma linha) e o bloqueamento de colunas
    foi devido à heterogeneidade entre os tamanhos dos bulbos de
    alho.
}
\examples{

library(lattice)

data(StorckEg2.3.5)
str(StorckEg2.3.5)

xyplot(rend ~ cult, data = StorckEg2.3.5,
       xlab = "Cultivares", ylab = "Rendimento")

}
\keyword{DQL}

