% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StorckTb101.R
\name{StorckTb101}
\alias{StorckTb101}
\title{Grupo de Experimentos de Competição de Cultivares de Milho}
\format{Um \code{data.frame} com 96 observações e 4 variáveis, em que

\describe{

\item{\code{loc}}{Fator categórico que representa os locais
    (ambientes) de instalação dos experimentos.}

\item{\code{cult}}{Fator categórico que representa as cultivares de
    milho.}

\item{\code{bloc}}{Fator categórico que representa os blocos em cada
    um dos locais.}

\item{\code{result}}{Rendimento de grãos, em kg por parcela.}

}}
\source{
Storck, L., Garcia, B. C., Lopes, S. J., Estefanel,
    V. (2011). Experimentação Vegetal (3th ed.). Santa Maria, RS:
    UFSM. (Tabela 101, pág. 182)
}
\description{
Grupo de experimentos, avaliando 6 cultivres de milho,
    conduzidos em 4 locais (4 ambientes) da região central d Rio
    Grande do Sul, no delineamento de blocos ao acaso com 4
    repetições. Foi avaliado o rendimento de grãos.
}
\examples{

library(lattice)

data(StorckTb101)
str(StorckTb101)

xyplot(prod ~ cult | loc, groups = bloc, data = StorckTb101,
       type = "o", as.table = TRUE,
       xlab = "Cultivares",
       ylab = "Rendimento de grãos (kg/parcela)")

}
\keyword{DBC}
\keyword{GE}

