% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storck.R
\name{StorckTb60}
\alias{StorckTb60}
\title{Experimento Bifatorial em Delineamento de Blocos Ao Acaso}
\format{Um \code{data.frame} com 48 observações e 4 variáveis, em que

\describe{

\item{\code{cult}}{Fator categórico que representa 4 cultivares de
    milho.}

\item{\code{manejo}}{Fator cetegórico que representa 3 formas de
    manejo. }

\item{\code{bloc}}{Fator categórico que representa 4 blocos.}

\item{\code{result}}{Resultados (kg/parcela) medido em cada parcela.}

}}
\source{
STORCK et al. (2011), Tabela 60, pág. 138.
}
\description{
Resultados fictícios (kg/parcela) de um experimento
    bifatorial 4 \eqn{\times} 3, no delineamento de blocos ao acaso,
    com 4 repetições.
}
\examples{

library(lattice)

data(StorckTb60)
str(StorckTb60)

xyplot(prod ~ cult, groups = manejo, data = StorckTb60,
       type = c("p", "a"),
       xlab = "Tipos de manejo",
       ylab = "Produção (kg/parcela)",
       auto.key = list(title = "Manejo", cex.title = 1.1,
                       columns = 3))

}
\keyword{DBC}
\keyword{FAT2}
