% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vieira.R
\name{VieiraEx7.3}
\alias{VieiraEx7.3}
\title{Experimento de um Fator em DIC com N\enc{ú}{u}mero Diferente
    de Repeti\enc{çõ}{co}es}
\format{Um \code{data.frame} com 18 observações e 2 variáveis, em que

\describe{

\item{\code{tratamento}}{Tratamento aplicado na unidade
    experimental.}

\item{\code{valor}}{Valor da resposta medida nas unidades
    experimentais.}

}}
\source{
VIEIRA (1999), Exercício 7.3, pág. 85.
}
\description{
Dados de um experimento inteiramente ao acaso com número
    diferente de repetições.
}
\examples{

data(VieiraEx7.3)
str(VieiraEx7.3)

xtabs(~tratamento, data = VieiraEx7.3)

library(lattice)

xyplot(valor ~ tratamento, data = VieiraEx7.3,
       jitter.x = TRUE, type = c("p", "a"))

}
\keyword{DIC}

