% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vieira.R
\name{VieiraPg50.1}
\alias{VieiraPg50.1}
\title{Experimento em Delineamento Inteiramente Casualizado}
\format{Um \code{data.frame} com 30 observações e 2 variáveis, em que

\describe{

\item{\code{trat}}{Fator categórico de 5 níveis.}

\item{\code{resp}}{Variável resposta em números inteiros.}

}}
\source{
VIEIRA (1999), pág. 50, exercício 2.
}
\description{
Experimento em delineamento inteiramente casualizado que
    estudou o efeito de um fator de 5 níveis categóricos em uma
    resposta na escala dos números inteiros.
}
\examples{

data(VieiraPg50.1)

str(VieiraPg50.1)

library(lattice)

xyplot(resp ~ trat, data = VieiraPg50.1,
       jitter.x = TRUE,
       xlab = "Tratamento",
       ylab = "Resposta")

}
\keyword{DIC}
