% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb12.1}
\alias{ZimmermannTb12.1}
\title{Produtividade de Gr\enc{ã}{a}os de Arroz Irrigado}
\format{Um \code{data.frame} com 42 observações e 3 variáveis

\describe{

\item{\code{geno}}{Fator de níveis nominais. Tratamento aplicado em
     arroz irrigado.}

\item{\code{rept}}{Número inteiro que identifica as repetições de
    cada tratamento.}

\item{\code{prod}}{Produção de grãos de arroz irrigado. A unidade de
    medida não é conhecida.}

}}
\source{
ZIMMERMANN (2004), pág. 249.
}
\description{
Dados do Ensaio 1 de um experimento em DIC, que estudou
    a produtividade de grãos de arroz em lavoura conduzida com
    inundação contínua durante todo o ciclo. O experimento teve seis
    repetições e sete tratamentos.
}
\examples{

library(lattice)

data(ZimmermannTb12.1)
str(ZimmermannTb12.1)

xyplot(prod ~ geno, data = ZimmermannTb12.1,
       type = c("p", "a"),
       xlab = "Tratamentos",
       ylab = "Produção de grãos de arroz irrigado")

aggregate(prod ~ geno, data = ZimmermannTb12.1,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

}
\keyword{DIC}
