% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZimmermannTb14.9.R
\name{ZimmermannTb14.9}
\alias{ZimmermannTb14.9}
\title{Produtividade de grãos de arroz com controle de insetos}
\format{Um \code{data.frame} com 36 observações e 4 variáveis

\describe{

\item{\code{linha}}{Fator de níveis numéricos. Indica a linha da 
    observação.}

\item{\code{coluna}}{Fator de níveis numéricos. Indica a coluna da 
    observação.}

\item{\code{inset}}{Indica o inseticida usado na observação.}

\item{\code{prod}}{Produção de grãos por área útil da parcela.}

}}
\source{
Zimmermann, F. J. (2004). Estatística aplicada à pesquisa
    agrícola (1st ed.). Santo Antônio de Goiás, GO: Embrapa Arroz e
    Feijão. (pg 297)
}
\description{
Dados de um ensaio de controle de insetos/pragas de arroz.
    O experimento foi feito em delineamento quadrado latino, mas há uma 
    um dado perdido (linha 6, coluna 2). Esta observação recebeu o valor 
    zero.
}
\examples{

library(lattice)
library(latticeExtra)
library(reshape)
 
data(ZimmermannTb14.9)

cast(ZimmermannTb14.9, linha ~ coluna, value = "inset")

levelplot(prod ~ linha + coluna,
          data = ZimmermannTb14.9, aspect = "iso",
          panel = function(x, y, z, subscripts, ...) {
            panel.levelplot(x, y, z, subscripts = subscripts)
            panel.text(x, y, ZimmermannTb14.9$inset[subscripts],
                       pos = 3)
            panel.text(x, y, sprintf("\%0.1f", z), pos = 1)
          })

xyplot(prod ~ inset, data = ZimmermannTb14.9, type=c("p","a"),
       xlab="Inseticidas", ylab="Produção", 
       main="Controle de Insetos")
}
\keyword{DQL}

