% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb3.12}
\alias{ZimmermannTb3.12}
\title{Dados de Mat\enc{é}{e}ria seca em Plantas}
\format{Um \code{data.frame} com 75 observações e 4 variáveis, em que

\describe{

\item{\code{solo}}{Fator de níveis nominais representados por
    inteiros. Indica o tipo de solo aonde o arroz foi plantado.}

\item{\code{planta}}{Fator de níveis numéricos que identifica as
    plantas dentro dos vasos.}

\item{\code{vaso}}{Fator de níveis numéricos que identifica os vasos
    de cada tratamento.}

\item{\code{prod}}{Produção de matéria seca das plantas, em gramas.}

}}
\source{
ZIMMERMANN (2004), Table 3.12, pág. 62.
}
\description{
Experimento em DIC que estudou a produção de matéria
    seca em plantas de arroz, em gramas. No experimento foram
    utilizados cinco vasos para cada um de quatro tratamentos e três
    plantas para cada vaso (parcela).
}
\examples{

library(lattice)

data(ZimmermannTb3.12)
str(ZimmermannTb3.12)

xtabs(~solo + vaso, data = ZimmermannTb3.12)

aggregate(prod ~ solo, data = ZimmermannTb3.12,
          FUN = function(x) { c(mean = mean(x), var = var(x)) })

xyplot(prod ~ solo, groups = vaso, data = ZimmermannTb3.12,
       type = c("p", "a"), jitter.x = TRUE,
       xlab = "Tratamentos",
       ylab = "Produção de matéria seca das plantas (g)")

}
\keyword{DIC}

