% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zimmermann.R
\name{ZimmermannTb3.5}
\alias{ZimmermannTb3.5}
\title{Fracionamento da Aduba\enc{çã}{ca}o na Produ\enc{çã}{ca}o de
    Gr\enc{ã}{a}os de Arroz Irrigado}
\format{Um \code{data.frame} com 32 observações e 3 variáveis

\describe{

\item{\code{trat}}{Fator de níveis categóricos que representa as
    formas de adubação: 1 = 80 kg/ha no plantio, 2 = 40 kg/ha 40 dias
    no plantio após a emergência (DAE), 3 = 13.2 kg/ha no plantio e
    66.8 kg/ha aos 40 DAE, e 4 = 13.2 kg/ha no plantio e 33.4 kg/ha
    aos 40 e 60 DAE.}

\item{\code{rept}}{Inteiro que identifica as repretições de uma forma
    de adubação.}

\item{\code{prod}}{Produção de grãos de arroz irrigado (ka
    ha\eqn{^{-1}}).}

}}
\source{
ZIMMERMANN (2004), Tabela 3.5, pág. 54.
}
\description{
Dados de um estudo sobre adubação nitrogenada na cultura
    do arroz irrigado. Foram testadas quatros formas de aplicação do
    adubo fazendo o fracionamento das quantidades em diferentes
    épocas. O experimento foi instalado em delineamento inteiramente
    casualizado com 8 repetições.
}
\examples{

library(lattice)

data(ZimmermannTb3.5)
str(ZimmermannTb3.5)

# Quantidade[época].
adub <- expression(80[0],
                   40[0] + 40[40],
                   13.2[0] + 66.8[40],
                   13.2[0] + 33.4[40] + 33.4[60])

xyplot(prod ~ adub, data = ZimmermannTb3.5,
       type = c("p", "a"), jitter.x = TRUE,
       xlab = "Formas de adubação",
       ylab = expression("Produção de arroz"~(kg~ha^{-1})),
       scales = list(x = list(labels = adub)))

aggregate(prod ~ adub, data = ZimmermannTb3.5,
          FUN = function(x) { c(mean = mean(x), sd = sd(x)) })

}
\keyword{DIC}
\keyword{contrastes}

